### 2025-03-31 #7964 @editorialbot
[PRE REVIEW]: mlr3extralearners: Expanding the mlr3 Ecosystem with Community-Driven Learner Integration
**Submitting author:** @sebffischer (<a href="http://orcid.org/0000-0002-9609-3197">Sebastian Fischer</a>)
**Repository:** https://github.com/mlr-org/mlr3extralearners
**Branch with paper.md** (empty if default branch): paper
**Version:** v1.0.0
