### 2025-03-31 f828782
fastai classifier experimental
