+++ attic/scratch_to_be_deleted_v2.R
@@ -1,0 +1,489 @@
+#' @title Extreme Gradient Boosting Classification Learner
+#'
+#' @name mlr_learners_classif.xgboost
+#'
+#' @description
+#' eXtreme Gradient Boosting classification.
+#' Calls [xgboost::xgb.train()] from package \CRANpkg{xgboost}.
+#'
+#' If not specified otherwise, the evaluation metric is set to the default `"logloss"`
+#' for binary classification problems and set to `"mlogloss"` for multiclass problems.
+#' This was necessary to silence a deprecation warning.
+#'
+#' Note that using the `watchlist` parameter directly will lead to problems when wrapping this [mlr3::Learner] in a
+#' `mlr3pipelines` `GraphLearner` as the preprocessing steps will not be applied to the data in the watchlist.
+#' See the section *Early Stopping and Validation* on how to do this.
+#'
+#' @template note_xgboost
+#' @section Initial parameter values:
+#' - `nrounds`:
+#'   - Actual default: no default.
+#'   - Adjusted default: 1000.
+#'   - Reason for change: Without a default construction of the learner would error.
+#'     The lightgbm learner has a default of 1000, so we use the same here.
+#' - `nthread`:
+#'   - Actual value: Undefined, triggering auto-detection of the number of CPUs.
+#'   - Adjusted value: 1.
+#'   - Reason for change: Conflicting with parallelization via \CRANpkg{future}.
+#' - `verbose`:
+#'   - Actual default: 1.
+#'   - Adjusted default: 0.
+#'   - Reason for change: Reduce verbosity.
+#'
+#' @section Early Stopping and Validation:
+#' In order to monitor the validation performance during the training, you can set the `$validate` field of the Learner.
+#' For information on how to configure the validation set, see the *Validation* section of [mlr3::Learner].
+#' This validation data can also be used for early stopping, which can be enabled by setting the `early_stopping_rounds` parameter.
+#' The final (or in the case of early stopping best) validation scores can be accessed via `$internal_valid_scores`, and the optimal `nrounds` via `$internal_tuned_values`.
+#' The internal validation measure can be set via the `eval_metric` parameter that can be a [mlr3::Measure], a function, or a character string for the internal xgboost measures.
+#' Using an [mlr3::Measure] is slower than the internal xgboost measures, but allows to use the same measure for tuning and validation.
+#'
+#' @inheritSection mlr_learners_regr.xgboost Offset
+#'
+#' @templateVar id classif.xgboost
+#' @template learner
+#'
+#' @references
+#' `r format_bib("chen_2016")`
+#'
+#' @export
+#' @template seealso_learner
+#' @template example_dontrun
+#' @examples
+#'
+#' \dontrun{
+#' # Train learner with early stopping on spam data set
+#' task = tsk("spam")
+#'
+#' # use 30 percent for validation
+#' # Set early stopping parameter
+#' learner = lrn("classif.xgboost",
+#'   nrounds = 100,
+#'   early_stopping_rounds = 10,
+#'   validate = 0.3
+#' )
+#'
+#' # Train learner with early stopping
+#' learner$train(task)
+#'
+#' # Inspect optimal nrounds and validation performance
+#' learner$internal_tuned_values
+#' learner$internal_valid_scores
+#' }
+LearnerClassifXgboost = R6Class("LearnerClassifXgboost",
+                                inherit = LearnerClassif,
+
+                                public = list(
+
+                                  #' @description
+                                  #' Creates a new instance of this [R6][R6::R6Class] class.
+                                  initialize = function() {
+
+                                    p_nrounds = p_int(1L,
+                                                      tags = c("train", "hotstart", "internal_tuning"),
[…]

[…]
