+++ R/learner_fastai_classif_fastai.R
@@ -1,9 +1,9 @@
-#' @title Classification Neural Network Learner
+#' @title FastAi Neural Network Tabular Classifier
 #' @author annanzrv
 #' @name mlr_learners_classif.fastai
 #'
 #' @description
-#' FIXME: BRIEF DESCRIPTION OF THE LEARNER.
+#' Fit simple feed forward neural networks for tabular classification tasks.
 #' Calls [fastai::fastai()] from FIXME: (CRAN VS NO CRAN): \CRANpkg{fastai} | 'fastai'.
 #'
 #' @section Initial parameter values:
@@ -23,6 +23,7 @@
 #' @export
 LearnerClassifFastAi = R6Class("LearnerClassifFastAi",
   inherit = LearnerClassif,
+
   public = list(
     #' @description
     #' Creates a new instance of this [R6][R6::R6Class] class.
@@ -35,11 +36,12 @@ LearnerClassifFastAi = R6Class("LearnerClassifFastAi",
         embed_p     = p_dbl(lower = 0L, upper = 1L, default = 0L, tags = "train"),  # Dropout probability for Embedding layer
         emb_szs     = p_uty(default=NULL, tags="train"),  # Sequence of (num_embeddings, embedding_dim) for each categorical variable
         epochs      = p_int(lower=1, default = 5, tags="train"), # Epochs
+        eval_metric = p_uty(tags = "train", custom_check = crate({function(x) check_true(any(is.character(x), is.function(x), inherits(x, "Measure")))})),
         layers      = p_uty(tags="train"),  # Sequence of ints used to specify the input and output size of each LinBnDrop layer
         loss_func   = p_uty(tags="train"),  # Defaults to fastai::CrossEntropyLossFlat()
         lr          = p_dbl(lower=0, default = 0.001, tags = "train"),  # Learning rate
         metrics     = p_uty(tags = "train"),  # optional list of metrics, e.g, fastai::Precision() or fastai::accuracy()
-        n_out       = p_int(default = NULL, tags="train"),  # ?
+        n_out       = p_int(tags="train"),  # ?
         num_workers = p_int(default = 0L, tags = "train"),  # how many subprocesses to use for data loading
         opt_func    = p_uty(tags="train"),  # Optimizer created when Learner.fit is called. E.g., fastai::Adam()
         pin_memory  = p_lgl(default = TRUE, tags = "train"),  # If True, the data loader will copy Tensors into CUDA pinned memory before returning them.
@@ -60,10 +62,12 @@ LearnerClassifFastAi = R6Class("LearnerClassifFastAi",
         param_set = param_set,
         properties = c("missings", "multiclass", "twoclass", "validation", "weights"),
         man = "mlr3extralearners::mlr_learners_classif.fastai",
-        label = "FastAi Neural Network Tabular Learner"
+        label = "FastAi Neural Network Tabular Classifier"
       )
     },
   ),
+
+
   active = list(
     #' @field internal_valid_scores (named `list()` or `NULL`)
     #' The validation scores extracted from `model$evaluation_log`.
@@ -97,10 +101,16 @@ LearnerClassifFastAi = R6Class("LearnerClassifFastAi",
       # get parameters for training
       pars = self$param_set$get_values(tags = "train")
 
-      pv_dt = formalArgs(fastai::TabularDataTable)
-      pv_dl = formalArgs(fastai:::fastai2$data$load$DataLoader)
-      pv_config = formalArgs(fastai::tabular_config)
-      pv_fit = formalArgs(fastai:::fastai2$learner$Learner$fit)
+      args_dt = formalArgs(fastai::TabularDataTable)
+      args_dl = formalArgs(fastai:::fastai2$data$load$DataLoader)
+      args_config = formalArgs(fastai::tabular_config)
[…]
