### 2025-04-02 3be0968
fix fit method
