+++ attic/scratch_to_be_deleted.R
@@ -16,15 +16,15 @@ df_fai = fastai::TabularDataTable(df, cat_names = cat_cols, cont_names = num_col
                                   y_names = task$target_names, splits=ids)
 
 dl = fastai::dataloaders(df_fai, bs=50)
-dl$train$y
-
-target_factor <- factor(dl$train$y)
-weights <- c(0.7, 0.3)
-weight_mapping <- setNames(weights, levels(target_factor))
-weights_for_factor <- weight_mapping[target_factor]
-class_weights <- unname(weights_for_factor)
-
-dl$train$wgts <- class_weights / sum(class_weights)
+# dl$train$y
+#
+# target_factor <- factor(dl$train$y)
+# weights <- c(0.7, 0.3)
+# weight_mapping <- setNames(weights, levels(target_factor))
+# weights_for_factor <- weight_mapping[target_factor]
+# class_weights <- unname(weights_for_factor)
+#
+# dl$train$wgts <- class_weights / sum(class_weights)
 
 # These params are not documented by fastai-r but only by fastai-py:
 # ps_dataloaders = ps(
@@ -37,7 +37,7 @@ dl$train$wgts <- class_weights / sum(class_weights)
 
 
 # Step 2: Define a set of configs for the learner
-config = fastai::tabular_config(embed_p = 0.3, use_bn = FALSE)
+config = fastai::tabular_config(embed_p = 0.3, use_bn = FALSE, epochs = 5)
 # ps_config = ps(
 #   ps       = p_uty(default = NULL, tags = "train"),  # Sequence of dropout probabilities
 #   embed_p  = p_dbl(lower = 0L, upper = 1L, default = 0L, tags = "train"),  # Dropout probability for Embedding layer
@@ -66,12 +66,13 @@ config = fastai::tabular_config(embed_p = 0.3, use_bn = FALSE)
 # Step 3: Define and fit the tabular learner
 tab_learner = fastai::tabular_learner(dl, layers=c(200,100,100),
                                 config = config,
-                                metrics = list(accuracy, RocAucBinary(),
-                                               Precision(), Recall(),
-                                               F1Score()))
+                                metrics = list(fastai::Accuracy(),
+                                               fastai::RocAucBinary(),
+                                               fastai::Precision(), fastai::Recall(),
+                                               fastai::F1Score()))
 invisible(tab_learner$remove_cb(tab_learner$progress))  # to avoid plot creation when internally validating
 
-# invisible(reticulate::py_capture_output(fit(tab_learner, n_epoch = 5, lr = 0.005)))
+invisible(reticulate::py_capture_output(fit(tab_learner, n_epoch = 5, lr = 0.005)))
 
 fit_eval = function(tab_learner, args) {
   eval_protocol = invoke(  # set this to self$eval_protocol
@@ -99,7 +100,7 @@ eval_protocol = fastai::fit(tab_learner, n_epoch = 5, lr = 0.005)
 
 
 # Prediction:
-predict(model, df[ids$test])
+predict(tab_learner, df[ids$test])
 
 
 
@@ -131,65 +132,15 @@ metric = function(pred, dtrain, msr = NULL, ...) {
 }
 
 
-tab_learner = fastai::tabular_learner(dl, layers=NULL,
-                                      config = NULL,
-                                      metrics = NULL)
+tab_learner = fastai::tabular_learner(dl, layers=c(100, 200, 100),
+                                      config = config,
+                                      metrics = fastai::AccumMetric(metric, flatten=FALSE, msr=logloss))
 
[…]
