+++ R/learner_fastai_classif_fastai.R
@@ -1,10 +1,10 @@
-#' @title FastAi Neural Network Tabular Classifier
+#' @title Classification Neural Network Learner
 #' @author annanzrv
 #' @name mlr_learners_classif.fastai
 #'
 #' @description
-#' Fit simple feed forward neural networks for tabular classification tasks.
-#' Calls [fastai::fastai()] from FIXME: (CRAN VS NO CRAN): \CRANpkg{fastai} | 'fastai'.
+#' Simple and fast neural nets for tabular data classification.
+#' Calls [fastai::fastai()] from package \CRANpkg{fastai}.
 #'
 #' @section Initial parameter values:
 #' FIXME: DEVIATIONS FROM UPSTREAM PARAMETERS. DELETE IF NOT APPLICABLE.
@@ -21,13 +21,14 @@
 #' @template seealso_learner
 #' @template example
 #' @export
-LearnerClassifFastAi = R6Class("LearnerClassifFastAi",
+LearnerClassifFastai = R6Class("LearnerClassifFastai",
   inherit = LearnerClassif,
-
   public = list(
+    eval_protocol = NULL,
     #' @description
     #' Creates a new instance of this [R6][R6::R6Class] class.
     initialize = function() {
+      # FIXME: MANUALLY ADD PARAMETERS BELOW AND THEN DELETE THIS LINE
       param_set = ps(
         act_cls     = p_uty(tags = "train"),  # Activation type for LinBnDrop layers, e.g., fastai::nn()$ReLU(inplace = TRUE)
         bn_cont     = p_lgl(default = TRUE, tags = "train"),  # Use BatchNorm1d on continuous variables
@@ -35,7 +36,7 @@ LearnerClassifFastAi = R6Class("LearnerClassifFastAi",
         drop_last   = p_lgl(default = FALSE, tags = "train"),  #  If True, then the last incomplete batch is dropped.
         embed_p     = p_dbl(lower = 0L, upper = 1L, default = 0L, tags = "train"),  # Dropout probability for Embedding layer
         emb_szs     = p_uty(default=NULL, tags="train"),  # Sequence of (num_embeddings, embedding_dim) for each categorical variable
-        epochs      = p_int(lower=1, default = 5, tags="train"), # Epochs
+        n_epoch     = p_int(lower=1, default = 5, tags="train"), # Epochs
         eval_metric = p_uty(tags = "train", custom_check = crate({function(x) check_true(any(is.character(x), is.function(x), inherits(x, "Measure")))})),
         layers      = p_uty(tags="train"),  # Sequence of ints used to specify the input and output size of each LinBnDrop layer
         loss_func   = p_uty(tags="train"),  # Defaults to fastai::CrossEntropyLossFlat()
@@ -54,19 +55,24 @@ LearnerClassifFastAi = R6Class("LearnerClassifFastAi",
         bs          = p_int(default = 50, tags="train") # how many samples per batch to load
       )
 
+      # Epochs is set to default parameter.
+      if (is.null(parmset$values$n_epoch)) {
+        param_set$values = list(n_epoch = 5)
+      }
+      self$eval_protocol = NULL
+
       super$initialize(
         id = "classif.fastai",
-        packages = c("mlr3extralearners", "fastai"),
+        packages = c("fastai", "reticulate"),
         feature_types = c("logical", "integer", "numeric", "factor", "ordered"),
         predict_types = c("response", "prob"),
         param_set = param_set,
-        properties = c("missings", "multiclass", "twoclass", "validation", "weights"),
+        properties = c("missings", "multiclass", "twoclass", "weights"),
         man = "mlr3extralearners::mlr_learners_classif.fastai",
         label = "FastAi Neural Network Tabular Classifier"
       )
-    },
+    }
   ),
-
 
   active = list(
[…]
