### 2025-04-07 d57abf4
add early stopping
