+++ R/learner_fastai_classif_fastai.R
@@ -7,7 +7,9 @@
 #' Calls [fastai::fastai()] from package \CRANpkg{fastai}.
 #'
 #' @section Initial parameter values:
-#' FIXME: DEVIATIONS FROM UPSTREAM PARAMETERS. DELETE IF NOT APPLICABLE.
+#' - `n_epoch`:
+#'   Needs to be set for [fastai::fit()] to work.
+#'   If no value is given, it is set to 5.
 #'
 #' @section Custom mlr3 parameters:
 #' FIXME: DEVIATIONS FROM UPSTREAM DEFAULTS. DELETE IF NOT APPLICABLE.
@@ -28,35 +30,36 @@ LearnerClassifFastai = R6Class("LearnerClassifFastai",
     #' @description
     #' Creates a new instance of this [R6][R6::R6Class] class.
     initialize = function() {
-      # FIXME: MANUALLY ADD PARAMETERS BELOW AND THEN DELETE THIS LINE
       param_set = ps(
-        act_cls     = p_uty(tags = "train"),  # Activation type for LinBnDrop layers, e.g., fastai::nn()$ReLU(inplace = TRUE)
-        bn_cont     = p_lgl(default = TRUE, tags = "train"),  # Use BatchNorm1d on continuous variables
-        bn_final    = p_lgl(default = FALSE, tags = "train"),  # Use BatchNorm1d on final layer
-        drop_last   = p_lgl(default = FALSE, tags = "train"),  #  If True, then the last incomplete batch is dropped.
-        embed_p     = p_dbl(lower = 0L, upper = 1L, default = 0L, tags = "train"),  # Dropout probability for Embedding layer
-        emb_szs     = p_uty(default=NULL, tags="train"),  # Sequence of (num_embeddings, embedding_dim) for each categorical variable
-        n_epoch     = p_int(lower=1, default = 5, tags="train"), # Epochs
-        eval_metric = p_uty(tags = "train", custom_check = crate({function(x) check_true(any(is.character(x), is.function(x), inherits(x, "Measure")))})),
-        layers      = p_uty(tags="train"),  # Sequence of ints used to specify the input and output size of each LinBnDrop layer
-        loss_func   = p_uty(tags="train"),  # Defaults to fastai::CrossEntropyLossFlat()
-        lr          = p_dbl(lower=0, default = 0.001, tags = "train"),  # Learning rate
-        metrics     = p_uty(tags = "train"),  # optional list of metrics, e.g, fastai::Precision() or fastai::accuracy()
-        n_out       = p_int(tags="train"),  # ?
-        num_workers = p_int(default = 0L, tags = "train"),  # how many subprocesses to use for data loading
-        opt_func    = p_uty(tags="train"),  # Optimizer created when Learner.fit is called. E.g., fastai::Adam()
-        pin_memory  = p_lgl(default = TRUE, tags = "train"),  # If True, the data loader will copy Tensors into CUDA pinned memory before returning them.
-        ps          = p_uty(default = NULL, tags = "train"),  # Sequence of dropout probabilities
-        shuffle     = p_lgl(default = FALSE, tags = "train"),  # If True, then data is shuffled every time dataloader is fully read/iterated.
-        train_bn    = p_lgl(default = TRUE, tags = "train"),  # controls if BatchNorm layers are trained
-        wd_bn_bias  = p_lgl(default = FALSE, tags = "train"),  # controls if weight decay is applied to BatchNorm layers and bias
-        use_bn      = p_lgl(default = TRUE, tags = "train"),  # Use BatchNorm1d in LinBnDrop layers
-        y_range     = p_uty(default = NULL, tags = "train"),  # Low and high for SigmoidRange activation (see below)
[…]
