### 2025-04-12 #7964 @RichardLitt
Thanks @JonnyTran. Best of luck with things.

@salbalkus @hbaniecki or @kelly-sovacool: Would you be interested in reviewing this paper?
