### 2025-04-14 371db82
fix: params and math rendering
