### 2025-04-14 c6ee2e3
fix: params and math rendering (#421)
