+++ NAMESPACE
@@ -33,6 +33,7 @@ export(LearnerClassifMob)
 export(LearnerClassifMultilayerPerceptron)
 export(LearnerClassifNaiveBayesMultinomial)
 export(LearnerClassifNaiveBayesWeka)
+export(LearnerClassifObliqueRandomForest)
 export(LearnerClassifOneR)
 export(LearnerClassifPART)
 export(LearnerClassifPriorityLasso)
@@ -83,6 +84,7 @@ export(LearnerRegrM5Rules)
 export(LearnerRegrMars)
 export(LearnerRegrMob)
 export(LearnerRegrMultilayerPerceptron)
+export(LearnerRegrObliqueRandomForest)
 export(LearnerRegrPriorityLasso)
 export(LearnerRegrQGam)
 export(LearnerRegrREPTree)
