+++ man/mlr_learners_regr.mqgam.Rd
@@ -1,0 +1,146 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/learner_qgam_regr_mqgam.R
+\name{mlr_learners_regr.mqgam}
+\alias{mlr_learners_regr.mqgam}
+\alias{LearnerRegrMQGam}
+\title{Generalized Additive Quantile Regression Model for Multiple Quantiles}
+\description{
+Quantile Regression with generalized additive models for fitting a learner on multiple quantiles simultaneously
+Calls \code{\link[qgam:mqgam]{qgam::mqgam()}} from package \CRANpkg{qgam}.
+}
+\section{Form}{
+
+For the \code{form} parameter, a gam formula specific to the \link[mlr3:Task]{Task} is required (see example and \code{?mgcv::formula.gam}).
+If no formula is provided, a fallback formula using all features in the task is used that will make the Learner behave like Linear Quantile Regression.
+The features specified in the formula need to be the same as columns with col_roles "feature" in the task.
+}
+
+\section{Quantile}{
+
+The quantiles for the Learner, i.e. \code{qu} parameter from \code{\link[qgam:mqgam]{qgam::mqgam()}}, is set using the values specified in \code{learner$quantiles}.
+The response quantile can be specified using \code{learner$quantile_response}
+}
+
+\section{Dictionary}{
+
+This \link[mlr3:Learner]{Learner} can be instantiated via \link[mlr3:mlr_sugar]{lrn()}:
+
+\if{html}{\out{<div class="sourceCode">}}\preformatted{lrn("regr.mqgam")
+}\if{html}{\out{</div>}}
+}
+
+\section{Meta Information}{
+
+\itemize{
+\item Task type: \dQuote{regr}
+\item Predict Types: \dQuote{response}, \dQuote{se}, \dQuote{quantiles}
+\item Feature Types: \dQuote{logical}, \dQuote{integer}, \dQuote{numeric}, \dQuote{factor}
+\item Required Packages: \CRANpkg{mlr3}, \CRANpkg{qgam}
+}
+}
+
+\section{Parameters}{
+\tabular{lllll}{
+   Id \tab Type \tab Default \tab Levels \tab Range \cr
+   form \tab untyped \tab - \tab  \tab - \cr
+   lsig \tab numeric \tab - \tab  \tab \eqn{(-\infty, \infty)}{(-Inf, Inf)} \cr
+   err \tab numeric \tab - \tab  \tab \eqn{[0, 1]}{[0, 1]} \cr
+   cluster \tab untyped \tab NULL \tab  \tab - \cr
+   multicore \tab logical \tab - \tab TRUE, FALSE \tab - \cr
+   ncores \tab numeric \tab - \tab  \tab \eqn{(-\infty, \infty)}{(-Inf, Inf)} \cr
+   paropts \tab untyped \tab list() \tab  \tab - \cr
+   link \tab untyped \tab "identity" \tab  \tab - \cr
+   argGam \tab untyped \tab - \tab  \tab - \cr
+   discrete \tab logical \tab FALSE \tab TRUE, FALSE \tab - \cr
+}
+}
+
+\examples{
+\dontshow{if (requireNamespace("qgam", quietly = TRUE)) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
+# simple example
+t = mlr3::tsk("mtcars")
+l = mlr3::lrn("regr.mqgam")
+t$select(c("cyl", "am", "disp", "hp"))
+l$param_set$values$form = mpg ~ cyl + am + s(disp) + s(hp)
+l$quantiles = c(0.05, 0.5, 0.95)
+l$quantile_response = 0.5
+l$train(t)
[…]
