+++ NAMESPACE
@@ -81,6 +81,7 @@ export(LearnerRegrLinearRegression)
 export(LearnerRegrLmer)
 export(LearnerRegrM5P)
 export(LearnerRegrM5Rules)
+export(LearnerRegrMQGam)
 export(LearnerRegrMars)
 export(LearnerRegrMob)
 export(LearnerRegrMultilayerPerceptron)
