### 2025-04-14 d89d59d
fix examples in qgam learner
