+++ R/learner_qgam_regr_mqgam.R
@@ -7,12 +7,15 @@
 #' Calls [qgam::mqgam()] from package \CRANpkg{qgam}.
 #'
 #' @section Form:
-#' For the `form` parameter, a gam formula specific to the [Task][mlr3::Task] is required (see example and `?mgcv::formula.gam`).
-#' If no formula is provided, a fallback formula using all features in the task is used that will make the Learner behave like Linear Quantile Regression.
+#' For the `form` parameter, a gam formula specific to the [Task][mlr3::Task] is required
+#' (see example and `?mgcv::formula.gam`).
+#' If no formula is provided, a fallback formula using all features in the task is used that will make the Learner
+#' behave like Linear Quantile Regression.
 #' The features specified in the formula need to be the same as columns with col_roles "feature" in the task.
 #'
 #' @section Quantile:
-#' The quantiles for the Learner, i.e. `qu` parameter from [qgam::mqgam()], is set using the values specified in `learner$quantiles`.
+#' The quantiles for the Learner, i.e. `qu` parameter from [qgam::mqgam()],
+#' is set using the values specified in `learner$quantiles`.
 #' The response quantile can be specified using `learner$quantile_response`
 #'
 #' @templateVar id regr.mqgam
@@ -35,118 +38,118 @@
 #' l$predict(t)
 #' @export
 LearnerRegrMQGam = R6Class("LearnerRegrMQGam",
-                          inherit = LearnerRegr,
-                          public = list(
-                            #' @description
-                            #' Creates a new instance of this [R6][R6::R6Class] class.
-                            initialize = function() {
-                              param_set = ps(
-                                form          = p_uty(tags = "train"),
-                                lsig          = p_dbl(tags = "train"),
-                                err           = p_dbl(lower = 0, upper = 1, tags = "train"),
-                                cluster       = p_uty(default = NULL, tags = "train"),
-                                multicore     = p_lgl(tags = "train"),
-                                ncores        = p_dbl(tags = "train"),
-                                paropts       = p_uty(default = list(), tags = "train"),
-                                link          = p_uty(default = "identity", tags = "train"),
-                                argGam        = p_uty(custom_check = crate(function(x) {
-                                  checkmate::check_list(x, names = "unique", null.ok = TRUE)
-                                }), tags = "train"),
-                                discrete      = p_lgl(default = FALSE, tags = "train")
-                              )
+  inherit = LearnerRegr,
+  public = list(
+    #' @description
+    #' Creates a new instance of this [R6][R6::R6Class] class.
+    initialize = function() {
+      param_set = ps(
+        form          = p_uty(tags = "train"),
+        lsig          = p_dbl(tags = "train"),
+        err           = p_dbl(lower = 0, upper = 1, tags = "train"),
+        cluster       = p_uty(default = NULL, tags = "train"),
+        multicore     = p_lgl(tags = "train"),
+        ncores        = p_dbl(tags = "train"),
+        paropts       = p_uty(default = list(), tags = "train"),
+        link          = p_uty(default = "identity", tags = "train"),
+        argGam        = p_uty(custom_check = crate(function(x) {
+          checkmate::check_list(x, names = "unique", null.ok = TRUE)
+        }), tags = "train"),
+        discrete      = p_lgl(default = FALSE, tags = "train")
+      )
 
-                              super$initialize(
-                                id = "regr.mqgam",
-                                packages = "qgam",
-                                feature_types = c("logical", "integer", "numeric", "factor"),
-                                predict_types = c("response", "se", "quantiles"),
-                                param_set = param_set,
-                                properties = "weights",
-                                man = "mlr3extralearners::mlr_learners_regr.mqgam",
-                                label = "Regression Quantile Generalized Additive Model Learner"
-                              )
+      super$initialize(
+        id = "regr.mqgam",
+        packages = "qgam",
[…]
