### 2025-04-22 796c06a
feat: add tabpfn classifier
