+++ tests/testthat/test_tabpfn_classif_tabpfn.R
@@ -1,0 +1,7 @@
+test_that("autotest tabpfn", {
+  set.seed(1)
+  learner = lrn("classif.tabpfn", train_mode = "both")
+  expect_learner(learner)
+  result = run_autotest(learner, N = 400, check_replicable = FALSE)
+  expect_true(result, info = result$error)
+})
