+++ NAMESPACE
@@ -46,6 +46,7 @@ export(LearnerClassifRandomTree)
 export(LearnerClassifSGD)
 export(LearnerClassifSMO)
 export(LearnerClassifSimpleLogistic)
+export(LearnerClassifTabPFN)
 export(LearnerClassifVotedPerceptron)
 export(LearnerDensKDEks)
 export(LearnerDensLocfit)
