### 2025-04-22 4ab02ad
feat: implement and check more params
