+++ R/learner_tabpfn_classif_tabpfn.R
@@ -5,16 +5,17 @@
 #' @description
 #' Foundation model for tabular data.
 #' Uses \CRANpgk{reticulate} to interface with the \CRANpkg{tabpfn} Python package.
+#' 
+#' @section Initial parameter values:
+#' - `n_jobs` is initialized to 1 to avoid threading conflicts with \CRANpkg{future}.#'
 #'
 #' @section Custom mlr3 parameters:
 #' - `train_mode`: Controls the behavior of the `$train()` method.
 #'   Possible options are `"fit"`, `"store"`, and `"both"` (default).
-#' 
 #'   - If `"fit"` is selected, the `$train()` function creates a Python object
 #'     of the class `TabPFNClassifier` and calls its `fit()` function.
 #'   - If `"store"` is selected, the `$train()` function only stores the training data.
 #'   - If `"both"` is selected, the `$train()` function does both.
-#' 
 #'   The fitted model (Python object), if created, is stored in the `$model$fitted` slot.
 #'   The training data, if stored, can be accessed in the `$model$X` and $model$y` slots.
 #'   Note that if you save this learner and load it later, the fitted model (Python object)
@@ -23,8 +24,14 @@
 #'   the training data.
 #'   When calling `$predict()` without a fitted model stored, the model will be fitted
 #'   on-the-fly and stored in `$model$fitted` henceforth.
-#' 
-#' @section Initial parameter values:
+#' - `categorical_feature_indices` uses R indexing instead of zero-based Python indexing.
+#' - `device` must be a string.
+#'   If set to `"auto"`, the behavior is the same as original.
+#'   Otherwise, the string is passed as argument to `torch.device()` to create a device.
+#' - `inference_precision` must be a `torch.dtype` name (e.g., `"float32"`
+#'   is interpreted as `torch.float32`) or `"auto"` or `"autocast"`.
+#' - `n_jobs` must be a positive integer specifying the number of CPU cores to use.
+#'   The `-1` option is disabled.
 #' 
 #' @template learner
 #'
@@ -50,12 +57,22 @@ LearnerClassifTabPFN = R6Class("LearnerClassifTabPFN",
         model_path = p_uty(default = "auto", tags = "train", custom_check = check_string),
         device = p_uty(default = "auto", tags = "train", custom_check = check_string),
         ignore_pretraining_limits = p_lgl(default = FALSE, tags = "train"),
-        inference_precision = p_uty(default = "auto", tags = "train"),
+        inference_precision = p_fct(
+          # do we need more dtypes?
+          c("float32", "float64", "bfloat16", "float16", "auto", "autocast"),
+          default = "auto",
+          tags = "train"),
         fit_mode = p_fct(
           c("low_memory", "fit_preprocessors", "fit_with_cache"),
           default = "fit_preprocessors",
           tags = "train"),
-        memory_saving_mode = p_dbl(lower = 0, tags = "train"),
+        memory_saving_mode = p_uty(default = "auto", tags = "train", custom_check = function(x) {
+          if ((x == "auto") || test_flag(x) || test_number(x, lower = 0)) {
+            TRUE
+          } else {
+            "Invalid value for memory_saving_mode. Must be 'auto', a TRUE/FALSE value, or a number > 0."
+          }
+        }),
         random_state = p_int(default = 0, tags = "train"),
         n_jobs = p_int(lower = 1, default = 1, tags = "train")
       )
@@ -64,6 +81,8 @@ LearnerClassifTabPFN = R6Class("LearnerClassifTabPFN",
       ps$values$train_mode = "both"
       # set n_jobs to 1 as mlr3 has its own parallelization
       ps$values$n_jobs = 1
+      # set a seed for the python `fit` function, following TabPFN's default
[…]
