### 2025-04-23 d44c4d3
docs: add section on installation
