+++ R/learner_tabpfn_classif_tabpfn.R
@@ -104,6 +104,9 @@ LearnerClassifTabPFN = R6Class("LearnerClassifTabPFN",
 
   private = list(
     .train = function(task) {
+      reticulate::py_require("tabpfn")
+      reticulate::py_require("torch")
+
       pars = self$param_set$get_values(tags = "train")
       train_mode = pars$train_mode
       if (train_mode == "both") {
@@ -114,7 +117,7 @@ LearnerClassifTabPFN = R6Class("LearnerClassifTabPFN",
 
       # create torch device
       if (!is.null(pars$device)) {
-        torch = reticulate::py_require("torch")
+        torch = reticulate::import("torch")
         pars$device = torch$device(pars$device)
       }
       # create dtype object
