### 2025-04-25 3c722e1
exclude sanity checks and replicability in tests
