### 2025-04-25 f4f389c
fix paramtest for classifier
