+++ tests/testthat/test_fastai_classif_fastai.R
@@ -1,7 +1,7 @@
 test_that("autotest", {
   learner = lrn("classif.fastai")
-  expect_learner(learner)
-  # note that you can skip tests using the exclude argument
-  result = run_autotest(learner, check_replicable=FALSE, exclude = "sanity")
+  expect_learner(learner, check_man = FALSE)
+  # results not replicable, because torch seed must be set in the python backend
+  result = run_autotest(learner, check_replicable=FALSE, exclude="sanity")
   expect_true(result, info = result$error)
 })
