### 2025-04-28 #7964 @Nowosad
Thank you for reaching out @RichardLitt. I took a look at the code and the paper, but to be honest, it's a bit above my head. I don't think I'd be able to give the thorough review it deserves.
