### 2025-05-01 #7964 @kelly-sovacool
@RichardLitt I can review this paper
