### 2025-05-05 f26f747
feat: pickle for marshaling
