### 2025-05-05 1b0237f
fix: wrap model in a list
