### 2025-05-05 d606cfe
fix: remove train_mode
