+++ R/learner_fastai_regr_fastai.R
@@ -1,9 +1,9 @@
-#' @title Classification Neural Network Learner
+#' @title Regression Neural Network Learner
 #' @author annanzrv
-#' @name mlr_learners_classif.fastai
+#' @name mlr_learners_regr.fastai
 #'
 #' @description
-#' Simple and fast neural nets for tabular data classification.
+#' Simple and fast neural nets for tabular data regression
 #' Calls [fastai::fastai()] from package \CRANpkg{fastai}.
 #'
 #' @section Initial parameter values:
@@ -12,9 +12,12 @@
 #'   If no value is given, it is set to 5.
 #' - `eval_metric`:
 #'   Needs to be set for [fastai::predict()] to work.
-#'   If no value is given, it is set to [mlr3measures::classif.logloss].
+#'   If no value is given, it is set to [mlr3measures::regr.rmse].
 #'
-#' @templateVar id classif.fastai
+#' @section Custom mlr3 parameters:
+#' FIXME: DEVIATIONS FROM UPSTREAM DEFAULTS. DELETE IF NOT APPLICABLE.
+#'
+#' @templateVar id regr.fastai
 #' @template learner
 #'
 #' @references
@@ -23,8 +26,8 @@
 #' @template seealso_learner
 #' @template example
 #' @export
-LearnerClassifFastai = R6Class("LearnerClassifFastai",
-  inherit = LearnerClassif,
+LearnerRegrFastai = R6Class("LearnerRegrFastai",
+  inherit = LearnerRegr,
   public = list(
     eval_protocol = NULL,
     #' @description
@@ -64,19 +67,19 @@ LearnerClassifFastai = R6Class("LearnerClassifFastai",
         param_set$set_values(n_epoch = 5)
       }
       if (is.null(param_set$values$eval_metric)) {
-        param_set$set_values(eval_metric = msr("classif.logloss"))
+        param_set$set_values(eval_metric = msr("regr.rmse"))
       }
       self$eval_protocol = NULL
 
       super$initialize(
-        id = "classif.fastai",
-        packages = c("fastai", "reticulate"),
+        id = "regr.fastai",
+        packages = "fastai",
         feature_types = c("logical", "integer", "numeric", "factor", "ordered"),
-        predict_types = c("response", "prob"),
+        predict_types = "response",
         param_set = param_set,
-        properties = c("multiclass", "twoclass", "weights", "validation"),
-        man = "mlr3extralearners::mlr_learners_classif.fastai",
-        label = "FastAi Neural Network Tabular Classifier"
+        properties = c("weights", "validation"),
+        man = "mlr3extralearners::mlr_learners_regr.fastai",
+        label = "FastAi Neural Network Tabular Regression"
       )
     }
   ),
@@ -113,12 +116,12 @@ LearnerClassifFastai = R6Class("LearnerClassifFastai",
     .train = function(task) {
       formula = task$formula()
       data = task$data()
-      cat_cols = task$feature_types[type != "numeric"]$id
+      cat_cols = task$feature_types[type != "numeric", id]
       if ((length(cat_cols) > 0) && (is.null(self$param_set$get_values(tags = "train")$procs))) {
         print("setting categorical vars")
         self$param_set$set_values(procs = list(fastai::Categorify()))
       }
-      num_cols = task$feature_types[type == "numeric"]$id
+      num_cols = task$feature_types[type == "numeric", id]
       comp = NULL  # comparison that fastai uses for internal validation
 
       # get parameters for training
@@ -146,17 +149,11 @@ LearnerClassifFastai = R6Class("LearnerClassifFastai",
         # set internal validation metric and convert it to format compatible with fastai
         measure = pars$eval_metric
         if (inherits(measure, "Measure")) {  # measure from mlr3measures
[…]
