+++ tests/testthat/test_fastai_regr_fastai.R
@@ -1,0 +1,7 @@
+test_that("autotest", {
+  learner = lrn("regr.fastai")
+  expect_learner(learner)
+  # note that you can skip tests using the exclude argument
+  result = run_autotest(learner)
+  expect_true(result, info = result$error)
+})
