+++ attic/scratch_to_be_deleted.R
@@ -1,24 +1,20 @@
-reticulate::use_condaenv('r-reticulate',required = TRUE)
+#reticulate::use_condaenv('r-reticulate',required = TRUE)
 
 
-task = tsk("sonar")
+task = tsk("boston_housing")
+numeric_features = c(task$feature_types[type == "numeric", id], task$target_names)
 df = task$data()
-ids = partition(task)
+df = df[, ..numeric_features]
 
-learner = lrn("classif.fastai")
-tasks = generate_tasks(learner)
-task = tasks$sanity_switched
 # Step 1: Load data into fastai-compatible format:
 # We  need to divide column names into Categorical columns and Continuous columns
 # We can also define several data preprocessors `procs = list(FillMissing(),Categorify(),Normalize())` and pass them to `TabularDataTable`
 
-cat_cols = task$feature_types[type != "numeric"]$id
-num_cols = task$feature_types[type == "numeric"]$id
-
-df_fai = fastai::TabularDataTable(df, cat_names = cat_cols, cont_names = num_cols,
+task = tsk("sonar")
+df = task$data()
+df_fai = fastai::TabularDataTable(df, #cat_names = cat_cols, cont_names = num_cols,
                                   y_names = task$target_names, splits=NULL)
-
-dl = fastai::dataloaders(df_fai, bs=50)
+dl = fastai::dataloaders(df_fai)
 # dl$train$y
 #
 # target_factor <- factor(dl$train$y)
@@ -69,8 +65,8 @@ config = fastai::tabular_config(embed_p = 0.3, use_bn = FALSE)
 # Step 3: Define and fit the tabular learner
 tab_learner = fastai::tabular_learner(dl, layers=c(200,100,100))
 invisible(tab_learner$remove_cb(tab_learner$progress))  # to avoid plot creation when internally validating
+fastai::fit(tab_learner, n_epoch = 5, lr = 0.005)
 
-invisible(reticulate::py_capture_output(fit(tab_learner, n_epoch = 5, lr = 0.005)))
 
 fit_eval = function(tab_learner, args) {
   eval_protocol = invoke(  # set this to self$eval_protocol
@@ -152,3 +148,122 @@ learner$eval_protocol
 learner$internal_tuned_values
 learner$internal_valid_scores
 
+
+
+
+
+
+
+learner = lrn("classif.fastai")
+tasks = generate_tasks(learner)
+learner$train(tasks$sanity_switched)
+pred = learner$predict(tasks$sanity_switched)
+pred$score(mlr3::msr("classif.ce"))
+learner$eval_protocol
+
+learner = lrn("classif.fastai")
+learner$train(tasks$sanity)
+pred = learner$predict(tasks$sanity)
+pred$score(mlr3::msr("classif.ce"))
+learner$eval_protocol
+
+
+
+
+task = tsk("boston_housing")
+ids = partition(task)
+df = task$data()
+# numeric_features = c(task$feature_types[type %in% c("numeric", "integer"), id], task$target_names)
+# df = task$data()
+# df = df[, ..numeric_features]
+
+cat_cols = task$feature_types[type != "numeric", id]
+num_cols = task$feature_types[type == "numeric", id]
+
+df_fai <- invoke(
+  fastai::TabularDataTable,
+  df = df,
+  cat_names = cat_cols,
+  cont_names = num_cols,
+  y_names = task$target_names,
+  procs = list(fastai::Categorify()),
+  splits = ids
+)
+
+dl <- invoke(
+  fastai::dataloaders,
+  df_fai
+)
+
+tab_learner <- fastai::tabular_learner(dl, metrics = list(fastai::msle()))
[…]
