+++ man/mlr_learners_classif.tabpfn.Rd
@@ -6,27 +6,42 @@
 \title{TabPFN Classification Learner}
 \description{
 Foundation model for tabular data.
-Uses \CRANpgk{reticulate} to interface with the \CRANpkg{tabpfn} Python package.
+Uses \CRANpgk{reticulate} to interface with the \href{https://github.com/PriorLabs/TabPFN}{\code{tabpfn}} Python package.
 }
-\section{Custom mlr3 parameters}{
+\section{Installation}{
 
-\itemize{
-\item \code{train_mode}: Controls the behavior of the \verb{$train()} method.
-Possible options are \code{"fit"}, \code{"store"}, and \code{"both"} (default).
-\itemize{
-\item If \code{"fit"} is selected, the \verb{$train()} function creates a Python object
-of the class \code{TabPFNClassifier} and calls its \code{fit()} function.
-\item If \code{"store"} is selected, the \verb{$train()} function only stores the training data.
-\item If \code{"both"} is selected, the \verb{$train()} function does both.
+While the Python dependencies are handled via \code{reticulate::py_require()}, you can
+manually specify a virtual environment by calling \code{reticulate::use_virtualenv()}
+prior to calling the \verb{$train()} function.
+In this virtual environment, the \code{tabpfn} package and its dependencies must be installed.
 }
 
-The fitted model (Python object), if created, is stored in the \verb{$model$fitted} slot.
-The training data, if stored, can be accessed in the \verb{$model$X} and $model$y\verb{slots. Note that if you save this learner and load it later, the fitted model (Python object) in}$model$fitted\verb{will be corrupted due to serialization issues. Therefore, it is recommended to use the}"store"\code{or}"both"\verb{option to store the training data. When calling}$predict()\verb{without a fitted model stored, the model will be fitted on-the-fly and stored in}$model$fitted` henceforth.
-}
+\section{Saving a Learner}{
+
+In order to save a \code{LearnerClassifTabPFN} for later usage,
+it is necessary to call the \verb{$marshal()} method on the \code{Learner}
+before writing it to disk, as the object will otherwise not be saved correctly.
+After loading a marshaled \code{LearnerClassifTabPFN} into R again,
+you then need to call \verb{$unmarshal()} to transform it into a useable state.
 }
 
 \section{Initial parameter values}{
 
+\itemize{
+\item \code{n_jobs} is initialized to 1 to avoid threading conflicts with \CRANpkg{future}.#'
+}
+}
+
+\section{Custom mlr3 parameters}{
+
+\itemize{
+\item \code{categorical_feature_indices} uses R indexing instead of zero-based Python indexing.
+\item \code{device} must be a string.
+If set to \code{"auto"}, the behavior is the same as original.
+Otherwise, the string is passed as argument to \code{torch.device()} to create a device.
+\item \code{inference_precision} must be \code{"auto"} or \code{"autocast"}.
+Passing \code{torch.dtype} is currently not supported.
+}
 }
 
 \section{Dictionary}{
@@ -41,16 +56,14 @@ This [Learner][mlr3::Learner] can be instantiated via [lrn()][mlr3::lrn]:
 }
 
 \references{
[…]
