### 2025-05-05 d15a988
fix: convert na to nan
