### 2025-05-05 fd1b97d
test: add tests for custom params
