### 2025-05-05 cc9ef78
fix: allow "auto" device
