### 2025-05-05 f116566
fix: ensure integer indices
