### 2025-05-05 549ebd1
docs: add inference_config, fix typo
