### 2025-05-05 4374bca
chore: handle dependency for marshaling
