+++ NAMESPACE
@@ -1,7 +1,7 @@
 # Generated by roxygen2: do not edit by hand
 
-S3method(marshal_model,classif.tabpfn_model)
-S3method(unmarshal_model,classif.tabpfn_model_marshaled)
+S3method(mlr3::marshal_model,classif.tabpfn_model)
+S3method(mlr3::unmarshal_model,classif.tabpfn_model_marshaled)
 export(LearnerClassifAbess)
 export(LearnerClassifAdaBoostM1)
 export(LearnerClassifBart)
