+++ NAMESPACE
@@ -1,7 +1,7 @@
 # Generated by roxygen2: do not edit by hand
 
-S3method(mlr3::marshal_model,classif.tabpfn_model)
-S3method(mlr3::unmarshal_model,classif.tabpfn_model_marshaled)
+S3method(marshal_model,classif.tabpfn_model)
+S3method(unmarshal_model,classif.tabpfn_model_marshaled)
 export(LearnerClassifAbess)
 export(LearnerClassifAdaBoostM1)
 export(LearnerClassifBart)
@@ -147,7 +147,9 @@ importFrom(mlr3,LearnerClassif)
 importFrom(mlr3,LearnerRegr)
 importFrom(mlr3,lrn)
 importFrom(mlr3,lrns)
+importFrom(mlr3,marshal_model)
 importFrom(mlr3,mlr_learners)
+importFrom(mlr3,unmarshal_model)
 importFrom(stats,as.formula)
 importFrom(stats,formula)
 importFrom(stats,na.omit)
