### 2025-05-11 b7b3b37
feat: add tabpfn regr learner
