+++ R/learner_tabpfn_regr_tabpfn.R
@@ -1,0 +1,227 @@
+#' @title TabPFN Regression Learner
+#' @author b-zhou
+#' @name mlr_learners_regr.tabpfn
+#'
+#' @description
+#' Foundation model for tabular data.
+#' Uses \CRANpkg{reticulate} to interface with the [`tabpfn`](https://github.com/PriorLabs/TabPFN) Python package.
+#'
+#' @section Installation:
+#' While the Python dependencies are handled via `reticulate::py_require()`, you can
+#' manually specify a virtual environment by calling `reticulate::use_virtualenv()`
+#' prior to calling the `$train()` function.
+#' In this virtual environment, the `tabpfn` package and its dependencies must be installed.
+#'
+#' @section Saving a Learner:
+#' In order to save a `LearnerRegrTabPFN` for later usage,
+#' it is necessary to call the `$marshal()` method on the `Learner`
+#' before writing it to disk, as the object will otherwise not be saved correctly.
+#' After loading a marshaled `LearnerRegrTabPFN` into R again,
+#' you then need to call `$unmarshal()` to transform it into a useable state.
+#'
+#' @section Initial parameter values:
+#' - `n_jobs` is initialized to 1 to avoid threading conflicts with \CRANpkg{future}.#'
+#'
+#' @section Custom mlr3 parameters:
+#' 
+#' - `output_type` corresponds to the same argument of the `.predict()` method of the `TabPFNRegressor` class, but only supports the options `"mean"`, `"median"` and `"mode"`.
+#'   The point predictions are stored as `$response` of the prediction object.
+#'
+#' - `categorical_feature_indices` uses R indexing instead of zero-based Python indexing.
+#'
+#' - `device` must be a string.
+#'   If set to `"auto"`, the behavior is the same as original.
+#'   Otherwise, the string is passed as argument to `torch.device()` to create a device.
+#'
+#' - `inference_precision` must be `"auto"` or `"autocast"`.
+#'   Passing `torch.dtype` is currently not supported.
+#'
+#' - `inference_config` is currently not supported.
+#'
+#' @template learner
+#'
+#' @references
+#' `r format_bib("hollmann2025tabpfn", "hollmann2023tabpfn")`
+#'
+#' @template seealso_learner
+#' @template example
+#' @export
+LearnerRegrTabPFN = R6Class("LearnerRegrTabPFN",
+  inherit = LearnerRegr,
+  public = list(
+    #' @description
+    #' Creates a new instance of this [R6][R6::R6Class] class.
+    initialize = function() {
+      ps = ps(
+        output_type = p_fct(
+          c("mean", "median", "mode"),
+          default = "mean",
+          tags = "predict"
+        ),
+        n_estimators = p_int(lower = 1, default = 4, tags = "train"),
+        categorical_features_indices = p_uty(tags = "train", custom_check = function(x) {
+          # R indexing is used
+          check_integerish(x, lower = 1, any.missing = FALSE, min.len = 1)
+        }),
+        softmax_temperature = p_dbl(default = 0.9, lower = 0, tags = "train"),
+        balance_probabilities = p_lgl(default = FALSE, tags = "train"),
+        average_before_softmax = p_lgl(default = FALSE, tags = "train"),
+        model_path = p_uty(default = "auto", tags = "train", custom_check = check_string),
+        device = p_uty(default = "auto", tags = "train", custom_check = check_string),
+        ignore_pretraining_limits = p_lgl(default = FALSE, tags = "train"),
+        inference_precision = p_fct(
+          # torch.dtype option is currently not supported
+          c("auto", "autocast"),
+          default = "auto",
+          tags = "train"
[…]
