### 2025-05-11 fb071a1
doc: update docs for tabpfn
