### 2025-05-12 585ddf0
Update NEWS.md
Co-authored-by: Sebastian Fischer <sebf.fischer@gmail.com>
