+++ NEWS.md
@@ -1,7 +1,7 @@
 # dev
 
 * feat: Support offset in learners `regr|classif.mgcv`, `regr.glm` and `regr.lmer`.
-* feat: Add `LearnerClassifTabPFN`.
+* feat: Add `LearnerClassifTabPFN` and `LearnerRegrTabPFN`.
 
 # mlr3extralearners 1.0.0
 
