### 2025-05-12 5a364dd
Update R/learner_tabpfn_classif_tabpfn.R
Co-authored-by: Sebastian Fischer <sebf.fischer@gmail.com>
