### 2025-05-12 b9ee603
Update R/learner_tabpfn_classif_tabpfn.R
Co-authored-by: Sebastian Fischer <sebf.fischer@gmail.com>
