### 2025-05-12 b61f233
Update R/learner_tabpfn_classif_tabpfn.R
Co-authored-by: Sebastian Fischer <sebf.fischer@gmail.com>
