+++ R/learner_tabpfn_classif_tabpfn.R
@@ -123,7 +123,6 @@ LearnerClassifTabPFN = R6Class("LearnerClassifTabPFN",
   private = list(
     .train = function(task) {
       reticulate::py_require(c("torch", "tabpfn"))
-      reticulate::py_require("tabpfn")
       tabpfn = reticulate::import("tabpfn")
 
       pars = self$param_set$get_values(tags = "train")
