### 2025-05-12 1086e60
feat: Learner for quantile regression GAMs for multiple quantiles (`mqgam`) (#422)
* mqgam regression learner

* fix examples in qgam learner

* fix style issues

* small fixes

* fix example

* ...

* ...

* ...

---------

Co-authored-by: be-marc <marcbecker@posteo.de>
