+++ dev/pkgdown.yml
@@ -1,12 +1,12 @@
 pandoc: 3.1.11
-pkgdown: 2.1.1
+pkgdown: 2.1.2
 pkgdown_sha: ~
 articles:
   common_issues: common_issues.html
   extending: extending.html
   articles/learner_status: learner_status.html
   articles/list_learners: list_learners.html
-last_built: 2025-04-16T15:36Z
+last_built: 2025-05-12T15:51Z
 urls:
   reference: https://mlr3extralearners.mlr-org.com/reference
   article: https://mlr3extralearners.mlr-org.com/articles
