+++ tests/testthat/test_paramtest_klaR_classif_rda.R
@@ -1,0 +1,33 @@
+test_that("classif.rda train", {
+  learner = lrn("classif.rda")
+  fun = klaR:::rda.default
+  exclude = c(
+    "x",
+    "object", # handled internally
+    "data", # handled internally
+    "grouping",  # optional vector specifying the class for each observation - not implemented
+    "prior",  # optional prior probabilities for the classes - not implemented
+    "regularization",  # handled internally by klaR
+    "startsimplex"  # optional starting simplex for the Nelder-Mead-minimization- not implemented
+  )
+
+  # note that you can also pass a list of functions in case $.train calls more than one
+  # function, e.g. for control arguments
+  paramtest = run_paramtest(learner, fun, exclude, tag = "train")
+  expect_paramtest(paramtest)
+})
+
+test_that("classif.rda predict", {
+  learner = lrn("classif.rda")
+  fun = klaR:::predict.rda # nolint
+  exclude = c(
+    "object", # handled internally
+    "data", # handled internally
+    "newdata", # handled internally
+    "posterior",  # must be set to True
+    "aslist"  # must be set to True
+  )
+
+  paramtest = run_paramtest(learner, fun, exclude, tag = "predict")
+  expect_paramtest(paramtest)
+})
