+++ man/mlr_learners_classif.rda.Rd
@@ -1,54 +1,60 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/learner_Cubist_regr_cubist.R
-\name{mlr_learners_regr.cubist}
-\alias{mlr_learners_regr.cubist}
-\alias{LearnerRegrCubist}
-\title{Regression Cubist Learner}
+% Please edit documentation in R/learner_klaR_classif_rda.R
+\name{mlr_learners_classif.rda}
+\alias{mlr_learners_classif.rda}
+\alias{LearnerClassifRda}
+\title{Classification Discriminant Analysis Learner}
 \description{
-Rule-based model that is an extension of Quinlan's M5 model tree. Each tree contains
-linear regression models at the terminal leaves.
-Calls \code{\link[Cubist:cubist]{Cubist::cubist()}} from \CRANpkg{Cubist}.
+Regularized Discriminant Analysis.
+Calls \code{\link[klaR:rda]{klaR::rda()}} from: \CRANpkg{klaR}.
 }
 \section{Dictionary}{
 
 This \link[mlr3:Learner]{Learner} can be instantiated via \link[mlr3:mlr_sugar]{lrn()}:
 
-\if{html}{\out{<div class="sourceCode">}}\preformatted{lrn("regr.cubist")
+\if{html}{\out{<div class="sourceCode">}}\preformatted{lrn("classif.rda")
 }\if{html}{\out{</div>}}
 }
 
 \section{Meta Information}{
 
 \itemize{
-\item Task type: \dQuote{regr}
-\item Predict Types: \dQuote{response}
-\item Feature Types: \dQuote{integer}, \dQuote{numeric}, \dQuote{character}, \dQuote{factor}, \dQuote{ordered}
-\item Required Packages: \CRANpkg{mlr3}, \CRANpkg{mlr3extralearners}, \CRANpkg{Cubist}
+\item Task type: \dQuote{classif}
+\item Predict Types: \dQuote{response}, \dQuote{prob}
+\item Feature Types: \dQuote{logical}, \dQuote{integer}, \dQuote{numeric}, \dQuote{factor}, \dQuote{ordered}
+\item Required Packages: \CRANpkg{mlr3}, \CRANpkg{klaR}
 }
 }
 
 \section{Parameters}{
 \tabular{lllll}{
    Id \tab Type \tab Default \tab Levels \tab Range \cr
-   committees \tab integer \tab - \tab  \tab \eqn{[1, 100]}{[1, 100]} \cr
-   unbiased \tab logical \tab FALSE \tab TRUE, FALSE \tab - \cr
-   rules \tab integer \tab 100 \tab  \tab \eqn{[1, \infty)}{[1, Inf)} \cr
-   extrapolation \tab numeric \tab 100 \tab  \tab \eqn{[0, 100]}{[0, 100]} \cr
-   sample \tab integer \tab 0 \tab  \tab \eqn{[0, \infty)}{[0, Inf)} \cr
-   seed \tab integer \tab - \tab  \tab \eqn{(-\infty, \infty)}{(-Inf, Inf)} \cr
-   label \tab untyped \tab "outcome" \tab  \tab - \cr
-   neighbors \tab integer \tab - \tab  \tab \eqn{[0, 9]}{[0, 9]} \cr
+   lambda \tab numeric \tab - \tab  \tab \eqn{[0, 1]}{[0, 1]} \cr
+   gamma \tab numeric \tab - \tab  \tab \eqn{[0, 1]}{[0, 1]} \cr
+   crossval \tab logical \tab TRUE \tab TRUE, FALSE \tab - \cr
[…]
