+++ DESCRIPTION
@@ -61,6 +61,7 @@ Suggests:
     jsonlite,
     keras (>= 2.3.0),
     kernlab,
+    klaR,
     knitr,
     ks,
     LiblineaR,
