+++ NAMESPACE
@@ -43,6 +43,7 @@ export(LearnerClassifRandomForestSRC)
 export(LearnerClassifRandomForestWeka)
 export(LearnerClassifRandomPlantedForest)
 export(LearnerClassifRandomTree)
+export(LearnerClassifRda)
 export(LearnerClassifSGD)
 export(LearnerClassifSMO)
 export(LearnerClassifSimpleLogistic)
