+++ R/bibentries.R
@@ -712,6 +712,31 @@ bibentries = c( # nolint start
     title = "Large margin classification using the perceptron algorithm",
     year = "1998"
   ),
+  friedman1989rda = bibentry("article",
+   ISSN = "01621459, 1537274X",
+   URL = "http://www.jstor.org/stable/2289860}",
+   author = "Jerome H. Friedman",
+   journal = "Journal of the American Statistical Association",
+   number = "405",
+   pages = "165--175",
+   publisher = "American Statistical Association, Taylor & Francis, Ltd.",
+   title = "Regularized Discriminant Analysis",
+   volume = "84",
+   year = "1989"
+  ),
+  ahdesmaki2010sda = bibentry("article",
+    title="Feature selection in omics prediction problems using cat scores and false nondiscovery rate control",
+    volume="4",
+    ISSN="1932-6157",
+    url="http://dx.doi.org/10.1214/09-AOAS277",
+    DOI="10.1214/09-aoas277",
+    number="1",
+    journal="The Annals of Applied Statistics",
+    publisher="Institute of Mathematical Statistics",
+    author="Ahdesmäki, Miika and Strimmer, Korbinian",
+    year="2010",
+    month="03"
+  ),
   faisolo2017qgam = bibentry("article",
     title = "Fast Calibrated Additive Quantile Regression",
     author = "Fasiolo, Matteo and Wood, Simon N. and Zaffran, Margaux and Nedellec, Raphael and Goude, Yannig",
