+++ tests/testthat/test_sda_classif_sda.R
@@ -1,7 +1,7 @@
 test_that("autotest", {
-  learner = lrn("classif.rda")
+  learner = lrn("classif.sda")
   expect_learner(learner)
   # note that you can skip tests using the exclude argument
-  result = run_autotest(learner, exclude = c("utf8_feature_names"))  # ä + ẞ problematic
+  result = run_autotest(learner)
   expect_true(result, info = result$error)
 })
