+++ man/mlr_learners_classif.sda.Rd
@@ -1,18 +1,18 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/learner_klaR_classif_rda.R
-\name{mlr_learners_classif.rda}
-\alias{mlr_learners_classif.rda}
-\alias{LearnerClassifRda}
-\title{Classification Discriminant Analysis Learner}
+% Please edit documentation in R/learner_sda_classif_sda.R
+\name{mlr_learners_classif.sda}
+\alias{mlr_learners_classif.sda}
+\alias{LearnerClassifSda}
+\title{Shrinkage Discriminant Analysis}
 \description{
-Regularized Discriminant Analysis.
-Calls \code{\link[klaR:rda]{klaR::rda()}} from: \CRANpkg{klaR}.
+Shrinkage Discriminant Analysis for classification.
+Calls \code{\link[sda:sda]{sda::sda()}} from \CRANpkg{sda}.
 }
 \section{Dictionary}{
 
 This \link[mlr3:Learner]{Learner} can be instantiated via \link[mlr3:mlr_sugar]{lrn()}:
 
-\if{html}{\out{<div class="sourceCode">}}\preformatted{lrn("classif.rda")
+\if{html}{\out{<div class="sourceCode">}}\preformatted{lrn("classif.sda")
 }\if{html}{\out{</div>}}
 }
 
@@ -21,8 +21,8 @@ This \link[mlr3:Learner]{Learner} can be instantiated via \link[mlr3:mlr_sugar]{
 \itemize{
 \item Task type: \dQuote{classif}
 \item Predict Types: \dQuote{response}, \dQuote{prob}
-\item Feature Types: \dQuote{logical}, \dQuote{integer}, \dQuote{numeric}, \dQuote{factor}, \dQuote{ordered}
-\item Required Packages: \CRANpkg{mlr3}, \CRANpkg{klaR}
+\item Feature Types: \dQuote{integer}, \dQuote{numeric}, \dQuote{factor}, \dQuote{ordered}
+\item Required Packages: \CRANpkg{mlr3}, \CRANpkg{sda}
 }
 }
 
@@ -30,28 +30,17 @@ This \link[mlr3:Learner]{Learner} can be instantiated via \link[mlr3:mlr_sugar]{
 \tabular{lllll}{
    Id \tab Type \tab Default \tab Levels \tab Range \cr
    lambda \tab numeric \tab - \tab  \tab \eqn{[0, 1]}{[0, 1]} \cr
-   gamma \tab numeric \tab - \tab  \tab \eqn{[0, 1]}{[0, 1]} \cr
-   crossval \tab logical \tab TRUE \tab TRUE, FALSE \tab - \cr
-   fold \tab integer \tab 10 \tab  \tab \eqn{[1, \infty)}{[1, Inf)} \cr
-   train.fraction \tab numeric \tab 0.5 \tab  \tab \eqn{[0, 1]}{[0, 1]} \cr
-   output \tab logical \tab FALSE \tab TRUE, FALSE \tab - \cr
-   schedule \tab integer \tab 2 \tab  \tab \eqn{(-\infty, \infty)}{(-Inf, Inf)} \cr
-   T.start \tab numeric \tab 0.1 \tab  \tab \eqn{[0, \infty)}{[0, Inf)} \cr
-   halflife \tab numeric \tab 50 \tab  \tab \eqn{[0, \infty)}{[0, Inf)} \cr
-   zero.temp \tab numeric \tab 0.01 \tab  \tab \eqn{[0, \infty)}{[0, Inf)} \cr
-   alpha \tab numeric \tab 2 \tab  \tab \eqn{[1, \infty)}{[1, Inf)} \cr
[…]
