+++ NAMESPACE
@@ -46,6 +46,7 @@ export(LearnerClassifRandomTree)
 export(LearnerClassifRda)
 export(LearnerClassifSGD)
 export(LearnerClassifSMO)
+export(LearnerClassifSda)
 export(LearnerClassifSimpleLogistic)
 export(LearnerClassifVotedPerceptron)
 export(LearnerDensKDEks)
