+++ DESCRIPTION
@@ -100,6 +100,7 @@ Suggests:
     RWeka,
     sandwich,
     set6,
+    sda,
     sm,
     stats,
     survival,
