### 2025-05-18 0a2e563
refactor: use same class for marshaled models
