+++ R/learner_tabpfn_regr_tabpfn.R
@@ -169,7 +169,7 @@ LearnerRegrTabPFN = R6Class("LearnerRegrTabPFN",
       # fit model
       fitted = mlr3misc::invoke(regressor$fit, X = x_py, y = y_py)
 
-      structure(list(fitted = fitted), class = "regr.tabpfn_model")
+      structure(list(fitted = fitted), class = "tabpfn_model")
     },
 
     .predict = function(task) {
@@ -205,7 +205,7 @@ LearnerRegrTabPFN = R6Class("LearnerRegrTabPFN",
 
 
 #' @export
-marshal_model.regr.tabpfn_model = function(model, inplace = FALSE, ...) {
+marshal_model.tabpfn_model = function(model, inplace = FALSE, ...) {
   # pickle should be available in any python environment
   pickle = reticulate::import("pickle")
   # save model as bytes
@@ -216,13 +216,13 @@ marshal_model.regr.tabpfn_model = function(model, inplace = FALSE, ...) {
   structure(list(
     marshaled = pickled,
     packages = "mlr3extralearners"
-  ), class = c("regr.tabpfn_model_marshaled", "marshaled"))
+  ), class = c("tabpfn_model_marshaled", "marshaled"))
 }
 
 #' @export
-unmarshal_model.regr.tabpfn_model_marshaled = function(model, inplace = FALSE, ...) {
+unmarshal_model.tabpfn_model_marshaled = function(model, inplace = FALSE, ...) {
   pickle = reticulate::import("pickle")
   # unpickle
   fitted = pickle$loads(reticulate::r_to_py(model$marshaled))
-  structure(list(fitted = fitted), class = "regr.tabpfn_model")
+  structure(list(fitted = fitted), class = "tabpfn_model")
 }
