### 2025-05-18 e9ee9a4
docs: use templates and inheritance
