+++ R/learner_tabpfn_classif_tabpfn.R
@@ -5,23 +5,10 @@
 #' @description
 #' Foundation model for tabular data.
 #' Uses \CRANpkg{reticulate} to interface with the [`tabpfn`](https://github.com/PriorLabs/TabPFN) Python package.
-#'
-#' @section Installation:
-#' While the Python dependencies are handled via `reticulate::py_require()`, you can
-#' manually specify a virtual environment by calling `reticulate::use_virtualenv()`
-#' prior to calling the `$train()` function.
-#' In this virtual environment, the `tabpfn` package and its dependencies must be installed.
-#'
-#' @section Saving a Learner:
-#' In order to save a `LearnerClassifTabPFN` for later usage,
-#' it is necessary to call the `$marshal()` method on the `Learner`
-#' before writing it to disk, as the object will otherwise not be saved correctly.
-#' After loading a marshaled `LearnerClassifTabPFN` into R again,
-#' you then need to call `$unmarshal()` to transform it into a useable state.
-#'
-#' @section Initial parameter values:
-#' - `n_jobs` is initialized to 1 to avoid threading conflicts with \CRANpkg{future}.
-#'
+#' 
+#' @templateVar class LearnerClassifTabPFN
+#' @template sections_tabpfn
+#' 
 #' @section Custom mlr3 parameters:
 #'
 #' - `categorical_feature_indices` uses R indexing instead of zero-based Python indexing.
@@ -34,7 +21,8 @@
 #'   Passing `torch.dtype` is currently not supported.
 #'
 #' - `inference_config` is currently not supported.
-#'
+#' 
+#' @templateVar id classif.tabpfn
 #' @template learner
 #'
 #' @references
