+++ man/mlr_learners_regr.tabpfn.Rd
@@ -28,14 +28,15 @@ you then need to call \verb{$unmarshal()} to transform it into a useable state.
 \section{Initial parameter values}{
 
 \itemize{
-\item \code{n_jobs} is initialized to 1 to avoid threading conflicts with \CRANpkg{future}.#'
+\item \code{n_jobs} is initialized to 1 to avoid threading conflicts with \CRANpkg{future}.
 }
 }
 
 \section{Custom mlr3 parameters}{
 
 \itemize{
-\item \code{output_type} corresponds to the same argument of the \code{.predict()} method of the \code{TabPFNRegressor} class, but only supports the options \code{"mean"}, \code{"median"} and \code{"mode"}.
+\item \code{output_type} corresponds to the same argument of the \code{.predict()} method of the \code{TabPFNRegressor} class,
+but only supports the options \code{"mean"}, \code{"median"} and \code{"mode"}.
 The point predictions are stored as \verb{$response} of the prediction object.
 \item \code{categorical_feature_indices} uses R indexing instead of zero-based Python indexing.
 \item \code{device} must be a string.
@@ -49,15 +50,67 @@ Passing \code{torch.dtype} is currently not supported.
 
 \section{Dictionary}{
 
-This [Learner][mlr3::Learner] can be instantiated via [lrn()][mlr3::lrn]:
+This \link[mlr3:Learner]{Learner} can be instantiated via \link[mlr3:mlr_sugar]{lrn()}:
 
-\if{html}{\out{<div class="sourceCode">}}\preformatted{lrn("
-
-[mlr3::Learner]: R:mlr3::Learner
-[mlr3::lrn]: R:mlr3::lrn
+\if{html}{\out{<div class="sourceCode">}}\preformatted{lrn("regr.tabpfn")
 }\if{html}{\out{</div>}}
 }
 
+\section{Meta Information}{
+
+\itemize{
+\item Task type: \dQuote{regr}
+\item Predict Types: \dQuote{response}, \dQuote{quantiles}
+\item Feature Types: \dQuote{logical}, \dQuote{integer}, \dQuote{numeric}
+\item Required Packages: \CRANpkg{mlr3}
+}
+}
+
+\section{Parameters}{
+\tabular{lllll}{
+   Id \tab Type \tab Default \tab Levels \tab Range \cr
+   output_type \tab character \tab mean \tab mean, median, mode \tab - \cr
+   n_estimators \tab integer \tab 4 \tab  \tab \eqn{[1, \infty)}{[1, Inf)} \cr
+   categorical_features_indices \tab untyped \tab - \tab  \tab - \cr
+   softmax_temperature \tab numeric \tab 0.9 \tab  \tab \eqn{[0, \infty)}{[0, Inf)} \cr
+   balance_probabilities \tab logical \tab FALSE \tab TRUE, FALSE \tab - \cr
+   average_before_softmax \tab logical \tab FALSE \tab TRUE, FALSE \tab - \cr
+   model_path \tab untyped \tab "auto" \tab  \tab - \cr
+   device \tab untyped \tab "auto" \tab  \tab - \cr
+   ignore_pretraining_limits \tab logical \tab FALSE \tab TRUE, FALSE \tab - \cr
+   inference_precision \tab character \tab auto \tab auto, autocast \tab - \cr
+   fit_mode \tab character \tab fit_preprocessors \tab low_memory, fit_preprocessors, fit_with_cache \tab - \cr
+   memory_saving_mode \tab untyped \tab "auto" \tab  \tab - \cr
+   random_state \tab integer \tab 0 \tab  \tab \eqn{(-\infty, \infty)}{(-Inf, Inf)} \cr
[…]

[…]
