+++ attic/scratch_to_be_deleted_v5.R
@@ -1,0 +1,184 @@
+#' @title Classification Neural Network Learner
+#' @author annanzrv
+#' @name mlr_learners_classif.tester
+#'
+#' @description
+#' Simple and fast neural nets for tabular data classification.
+#' Calls [fastai::fastai()] from package \CRANpkg{fastai}.
+#'
+#' @section Initial parameter values:
+#' - `n_epoch`:
+#'   Needs to be set for [fastai::fit()] to work.
+#'   If no value is given, it is set to 5.
+#' - `eval_metric`:
+#'   Needs to be set for [fastai::predict()] to work.
+#'   If no value is given, it is set to [mlr3measures::classif.logloss].
+#'
+#' @templateVar id classif.fastai
+#' @template learner
+#'
+#' @references
+#' `r format_bib(FIXME: ONE OR MORE REFERENCES FROM bibentries.R)`
+#'
+#' @template seealso_learner
+#' @template example
+#' @export
+LearnerClassifTester = R6Class("LearnerClassifTester",
+  inherit = LearnerClassif,
+  public = list(
+   eval_protocol = NULL,
+   #' @description
+   #' Creates a new instance of this [R6][R6::R6Class] class.
+   initialize = function() {
+     param_set = ps(
+       act_cls     = p_uty(tags = "train"),  # Activation type for LinBnDrop layers, e.g., fastai::nn()$ReLU(inplace = TRUE)
+       bn_cont     = p_lgl(default = TRUE, tags = "train"),  # Use BatchNorm1d on continuous variables
+       bn_final    = p_lgl(default = FALSE, tags = "train"),  # Use BatchNorm1d on final layer
+       drop_last   = p_lgl(default = FALSE, tags = "train"),  #  If True, then the last incomplete batch is dropped.
+       embed_p     = p_dbl(lower = 0L, upper = 1L, default = 0L, tags = "train"),  # Dropout probability for Embedding layer
+       emb_szs     = p_uty(default=NULL, tags="train"),  # Sequence of (num_embeddings, embedding_dim) for each categorical variable
+       n_epoch     = p_int(lower=1, default = 5, tags="train"), # Epochs
+       eval_metric = p_uty(tags = "train", custom_check = crate({function(x) check_true(any(is.character(x), is.function(x), inherits(x, "Measure")))})),
+       layers      = p_uty(tags="train"),  # Sequence of ints used to specify the input and output size of each LinBnDrop layer
+       loss_func   = p_uty(tags="train"),  # Defaults to fastai::CrossEntropyLossFlat()
+       lr          = p_dbl(lower=0, default = 0.001, tags = "train"),  # Learning rate
+       metrics     = p_uty(tags = "train"),  # optional list of metrics, e.g, fastai::Precision() or fastai::accuracy()
+       n_out       = p_int(tags="train"),  # ?
+       num_workers = p_int(default = 0L, tags = "train"),  # how many subprocesses to use for data loading
+       opt_func    = p_uty(tags="train"),  # Optimizer created when Learner.fit is called. E.g., fastai::Adam()
+       patience    = p_int(1L, default = 1, tags = c("train", "internal_validation")),  # number of epochs to wait when training has not improved model. add `depends = quote(early_stopping == TRUE`)`
+       pin_memory  = p_lgl(default = TRUE, tags = "train"),  # If True, the data loader will copy Tensors into CUDA pinned memory before returning them.
+       procs       = p_uty(default=NULL, tags="train"),  # fastai preprocessing steps such as fastai::Categorify, fastai::Normalize, fastai::fill_missing
+       ps          = p_uty(default = NULL, tags = "train"),  # Sequence of dropout probabilities
[…]
