+++ attic/scratch_to_be_deleted.R
@@ -168,45 +168,50 @@ pred$score(mlr3::msr("classif.ce"))
 learner$eval_protocol
 
 
+learner = lrn("classif.fastai")
+tasks = generate_tasks(learner)
 
-
-task = tsk("boston_housing")
-ids = partition(task)
+task = tasks$sanity
 df = task$data()
-# numeric_features = c(task$feature_types[type %in% c("numeric", "integer"), id], task$target_names)
-# df = task$data()
-# df = df[, ..numeric_features]
 
 cat_cols = task$feature_types[type != "numeric", id]
 num_cols = task$feature_types[type == "numeric", id]
 
-df_fai <- invoke(
+df_fai_n <- invoke(
   fastai::TabularDataTable,
   df = df,
   cat_names = cat_cols,
   cont_names = num_cols,
   y_names = task$target_names,
-  procs = list(fastai::Categorify()),
-  splits = ids
+  splits = NULL,
+  #procs = list(fastai::Categorify(), fastai::Normalize())
 )
 
-dl <- invoke(
+dl_n <- invoke(
   fastai::dataloaders,
-  df_fai
+  df_fai_n
 )
 
-tab_learner <- fastai::tabular_learner(dl, metrics = list(fastai::msle()))
-#invisible(tab_learner$remove_cb(tab_learner$progress))
+config_n = invoke(
+  fastai::tabular_config
+)
+tab_learner_n <- fastai::tabular_learner(dl, layers = NULL, metrics = fastai::accuracy(), config = config)
+invisible(tab_learner_n$remove_cb(tab_learner_n$progress))
+fit_eval = function(tab_learner_n) {
+  eval_protocol = invoke(
+    fastai::fit,
+    object = tab_learner_n,
+    n_epoch = 5
+  )
+  eval_protocol
+}
+# to prevent python from printing evaluation protocol do:
+eval =invisible(reticulate::py_capture_output(fit_eval(tab_learner)))
 
-eval = fastai::fit(tab_learner, 5)
+pred = predict(tab_learner, ordered_features(task, learner))
 
-test_dl = object$dls$test_dl(row)
-tab_learner$get_preds(dl = test_dl, with_decoded = FALSE)
-
-pred = predict(tab_learner, df)
-
-
-
+class_labels = levels(task$truth())
+class_labels[pred$class+1]
 
 
 
