### 2025-05-19 afe022d
Update description for klaR learner
