### 2025-05-19 8339f3d
clean scratch files
