### 2025-05-19 e632ee0
update description for rda learner
