+++ R/learner_qgam_regr_mqgam.R
@@ -97,10 +97,7 @@ LearnerRegrMQGam = R6Class("LearnerRegrMQGam",
 
       arg_gam_pars = pars$argGam
       pars = pars[names(pars) %nin% c("argGam", "link")]
-
-      if ("weights" %in% task$properties) {
-        arg_gam_pars = insert_named(arg_gam_pars, list(weights = task$weights$weight))
-      }
+      arg_gam_pars$weights = get_weights(task, private)
 
       if (is.null(pars$form)) {
         form = stats::reformulate(task$feature_names, response = task$target_names)
