+++ R/learner_sm_dens_nonpar.R
@@ -50,10 +50,7 @@ LearnerDensNonparametric = R6Class("LearnerDensNonparametric",
   private = list(
     .train = function(task) {
       pars = self$param_set$get_values(tags = "train")
-
-      if ("weights" %in% task$properties) {
-        pars$weights = task$weights$weight
-      }
+      pars$weights = get_weights(task, private)
 
       # TODO: Why is it done that way??
       pdf = function(x) {} # nolint
