+++ R/learner_lme4_regr_lmer.R
@@ -162,10 +162,7 @@ LearnerRegrLmer = R6Class("LearnerRegrLmer",
       # formula must be set manually to use the lme4 mixed effects syntax
       formula = pars_train$formula
       pars_train[["formula"]] = NULL
-
-      if ("weights" %in% task$properties) {
-        pars_train$weights = task$weights$weight
-      }
+      pars_train$weights = get_weights(task, private)
 
       if ("offset" %in% task$properties) {
         pars_train$offset = task$offset$offset
