+++ R/learner_glmnet_surv_glmnet.R
@@ -137,9 +137,7 @@ LearnerSurvGlmnet = R6Class("LearnerSurvGlmnet",
       target = task$truth()
       pv = self$param_set$get_values(tags = "train")
       pv$family = "cox"
-      if ("weights" %in% task$properties) {
-        pv$weights = task$weights$weight
-      }
+      pv$weights = get_weights(task, private)
       pv = glmnet_set_offset(task, "train", pv)
 
       list(
@@ -160,7 +158,7 @@ LearnerSurvGlmnet = R6Class("LearnerSurvGlmnet",
       pv = glmnet_set_offset(task, "predict", pv)
 
       # get survival matrix
-      fit = invoke(survival::survfit, formula = self$model$model,
+      fit = invoke(survival::survfit,formula = self$model$model,
                    x = self$model$x, y = self$model$y,
                    weights = self$model$weights, offset = self$model$offset,
                    newx = newdata, se.fit = FALSE, .args = pv)
