### 2025-05-19 7b37512
fix weight parameter name
