+++ R/learner_abess_regr_abess.R
@@ -85,7 +85,7 @@ LearnerRegrAbess = R6Class("LearnerRegrAbess",
     .train = function(task) {
       # get parameters for training
       pars = self$param_set$get_values(tags = "train")
-      pars$weights = get_weights(task, private)
+      pars$weight = get_weights(task, private)
 
       invoke(
         abess::abess,
