### 2025-05-19 6b812e8
add weight property
