+++ R/learner_Cubist_regr_cubist.R
@@ -43,6 +43,7 @@ LearnerRegrCubist = R6Class("LearnerRegrCubist",
         feature_types = c("integer", "numeric", "character", "factor", "ordered"),
         predict_types = "response",
         param_set = param_set,
+        properties = "weights",
         man = "mlr3extralearners::mlr_learners_regr.cubist",
         label = "Rule-based model"
       )
