### 2025-05-19 5b69afc
remove weights if not supported by learner
