### 2025-05-20 87bc70f
fix: use mlr3 release version (#429)
Also fixes the issue with the vignette rendering
