### 2025-05-21 43bd6fc
fix typo in bibentries
