### 2025-05-21 51c46b5
added new sparsedisrcrim diagonal lda learner
