+++ DESCRIPTION
@@ -22,7 +22,8 @@ Authors@R: c(
            comment = c(ORCID = "0000-0002-3609-8674")),
     person("Lukas", "Burk", , "burk@leibniz-bips.de", role = "ctb",
            comment = c(ORCID = "0000-0001-7528-3795")),
-    person("Lona", "Koers", , "lona.koers@gmail.com", role = "ctb")
+    person("Lona", "Koers", , "lona.koers@gmail.com", role = "ctb"),
+    person("Anna", "Nazarova", , "anna.nazarova@outlook.com", role = "ctb")
   )
 Description: Extra learners for use in mlr3.
 License: LGPL-3
@@ -100,6 +101,7 @@ Suggests:
     sandwich,
     set6,
     sm,
+    sparsediscrim,
     stats,
     survival,
     survivalmodels (>= 0.1.19),
