### 2025-05-21 d9fedda
fix encoding in bibentries
