### 2025-05-21 b4d17c0
add diagonal lda learner from sparsediscrim
