+++ tests/testthat/test_sparsediscrim_classif_dlda.R
@@ -1,0 +1,7 @@
+test_that("autotest", {
+  learner = lrn("classif.dlda")
+  expect_learner(learner, check_man = FALSE)
+  # note that you can skip tests using the exclude argument
+  result = run_autotest(learner)
+  expect_true(result, info = result$error)
+})
