### 2025-05-21 cb44936
adding shrinkage diagonal lda learner and min distance bayesian learner from sparsediscrim
