+++ tests/testthat/test_sparsediscrim_classif_sdlda.R
@@ -1,5 +1,5 @@
 test_that("autotest", {
-  learner = lrn("classif.dlda")
+  learner = lrn("classif.sdlda")
   expect_learner(learner)
   # note that you can skip tests using the exclude argument
   result = run_autotest(learner)
