+++ R/bibentries.R
@@ -723,6 +723,29 @@ bibentries = c( # nolint start
     publisher = "ASA Websites",
     doi = "10.1198/016214502753479248"
   ),
+  Srivastava2007mdeb = bibentry("article",
+    title = "Comparison of Discrimination Methods for High Dimensional Data",
+    author = "Srivastava, M. and Kubokawa, T.",
+    year = "2007",
+    journal = "Journal of the Japanese Statistical Association",
+    volume = "37",
+    number = "1",
+    pages = "123--134"
+  ),
+  Peng2009sdlda = bibentry("article",
+    ISSN = "0006341X, 15410420",
+    URL = "http://www.jstor.org/stable/20640622",
+    abstract = "High-dimensional data such as microarrays have brought us new statistical challenges. For example, using a large number of genes to classify samples based on a small number of microarrays remains a difficult problem. Diagonal discriminant analysis, support vector machines, and k-nearest neighbor have been suggested as among the best methods for small sample size situations, but none was found to be superior to others. In this article, we propose an improved diagonal discriminant approach through shrinkage and regularization of the variances. The performance of our new approach along with the existing methods is studied through simulations and applications to real data. These studies show that the proposed shrinkage-based and regularization diagonal discriminant methods have lower misclassification rates than existing methods in many cases.",
+    author = "Herbert Pang and Tiejun Tong and Hongyu Zhao",
+    journal = "Biometrics",
+    number = "4",
+    pages = "1021--1029",
+    publisher = "[Wiley, International Biometric Society]",
+    title = "Shrinkage-Based Diagonal Discriminant Analysis and Its Applications in High-Dimensional Data",
+    urldate = "2025-05-21",
+    volume = "65",
+    year = "2009"
+  ),
   faisolo2017qgam = bibentry("article",
     title = "Fast Calibrated Additive Quantile Regression",
     author = "Fasiolo, Matteo and Wood, Simon N. and Zaffran, Margaux and Nedellec, Raphael and Goude, Yannig",
