+++ tests/testthat/test_paramtest_sparsediscrim_classif_sdlda.R
@@ -1,9 +1,7 @@
-test_that("classif.dlda train", {
-  learner = lrn("classif.dlda")
-  fun = sparsediscrim:::lda_diag.formula
+test_that("classif.sdlda train", {
+  learner = lrn("classif.sdlda")
+  fun = sparsediscrim:::lda_shrink_cov.formula
   exclude = c(
-    "x",
-    "y",
     "object", # handled internally
     "data", # handled internally
     "formula" # handled internally
@@ -15,16 +13,15 @@ test_that("classif.dlda train", {
   expect_paramtest(paramtest)
 })
 
-test_that("classif.dlda predict", {
-  learner = lrn("classif.dlda")
-  fun = sparsediscrim:::predict.lda_diag # nolint
+test_that("classif.sdlda predict", {
+  learner = lrn("classif.sdlda")
+  fun = sparsediscrim:::predict.lda_shrink_cov # nolint
   exclude = c(
     "object", # handled internally
     "data", # handled internally
     "newdata", # handled internally
     "type" # handled internally
   )
-
   paramtest = run_paramtest(learner, fun, exclude, tag = "predict")
   expect_paramtest(paramtest)
 })
