+++ R/learner_sparsediscrim_classif_sdlda.R
@@ -20,26 +20,26 @@
 LearnerClassifSdlda = R6Class("LearnerClassifSdlda",
   inherit = LearnerClassif,
   public = list(
-  #' @description
-  #' Creates a new instance of this [R6][R6::R6Class] class.
-  initialize = function() {
+    #' @description
+    #' Creates a new instance of this [R6][R6::R6Class] class.
+    initialize = function() {
     # FIXME: MANUALLY ADD PARAMETERS BELOW AND THEN DELETE THIS LINE
-    param_set = ps(
-      num_alphas = p_int(default = 101, tags = "train"),
-      prior     = p_uty(default = NULL, tags = "train")
-    )
+      param_set = ps(
+        num_alphas = p_int(default = 101, tags = "train"),
+        prior      = p_uty(default = NULL, tags = "train")
+      )
 
-    super$initialize(
-      id = "classif.sdlda",
-      packages = "sparsediscrim",
-      feature_types = c("integer", "numeric"),
-      predict_types = c("response", "prob"),
-      param_set = param_set,
-      properties = c("multiclass", "twoclass"),
-      man = "mlr3extralearners::mlr_learners_classif.sdlda",
-      label = "Shrinkage-based Diagonal Linear Discriminant Analysis"
-     )
-   }
+      super$initialize(
+        id = "classif.sdlda",
+        packages = "sparsediscrim",
+        feature_types = c("integer", "numeric"),
+        predict_types = c("response", "prob"),
+        param_set = param_set,
+        properties = c("multiclass", "twoclass"),
+        man = "mlr3extralearners::mlr_learners_classif.sdlda",
+        label = "Shrinkage-based Diagonal Linear Discriminant Analysis"
+      )
+    }
   ),
   private = list(
     .train = function(task) {
