### 2025-05-25 6462d56
refactor: use init argument
