+++ R/learner_tabpfn_regr_tabpfn.R
@@ -74,11 +74,8 @@ LearnerRegrTabPFN = R6Class("LearnerRegrTabPFN",
           }
         }),
         random_state = p_int(default = 0, tags = "train"),
-        n_jobs = p_int(lower = 1, default = 1, special_vals = list(-1), tags = "train")
+        n_jobs = p_int(lower = 1, init = 1, special_vals = list(-1), tags = "train")
       )
-
-      # set n_jobs to 1 as mlr3 has its own parallelization
-      ps$values$n_jobs = 1
 
       super$initialize(
         id = "regr.tabpfn",
