+++ NEWS.md
@@ -1,8 +1,8 @@
 # dev
 
 * feat: Support offset in learners `regr|classif.mgcv`, `regr.glm` and `regr.lmer`.
-* feat: Added learner `LearnerRegrQGam` and `LearnerRegrMQGam`
-* feat: Added `LearnerClassifTabPFN` and `LearnerRegrTabPFN`.
+* feat: Added learners `LearnerRegrQGam` and `LearnerRegrMQGam`.
+* feat: Added learners `LearnerClassifTabPFN` and `LearnerRegrTabPFN`.
 
 # mlr3extralearners 1.0.0
 
