+++ R/learner_catboost_classif_catboost.R
@@ -272,7 +272,7 @@ LearnerClassifCatboost = R6Class("LearnerClassifCatboost",
       learn_pool = invoke(catboost::catboost.load_pool,
         data = task$data(cols = task$feature_names),
         label = label,
-        weight = task$weights$weight,
+        weight = get_weights(task, private),
         thread_count = self$param_set$values$thread_count)
 
       # early stopping
@@ -293,7 +293,7 @@ LearnerClassifCatboost = R6Class("LearnerClassifCatboost",
         invoke(catboost::catboost.load_pool,
           data = internal_valid_task$data(cols = internal_valid_task$feature_names),
           label = test_label,
-          weight = internal_valid_task$weights$weight,
+          weight = get_weights(internal_valid_task, private),
           thread_count = 1)
       }
 
