### 2025-05-26 78aef1f
remove get_weights helper
