+++ R/learner_partykit_regr_ctree.R
@@ -95,7 +95,7 @@ LearnerRegrCTree = R6Class("LearnerRegrCTree",
       pars_pargs = pars[names(pars) %in% formalArgs(mvtnorm::GenzBretz)]
       pars = pars[names(pars) %nin% formalArgs(mvtnorm::GenzBretz)]
       pars$pargs = invoke(mvtnorm::GenzBretz, pars_pargs)
-      pars$weights = get_weights(task, private)
+      pars$weights = private$.get_weights(task)
 
       invoke(partykit::ctree, formula = task$formula(),
         data = task$data(), .args = pars)
