+++ R/learner_catboost_regr_catboost.R
@@ -225,7 +225,7 @@ LearnerRegrCatboost = R6Class("LearnerRegrCatboost",
       learn_pool = invoke(catboost::catboost.load_pool,
         data = task$data(cols = task$feature_names),
         label = task$data(cols = task$target_names)[[1L]],
-        weight = get_weights(task, private),
+        weight = private$.get_weights(task),
         thread_count = self$param_set$values$thread_count)
 
       # early stopping
@@ -241,7 +241,7 @@ LearnerRegrCatboost = R6Class("LearnerRegrCatboost",
         invoke(catboost::catboost.load_pool,
           data = internal_valid_task$data(cols = internal_valid_task$feature_names),
           label = test_label,
-          weight = get_weights(internal_valid_task, private),
+          weight = private$.get_weights(internal_valid_task),
           thread_count = 1)
       }
 
