+++ R/learner_mgcv_regr_gam.R
@@ -104,7 +104,7 @@ LearnerRegrGam = R6Class("LearnerRegrGam",
       pars = pars[names(pars) %nin% formalArgs(mgcv::gam.control)]
 
       data = task$data(cols = c(task$feature_names, task$target_names))
-      pars$weights = get_weights(task, private)
+      pars$weights = private$.get_weights(task)
 
       if ("offset" %in% task$properties) {
         pars$offset = task$offset$offset
