+++ R/learner_mboost_surv_mboost.R
@@ -118,7 +118,7 @@ LearnerSurvMBoost = R6Class("LearnerSurvMBoost",
       }
 
       pars = self$param_set$get_values(tags = "train")
-      pars$weights = get_weights(task, private)
+      pars$weights = private$.get_weights(task)
 
       if ("offset" %in% task$properties) {
         pars$offset = task$offset$offset
