### 2025-05-26 f9a8cda
remove task$weights
