### 2025-05-27 f3f534e
feat: marshal xgboost
