### 2025-05-27 16b5314
refactor: move marshal functions to learner file
