### 2025-05-27 37f8736
chore: update namespace and rd file
