+++ man/mlr_learners_surv.xgboost.cox.Rd
@@ -28,6 +28,15 @@ By default, the Breslow estimator is used via \code{\link[mlr3proba:breslow]{mlr
 }
 }
 
+\section{Saving this learner}{
+
+In order to save a \code{LearnerSurvXgboostCox} for later usage,
+it is necessary to call the \verb{$marshal()} method on the \code{Learner}
+before writing it to disk, as the object will otherwise not be saved correctly.
+After loading a marshaled \code{LearnerSurvXgboostCox} into R again,
+you then need to call \verb{$unmarshal()} to transform it into a useable state.
+}
+
 \section{Initial parameter values}{
 
 \itemize{
@@ -190,6 +199,9 @@ Extracted from \code{model$evaluation_log}}
 
 \item{\code{validate}}{How to construct the internal validation data. This parameter can be either \code{NULL},
 a ratio, \code{"test"}, or \code{"predefined"}.}
+
+\item{\code{marshaled}}{(\code{logical(1)})\cr
+Whether the learner has been marshaled.}
 }
 \if{html}{\out{</div>}}
 }
@@ -198,6 +210,8 @@ a ratio, \code{"test"}, or \code{"predefined"}.}
 \itemize{
 \item \href{#method-LearnerSurvXgboostCox-new}{\code{LearnerSurvXgboostCox$new()}}
 \item \href{#method-LearnerSurvXgboostCox-importance}{\code{LearnerSurvXgboostCox$importance()}}
+\item \href{#method-LearnerSurvXgboostCox-marshal}{\code{LearnerSurvXgboostCox$marshal()}}
+\item \href{#method-LearnerSurvXgboostCox-unmarshal}{\code{LearnerSurvXgboostCox$unmarshal()}}
 \item \href{#method-LearnerSurvXgboostCox-clone}{\code{LearnerSurvXgboostCox$clone()}}
 }
 }
@@ -239,6 +253,42 @@ The importance scores are calculated with \code{\link[xgboost:xgb.importance]{xg
 
 \subsection{Returns}{
 Named \code{numeric()}.
+}
+}
+\if{html}{\out{<hr>}}
+\if{html}{\out{<a id="method-LearnerSurvXgboostCox-marshal"></a>}}
+\if{latex}{\out{\hypertarget{method-LearnerSurvXgboostCox-marshal}{}}}
+\subsection{Method \code{marshal()}}{
+Marshal the learner's model.
+\subsection{Usage}{
+\if{html}{\out{<div class="r">}}\preformatted{LearnerSurvXgboostCox$marshal(...)}\if{html}{\out{</div>}}
+}
+
+\subsection{Arguments}{
+\if{html}{\out{<div class="arguments">}}
+\describe{
+\item{\code{...}}{(any)\cr
+Additional arguments passed to \code{\link[=marshal_model]{marshal_model()}}.}
+}
+\if{html}{\out{</div>}}
+}
+}
+\if{html}{\out{<hr>}}
+\if{html}{\out{<a id="method-LearnerSurvXgboostCox-unmarshal"></a>}}
+\if{latex}{\out{\hypertarget{method-LearnerSurvXgboostCox-unmarshal}{}}}
+\subsection{Method \code{unmarshal()}}{
+Unmarshal the learner's model.
+\subsection{Usage}{
+\if{html}{\out{<div class="r">}}\preformatted{LearnerSurvXgboostCox$unmarshal(...)}\if{html}{\out{</div>}}
+}
+
+\subsection{Arguments}{
+\if{html}{\out{<div class="arguments">}}
+\describe{
+\item{\code{...}}{(any)\cr
+Additional arguments passed to \code{\link[=unmarshal_model]{unmarshal_model()}}.}
+}
+\if{html}{\out{</div>}}
 }
 }
 \if{html}{\out{<hr>}}
