### 2025-05-27 a9ccff3
test: add test for marshaling
