### 2025-05-28 37e1449
refactor: comments and style
