### 2025-05-28 b2f95e0
Update R/learner_xgboost_surv_xgboost_cox.R
Co-authored-by: Sebastian Fischer <sebf.fischer@gmail.com>
