### 2025-05-28 490d0be
Update R/learner_xgboost_surv_xgboost_cox.R
Co-authored-by: Sebastian Fischer <sebf.fischer@gmail.com>
