+++ R/learner_xgboost_surv_xgboost_cox.R
@@ -301,7 +301,6 @@ unmarshal_model.xgboost_cox_model_marshaled = function(model, ...) {
   tmp = tempfile(fileext = ".buffer")
   writeBin(model$train_data, tmp)
   train_data = xgboost::xgb.DMatrix(tmp)
-  # clean up
   unlink(tmp)
 
   structure(list(
