### 2025-05-28 be37920
Update R/learner_xgboost_surv_xgboost_cox.R
Co-authored-by: Sebastian Fischer <sebf.fischer@gmail.com>
