+++ dev/pkgdown.yml
@@ -1,12 +1,12 @@
 pandoc: 3.1.11
-pkgdown: 2.1.2
+pkgdown: 2.1.3
 pkgdown_sha: ~
 articles:
   common_issues: common_issues.html
   extending: extending.html
   articles/learner_status: learner_status.html
   articles/list_learners: list_learners.html
-last_built: 2025-05-20T07:05Z
+last_built: 2025-05-28T09:57Z
 urls:
   reference: https://mlr3extralearners.mlr-org.com/reference
   article: https://mlr3extralearners.mlr-org.com/articles
