### 2025-05-28 7848285
fix: Add marshaling for `surv.xgboost.cox` learner (#432)
