+++ R/learner_xgboost_surv_xgboost_cox.R
@@ -19,6 +19,13 @@
 #'
 #' @template note_xgboost
 #'
+#' @section Saving this learner:
+#' In order to save a `LearnerSurvXgboostCox` for later usage,
+#' it is necessary to call the `$marshal()` method on the `Learner`
+#' before writing it to disk, as the object will otherwise not be saved correctly.
+#' After loading a marshaled `LearnerSurvXgboostCox` into R again,
+#' you then need to call `$unmarshal()` to transform it into a useable state.
+#'
 #' @section Initial parameter values:
 #' - `nrounds` is initialized to 1000.
 #' - `nthread` is initialized to 1 to avoid conflicts with parallelization via \CRANpkg{future}.
@@ -127,7 +134,7 @@ LearnerSurvXgboostCox = R6Class("LearnerSurvXgboostCox",
         param_set = ps,
         predict_types = c("crank", "lp", "distr"),
         feature_types = c("integer", "numeric"),
-        properties = c("weights", "missings", "importance", "validation", "internal_tuning"),
+        properties = c("weights", "missings", "importance", "validation", "internal_tuning", "marshal"),
         packages = c("mlr3extralearners", "xgboost"),
         man = "mlr3extralearners::mlr_learners_surv.xgboost.cox",
         label = "Extreme Gradient Boosting Cox"
@@ -140,6 +147,21 @@ LearnerSurvXgboostCox = R6Class("LearnerSurvXgboostCox",
     #' @return Named `numeric()`.
     importance = function() {
       xgb_imp(self$model$model)
+    },
+
+    #' @description
+    #' Marshal the learner's model.
+    #' @param ... (any)\cr
+    #'   Additional arguments passed to [`marshal_model()`].
+    marshal = function(...) {
+      mlr3::learner_marshal(.learner = self, ...)
+    },
+    #' @description
+    #' Unmarshal the learner's model.
+    #' @param ... (any)\cr
+    #'   Additional arguments passed to [`unmarshal_model()`].
+    unmarshal = function(...) {
+      mlr3::learner_unmarshal(.learner = self, ...)
     }
   ),
 
@@ -163,6 +185,11 @@ LearnerSurvXgboostCox = R6Class("LearnerSurvXgboostCox",
         private$.validate = mlr3::assert_validate(rhs)
       }
       private$.validate
+    },
+    #' @field marshaled (`logical(1)`)\cr
+    #' Whether the learner has been marshaled.
+    marshaled = function() {
+      mlr3::learner_marshaled(self)
     }
   ),
 
@@ -204,10 +231,10 @@ LearnerSurvXgboostCox = R6Class("LearnerSurvXgboostCox",
         pv$watchlist = c(pv$watchlist, list(test = test_data))
       }
 
-      list(
+      structure(list(
         model = invoke(xgboost::xgb.train, data = data, .args = pv),
         train_data = data # for breslow
-      )
+      ), class = "xgboost_cox_model")
     },
 
     .predict = function(task) {
@@ -245,3 +272,38 @@ LearnerSurvXgboostCox = R6Class("LearnerSurvXgboostCox",
 )
 
 .extralrns_dict$add("surv.xgboost.cox", LearnerSurvXgboostCox)
+
+
+#' @export
+marshal_model.xgboost_cox_model = function(model, inplace = FALSE, ...) {
+  # xgb.DMatrix cannot be saved to a raw vector, but only to a file,
+  # so we save it to a temporary file and then read it back as a raw vector.
+  on.exit(unlink(tmp), add = TRUE)
+  tmp = tempfile(fileext = ".buffer")
+  xgboost::xgb.DMatrix.save(model$train_data, "xgb.data")
[…]
