+++ dev/pkgdown.yml
@@ -6,7 +6,7 @@ articles:
   extending: extending.html
   articles/learner_status: learner_status.html
   articles/list_learners: list_learners.html
-last_built: 2025-05-28T09:57Z
+last_built: 2025-05-28T10:25Z
 urls:
   reference: https://mlr3extralearners.mlr-org.com/reference
   article: https://mlr3extralearners.mlr-org.com/articles
