### 2025-06-01 675cb7c
adding evtree classif learner
