+++ NAMESPACE
@@ -13,6 +13,7 @@ export(LearnerClassifCatboost)
 export(LearnerClassifDecisionStump)
 export(LearnerClassifDecisionTable)
 export(LearnerClassifEarth)
+export(LearnerClassifEvtree)
 export(LearnerClassifFNN)
 export(LearnerClassifGAMBoost)
 export(LearnerClassifGBM)
@@ -67,6 +68,7 @@ export(LearnerRegrCubist)
 export(LearnerRegrDecisionStump)
 export(LearnerRegrDecisionTable)
 export(LearnerRegrEarth)
+export(LearnerRegrEvtree)
 export(LearnerRegrFNN)
 export(LearnerRegrGAMBoost)
 export(LearnerRegrGBM)
