+++ tests/testthat/test_paramtest_evtree_classif_evtree.R
@@ -44,4 +44,3 @@ test_that("classif.evtree control", {
   paramtest = run_paramtest(learner, fun, exclude, tag = "train")
   expect_paramtest(paramtest)
 })
-# prediction handled by partykit
