### 2025-06-01 5150380
adding ada classification learner
