+++ man/mlr_learners_classif.ada.Rd
@@ -1,18 +1,25 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/learner_abess_classif_abess.R
-\name{mlr_learners_classif.abess}
-\alias{mlr_learners_classif.abess}
-\alias{LearnerClassifAbess}
-\title{Classification Abess Learner}
+% Please edit documentation in R/learner_ada_classif_ada.R
+\name{mlr_learners_classif.ada}
+\alias{mlr_learners_classif.ada}
+\alias{LearnerClassifAdaBoosting}
+\title{Classification Boosting Learner}
 \description{
-Adaptive best-subset selection for classification.
-Calls \code{\link[abess:abess]{abess::abess()}} from \CRANpkg{abess}.
+Classification using ada.
+Calls \code{\link[ada:ada]{ada::ada()}} from \CRANpkg{ada}.
 }
+\section{Initial parameter values}{
+
+\itemize{
+\item \code{xval} is set to \code{0} by default for speed.
+}
+}
+
 \section{Dictionary}{
 
 This \link[mlr3:Learner]{Learner} can be instantiated via \link[mlr3:mlr_sugar]{lrn()}:
 
-\if{html}{\out{<div class="sourceCode">}}\preformatted{lrn("classif.abess")
+\if{html}{\out{<div class="sourceCode">}}\preformatted{lrn("classif.ada")
 }\if{html}{\out{</div>}}
 }
 
@@ -21,56 +28,40 @@ This \link[mlr3:Learner]{Learner} can be instantiated via \link[mlr3:mlr_sugar]{
 \itemize{
 \item Task type: \dQuote{classif}
 \item Predict Types: \dQuote{response}, \dQuote{prob}
-\item Feature Types: \dQuote{integer}, \dQuote{numeric}
-\item Required Packages: \CRANpkg{mlr3}, \CRANpkg{abess}
+\item Feature Types: \dQuote{logical}, \dQuote{integer}, \dQuote{numeric}, \dQuote{factor}, \dQuote{ordered}
+\item Required Packages: \CRANpkg{mlr3}, \CRANpkg{ada}, \CRANpkg{rpart}
 }
 }
 
 \section{Parameters}{
 \tabular{lllll}{
    Id \tab Type \tab Default \tab Levels \tab Range \cr
-   family \tab character \tab - \tab binomial, multinomial, ordinal \tab - \cr
-   tune.path \tab character \tab sequence \tab sequence, gsection \tab - \cr
-   tune.type \tab character \tab gic \tab gic, aic, bic, ebic, cv \tab - \cr
-   normalize \tab integer \tab NULL \tab  \tab \eqn{(-\infty, \infty)}{(-Inf, Inf)} \cr
-   support.size \tab untyped \tab NULL \tab  \tab - \cr
-   c.max \tab integer \tab 2 \tab  \tab \eqn{[1, \infty)}{[1, Inf)} \cr
-   gs.range \tab untyped \tab NULL \tab  \tab - \cr
-   lambda \tab numeric \tab 0 \tab  \tab \eqn{[0, \infty)}{[0, Inf)} \cr
-   always.include \tab untyped \tab NULL \tab  \tab - \cr
-   group.index \tab untyped \tab NULL \tab  \tab - \cr
-   init.active.set \tab untyped \tab NULL \tab  \tab - \cr
-   splicing.type \tab integer \tab 2 \tab  \tab \eqn{[1, 2]}{[1, 2]} \cr
-   max.splicing.iter \tab integer \tab 20 \tab  \tab \eqn{[1, \infty)}{[1, Inf)} \cr
[…]
