### 2025-06-01 cfebbb5
fix paramset and test
