### 2025-06-02 118d460
fix oob error and add entry to NEWS.md
