### 2025-06-02 d73423b
fix more learners
