+++ R/learner_randomForestSRC_regr_rfsrc.R
@@ -162,7 +162,7 @@ LearnerRegrRandomForestSRC = R6Class("LearnerRegrRandomForestSRC",
       pv = convert_ratio(pv, "sampsize", "sampsize.ratio", task$nrow)
       cores = pv$cores %??% 1L
       pv$cores = NULL
-      pv$case.wt = as.numeric(private$.get_weights(task)) # nolint
+      pv$case.wt = private$.get_weights(task) # nolint
 
       invoke(randomForestSRC::rfsrc,
         formula = task$formula(), data = task$data(),
