+++ R/learner_randomForestSRC_classif_rfsrc.R
@@ -173,7 +173,7 @@ LearnerClassifRandomForestSRC = R6Class("LearnerClassifRandomForestSRC",
       pv = convert_ratio(pv, "sampsize", "sampsize.ratio", task$nrow)
       cores = pv$cores %??% 1L
       pv$cores = NULL
-      pv$case.wt = as.numeric(private$.get_weights(task)) # nolint
+      pv$case.wt = private$.get_weights(task) # nolint
 
       invoke(randomForestSRC::rfsrc,
         formula = task$formula(), data = data.table::setDF(task$data()),
