+++ R/learner_randomForestSRC_classif_imbalanced_rfsrc.R
@@ -165,7 +165,7 @@ LearnerClassifImbalancedRandomForestSRC = R6Class("LearnerClassifImbalancedRando
       pv = convert_ratio(pv, "sampsize", "sampsize.ratio", task$nrow)
       cores = pv$cores %??% 1L
       pv$cores = NULL
-      pv$case.wt = as.numeric(private$.get_weights(task)) # nolint
+      pv$case.wt = private$.get_weights(task) # nolint
 
       invoke(randomForestSRC::imbalanced.rfsrc,
              formula = task$formula(), data = data.table::setDF(task$data()),
