### 2025-06-02 2ba0a63
add marginal.xvar parameter
