+++ R/learner_randomForestSRC_classif_rfsrc.R
@@ -115,7 +115,8 @@ LearnerClassifRandomForestSRC = R6Class("LearnerClassifRandomForestSRC",
         cores = p_int(default = 1L, lower = 1L, tags = c("train", "predict", "threads")),
         save.memory = p_lgl(default = FALSE, tags = "train"),
         perf.type = p_fct(levels = c("gmean", "misclass", "brier", "none"), tags = "train"), # nolint
-        case.depth = p_lgl(default = FALSE, tags = c("train", "predict"))
+        case.depth = p_lgl(default = FALSE, tags = c("train", "predict")),
+        marginal.xvar	= p_uty(default = NULL, tags = "predict")
       )
 
       super$initialize(
