### 2025-06-02 #8331 @editorialbot
[REVIEW]: mlr3extralearners: Expanding the mlr3 Ecosystem with Community-Driven Learner Integration
**Submitting author:** @sebffischer (<a href="http://orcid.org/0000-0002-9609-3197">Sebastian Fischer</a>)
**Repository:** https://github.com/mlr-org/mlr3extralearners
**Branch with paper.md** (empty if default branch): main
**Version:** 1.3.0
