+++ R/bibentries.R
@@ -720,5 +720,18 @@ bibentries = c( # nolint start
     volume = "116",
     pages = "1402--1412",
     doi = "10.1080/01621459.2020.1725521"
+  ),
+  howard_2020 = bibentry("article",
+    title     = "Fastai: A Layered API for Deep Learning",
+    author    = "Howard, Jeremy and Gugger, Sylvain",
+    year      = "2020",
+    journal   = "Information",
+    volume    = "11",
+    number    = "2",
+    pages     = "108",
+    doi       = "10.3390/info11020108",
+    issn      = "2078-2489",
+    publisher = "MDPI AG",
+    month     = "feb"
   )
 ) # nolint end
