### 2025-06-03 c7b91a8
feat: add learner weights (#428)
- Add the new version of learner weights to all learners
- Removes weights from learners that don't support it

The changes are based on mlr-org/mlr3learners#340 and
mlr-org/mlr3learners#338 in `mlr3learners`.

---------

Co-authored-by: Sebastian Fischer <sebf.fischer@gmail.com>
Co-authored-by: Marc Becker <33069354+be-marc@users.noreply.github.com>
Co-authored-by: Copilot <175728472+Copilot@users.noreply.github.com>
