+++ .github/workflows/devcmdcheck.yml
@@ -58,7 +58,8 @@ jobs:
           pak::pkg_install('reticulate')
           reticulate::install_miniconda(force = TRUE)
           reticulate::use_condaenv('r-reticulate')
-          install.packages('keras')
+          pak::pkg_install(c('eagerai/fastai', 'keras'))
+          fastai::install_fastai(gpu = FALSE, cuda_version = '11.6', overwrite = FALSE)
           reticulate::py_install("pytorch-cpu")
           reticulate::py_install("tensorflow-cpu", pip = TRUE)
           survivalmodels::install_pycox(pip = TRUE, install_torch = FALSE)
