+++ tests/testthat/test_paramtest_fastai_regr_fastai.R
@@ -1,26 +1,0 @@
-test_that("regr.fastai train", {
-  learner = lrn("regr.fastai")
-  fun = fastai::fastai
-  exclude = c(
-    "object", # handled internally
-    "data" # handled internally
-  )
-
-  # note that you can also pass a list of functions in case $.train calls more than one
-  # function, e.g. for control arguments
-  paramtest = run_paramtest(learner, fun, exclude, tag = "train")
-  expect_paramtest(paramtest)
-})
-
-test_that("regr.fastai predict", {
-  learner = lrn("regr.fastai")
-  fun = fastai::predict # nolint
-  exclude = c(
-    "object", # handled internally
-    "data", # handled internally
-    "newdata" # handled internally
-  )
-
-  paramtest = run_paramtest(learner, fun, exclude, tag = "predict")
-  expect_paramtest(paramtest)
-})
