### 2025-06-03 80f4c3d
fix: weights property
