+++ R/learner_aorsf_surv_aorsf.R
@@ -90,7 +90,7 @@ LearnerSurvAorsf = R6Class("LearnerSurvAorsf",
         feature_types = c("integer", "numeric", "factor", "ordered"),
         predict_types = c("crank", "distr", "response"),
         param_set = ps,
-        properties = c("oob_error", "importance", "missings"),
+        properties = c("oob_error", "importance", "missings", "weights"),
         man = "mlr3extralearners::mlr_learners_surv.aorsf",
         label = "Oblique Random Forest"
       )
