+++ tests/testthat/test_paramtest_adabag_classif_adabag.R
@@ -1,0 +1,26 @@
+test_that("classif.adabag train", {
+  learner = lrn("classif.adabag")
+  fun = adabag::adabag
+  exclude = c(
+    "object", # handled internally
+    "data" # handled internally
+  )
+
+  # note that you can also pass a list of functions in case $.train calls more than one
+  # function, e.g. for control arguments
+  paramtest = run_paramtest(learner, fun, exclude, tag = "train")
+  expect_paramtest(paramtest)
+})
+
+test_that("classif.adabag predict", {
+  learner = lrn("classif.adabag")
+  fun = adabag::predict # nolint
+  exclude = c(
+    "object", # handled internally
+    "data", # handled internally
+    "newdata" # handled internally
+  )
+
+  paramtest = run_paramtest(learner, fun, exclude, tag = "predict")
+  expect_paramtest(paramtest)
+})
