+++ man/mlr_learners_classif.adabag.Rd
@@ -1,18 +1,30 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/learner_abess_classif_abess.R
-\name{mlr_learners_classif.abess}
-\alias{mlr_learners_classif.abess}
-\alias{LearnerClassifAbess}
-\title{Classification Abess Learner}
+% Please edit documentation in R/learner_adabag_classif_adabag.R
+\name{mlr_learners_classif.adabag}
+\alias{mlr_learners_classif.adabag}
+\alias{LearnerClassifAdabag}
+\title{Classification Boosting Learner}
 \description{
-Adaptive best-subset selection for classification.
-Calls \code{\link[abess:abess]{abess::abess()}} from \CRANpkg{abess}.
+Classification boosting algorithm.
+Calls \code{\link[adabag:boosting]{adabag::boosting()}} from \CRANpkg{adabag}.
 }
+\section{Initial parameter values}{
+
+\itemize{
+\item \code{xval}:
+\itemize{
+\item Actual default: 10L
+\item Initial value: 0L
+\item Reason for change: Set to 0 for speed.
+}
+}
+}
+
 \section{Dictionary}{
 
 This \link[mlr3:Learner]{Learner} can be instantiated via \link[mlr3:mlr_sugar]{lrn()}:
 
-\if{html}{\out{<div class="sourceCode">}}\preformatted{lrn("classif.abess")
+\if{html}{\out{<div class="sourceCode">}}\preformatted{lrn("classif.adabag")
 }\if{html}{\out{</div>}}
 }
 
@@ -21,56 +33,34 @@ This \link[mlr3:Learner]{Learner} can be instantiated via \link[mlr3:mlr_sugar]{
 \itemize{
 \item Task type: \dQuote{classif}
 \item Predict Types: \dQuote{response}, \dQuote{prob}
-\item Feature Types: \dQuote{integer}, \dQuote{numeric}
-\item Required Packages: \CRANpkg{mlr3}, \CRANpkg{abess}
+\item Feature Types: \dQuote{logical}, \dQuote{integer}, \dQuote{numeric}, \dQuote{factor}, \dQuote{ordered}
+\item Required Packages: \CRANpkg{mlr3}, \CRANpkg{adabag}, \CRANpkg{rpart}
 }
 }
 
 \section{Parameters}{
 \tabular{lllll}{
    Id \tab Type \tab Default \tab Levels \tab Range \cr
-   family \tab character \tab - \tab binomial, multinomial, ordinal \tab - \cr
-   tune.path \tab character \tab sequence \tab sequence, gsection \tab - \cr
-   tune.type \tab character \tab gic \tab gic, aic, bic, ebic, cv \tab - \cr
-   normalize \tab integer \tab NULL \tab  \tab \eqn{(-\infty, \infty)}{(-Inf, Inf)} \cr
-   support.size \tab untyped \tab NULL \tab  \tab - \cr
-   c.max \tab integer \tab 2 \tab  \tab \eqn{[1, \infty)}{[1, Inf)} \cr
-   gs.range \tab untyped \tab NULL \tab  \tab - \cr
-   lambda \tab numeric \tab 0 \tab  \tab \eqn{[0, \infty)}{[0, Inf)} \cr
-   always.include \tab untyped \tab NULL \tab  \tab - \cr
-   group.index \tab untyped \tab NULL \tab  \tab - \cr
-   init.active.set \tab untyped \tab NULL \tab  \tab - \cr
-   splicing.type \tab integer \tab 2 \tab  \tab \eqn{[1, 2]}{[1, 2]} \cr
[…]
