### 2025-06-03 b0a6a44
update NEWS.md
