### 2025-06-10 e2202b0
fix weight parameter
