### 2025-06-11 0843c72
fix weight parameter
