### 2025-06-11 7ed2315
fix: tabpfn import
