+++ R/learner_tabpfn_regr_tabpfn.R
@@ -184,26 +184,3 @@ LearnerRegrTabPFN = R6Class("LearnerRegrTabPFN",
 
 .extralrns_dict$add("regr.tabpfn", LearnerRegrTabPFN)
 
-
-#' @export
-marshal_model.tabpfn_model = function(model, inplace = FALSE, ...) {
-  # pickle should be available in any python environment
-  pickle = reticulate::import("pickle")
-  # save model as bytes
-  pickled = pickle$dumps(model$fitted)
-  # ensure object is converted to R
-  pickled = as.raw(pickled)
-
-  structure(list(
-    marshaled = pickled,
-    packages = "mlr3extralearners"
-  ), class = c("tabpfn_model_marshaled", "marshaled"))
-}
-
-#' @export
-unmarshal_model.tabpfn_model_marshaled = function(model, inplace = FALSE, ...) {
-  pickle = reticulate::import("pickle")
-  # unpickle
-  fitted = pickle$loads(reticulate::r_to_py(model$marshaled))
-  structure(list(fitted = fitted), class = "tabpfn_model")
-}
