+++ R/learner_tabpfn_classif_tabpfn.R
@@ -144,8 +144,6 @@ LearnerClassifTabPFN = R6Class("LearnerClassifTabPFN",
     },
 
     .predict = function(task) {
-      message("Predict2")
-      message("Predict")
       reticulate::py_require("tabpfn")
       reticulate::import("tabpfn")
       model = self$model$fitted
@@ -176,7 +174,6 @@ LearnerClassifTabPFN = R6Class("LearnerClassifTabPFN",
 marshal_model.tabpfn_model = function(model, inplace = FALSE, ...) {
   reticulate::py_require(c("pickle", "tabpfn"))
   reticulate::import("tabpfn")
-  message("Marshal")
   # pickle should be available in any python environment
   pickle = reticulate::import("pickle")
   # save model as bytes
@@ -194,7 +191,6 @@ marshal_model.tabpfn_model = function(model, inplace = FALSE, ...) {
 unmarshal_model.tabpfn_model_marshaled = function(model, inplace = FALSE, ...) {
   reticulate::py_require(c("pickle", "tabpfn"))
   reticulate::import("tabpfn")
-  message("Unmarshal tabpfn")
   pickle = reticulate::import("pickle")
   # unpickle
   fitted = pickle$loads(reticulate::r_to_py(model$marshaled))
