+++ R/learner_tabpfn_classif_tabpfn.R
@@ -172,7 +172,7 @@ LearnerClassifTabPFN = R6Class("LearnerClassifTabPFN",
 
 #' @export
 marshal_model.tabpfn_model = function(model, inplace = FALSE, ...) {
-  reticulate::py_require(c("pickle", "tabpfn"))
+  reticulate::py_require("tabpfn")
   reticulate::import("tabpfn")
   # pickle should be available in any python environment
   pickle = reticulate::import("pickle")
@@ -189,10 +189,11 @@ marshal_model.tabpfn_model = function(model, inplace = FALSE, ...) {
 
 #' @export
 unmarshal_model.tabpfn_model_marshaled = function(model, inplace = FALSE, ...) {
-  reticulate::py_require(c("pickle", "tabpfn"))
+  reticulate::py_require("tabpfn")
   reticulate::import("tabpfn")
   pickle = reticulate::import("pickle")
   # unpickle
   fitted = pickle$loads(reticulate::r_to_py(model$marshaled))
   structure(list(fitted = fitted), class = "tabpfn_model")
 }
+
