### 2025-06-11 f2077e6
fix measure tests
