### 2025-06-12 f3c5614
fix: also force import in regr predict
