+++ tests/testthat/test_adabag_classif_adabag.R
@@ -2,6 +2,6 @@ test_that("autotest", {
   learner = lrn("classif.adabag")
   expect_learner(learner)
   # note that you can skip tests using the exclude argument
-  result = run_autotest(learner, exclude = "feat_all_binary || utf8_feature_names_binary")
+  result = run_autotest(learner, exclude = "utf8_feature_names")
   expect_true(result, info = result$error)
 })
