+++ R/learner_adabag_classif_adabag.R
@@ -109,8 +109,7 @@ LearnerClassifAdabag = R6Class("LearnerClassifAdabag",
       type = self$predict_type
 
       # adaboost requires target column
-      newdata[, task$target_names] = factor(rep(1, nrow(newdata)),
-        levels = task$class_names)
+      newdata[, task$target_names] = factor(rep(1, nrow(newdata)), levels = task$class_names)
 
       pred = invoke(predict, self$model, newdata = newdata, .args = pars)
 
