+++ man/mlr_learners_classif.bst.Rd
@@ -1,54 +1,88 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/learner_Cubist_regr_cubist.R
-\name{mlr_learners_regr.cubist}
-\alias{mlr_learners_regr.cubist}
-\alias{LearnerRegrCubist}
-\title{Regression Cubist Learner}
+% Please edit documentation in R/learner_bst_classif_bst.R
+\name{mlr_learners_classif.bst}
+\alias{mlr_learners_classif.bst}
+\alias{LearnerClassifBst}
+\title{Classification Gradient Boosting Learner}
 \description{
-Rule-based model that is an extension of Quinlan's M5 model tree. Each tree contains
-linear regression models at the terminal leaves.
-Calls \code{\link[Cubist:cubist]{Cubist::cubist()}} from \CRANpkg{Cubist}.
+Classification gradient boosting learner.
+Calls \code{\link[bst:bst]{bst::bst()}} from \CRANpkg{bst}.
 }
+\note{
+Only classification-appropriate loss functions are available for the \code{family} parameter.
+}
+\section{Initial parameter values}{
+
+\itemize{
+\item \code{Learner = "ls"}: Default base learner type
+\item \code{xval = 0}: No cross-validation
+\item \code{maxdepth = 1}: Maximum tree depth
+}
+}
+
 \section{Dictionary}{
 
 This \link[mlr3:Learner]{Learner} can be instantiated via \link[mlr3:mlr_sugar]{lrn()}:
 
-\if{html}{\out{<div class="sourceCode">}}\preformatted{lrn("regr.cubist")
+\if{html}{\out{<div class="sourceCode">}}\preformatted{lrn("classif.bst")
 }\if{html}{\out{</div>}}
 }
 
 \section{Meta Information}{
 
 \itemize{
-\item Task type: \dQuote{regr}
-\item Predict Types: \dQuote{response}
-\item Feature Types: \dQuote{integer}, \dQuote{numeric}, \dQuote{character}, \dQuote{factor}, \dQuote{ordered}
-\item Required Packages: \CRANpkg{mlr3}, \CRANpkg{mlr3extralearners}, \CRANpkg{Cubist}
+\item Task type: \dQuote{classif}
+\item Predict Types: \dQuote{response}, \dQuote{prob}
+\item Feature Types: \dQuote{numeric}
+\item Required Packages: \CRANpkg{mlr3}, \CRANpkg{bst}, \CRANpkg{rpart}
 }
 }
 
 \section{Parameters}{
 \tabular{lllll}{
    Id \tab Type \tab Default \tab Levels \tab Range \cr
-   committees \tab integer \tab - \tab  \tab \eqn{[1, 100]}{[1, 100]} \cr
-   unbiased \tab logical \tab FALSE \tab TRUE, FALSE \tab - \cr
-   rules \tab integer \tab 100 \tab  \tab \eqn{[1, \infty)}{[1, Inf)} \cr
-   extrapolation \tab numeric \tab 100 \tab  \tab \eqn{[0, 100]}{[0, 100]} \cr
-   sample \tab integer \tab 0 \tab  \tab \eqn{[0, \infty)}{[0, Inf)} \cr
-   seed \tab integer \tab - \tab  \tab \eqn{(-\infty, \infty)}{(-Inf, Inf)} \cr
-   label \tab untyped \tab "outcome" \tab  \tab - \cr
[…]
