+++ NAMESPACE
@@ -8,6 +8,7 @@ export(LearnerClassifAbess)
 export(LearnerClassifAdaBoostM1)
 export(LearnerClassifBart)
 export(LearnerClassifBayesNet)
+export(LearnerClassifBst)
 export(LearnerClassifC50)
 export(LearnerClassifCForest)
 export(LearnerClassifCTree)
@@ -62,6 +63,7 @@ export(LearnerDensPlugin)
 export(LearnerDensSpline)
 export(LearnerRegrAbess)
 export(LearnerRegrBart)
+export(LearnerRegrBst)
 export(LearnerRegrCForest)
 export(LearnerRegrCTree)
 export(LearnerRegrCatboost)
