### 2025-06-12 89adbb2
fix bibentries
