### 2025-06-13 fb8f203
try fixing bibentries again
