### 2025-06-14 6d428be
add marshaling to classif learner
