+++ man/mlr_learners_classif.fastai.Rd
@@ -6,7 +6,7 @@
 \title{Classification Neural Network Learner}
 \description{
 Simple and fast neural nets for tabular data classification.
-Calls \code{\link[fastai:fastai]{fastai::fastai()}} from package \CRANpkg{fastai}.
+Calls \code{\link[fastai:tabular_learner]{fastai::tabular_learner()}} from package \CRANpkg{fastai}.
 }
 \section{Initial parameter values}{
 
@@ -16,7 +16,7 @@ Needs to be set for \code{\link[fastai:reexports]{fastai::fit()}} to work.
 If no value is given, it is set to 5.
 \item \code{eval_metric}:
 Needs to be set for \code{\link[fastai:reexports]{fastai::predict()}} to work.
-If no value is given, it is set to \link[mlr3measures:classif.logloss]{mlr3measures::classif.logloss}.
+If no value is given, it is set to \code{fastai::accuracy()}.
 }
 }
 
@@ -30,14 +30,71 @@ This \link[mlr3:Learner]{Learner} can be instantiated via \link[mlr3:mlr_sugar]{
 
 \section{Meta Information}{
 
-\verb{r paste(mlr3misc::rd_info(mlr3::lrn("classif.fastai")), collapse = "\\n")}
+\itemize{
+\item Task type: \dQuote{classif}
+\item Predict Types: \dQuote{response}, \dQuote{prob}
+\item Feature Types: \dQuote{logical}, \dQuote{integer}, \dQuote{numeric}, \dQuote{factor}, \dQuote{ordered}
+\item Required Packages: \CRANpkg{mlr3}, \CRANpkg{fastai}, \CRANpkg{reticulate}
+}
 }
 
 \section{Parameters}{
-
-\verb{r paste(mlr3misc::rd_info(mlr3::lrn("classif.fastai")$param_set), collapse = "\\n")}
+\tabular{lllll}{
+   Id \tab Type \tab Default \tab Levels \tab Range \cr
+   act_cls \tab untyped \tab - \tab  \tab - \cr
+   bn_cont \tab logical \tab TRUE \tab TRUE, FALSE \tab - \cr
+   bn_final \tab logical \tab FALSE \tab TRUE, FALSE \tab - \cr
+   drop_last \tab logical \tab FALSE \tab TRUE, FALSE \tab - \cr
+   embed_p \tab numeric \tab 0 \tab  \tab \eqn{[0, 1]}{[0, 1]} \cr
+   emb_szs \tab untyped \tab NULL \tab  \tab - \cr
+   n_epoch \tab integer \tab - \tab  \tab \eqn{[1, \infty)}{[1, Inf)} \cr
+   eval_metric \tab untyped \tab - \tab  \tab - \cr
+   layers \tab untyped \tab - \tab  \tab - \cr
+   loss_func \tab untyped \tab - \tab  \tab - \cr
+   lr \tab numeric \tab 0.001 \tab  \tab \eqn{[0, \infty)}{[0, Inf)} \cr
+   metrics \tab untyped \tab - \tab  \tab - \cr
+   n_out \tab integer \tab - \tab  \tab \eqn{(-\infty, \infty)}{(-Inf, Inf)} \cr
+   num_workers \tab integer \tab 0 \tab  \tab \eqn{(-\infty, \infty)}{(-Inf, Inf)} \cr
+   opt_func \tab untyped \tab - \tab  \tab - \cr
+   patience \tab integer \tab 1 \tab  \tab \eqn{[1, \infty)}{[1, Inf)} \cr
+   pin_memory \tab logical \tab TRUE \tab TRUE, FALSE \tab - \cr
+   procs \tab untyped \tab NULL \tab  \tab - \cr
+   ps \tab untyped \tab NULL \tab  \tab - \cr
[…]
