+++ man/mlr_learners_surv.ctree.Rd
@@ -14,7 +14,7 @@ This learner returns two prediction types:
 \enumerate{
 \item \code{distr}: a survival matrix in two dimensions, where observations are
 represented in rows and time points in columns.
-Calculated using the internal \code{\link[partykit:predict.party]{partykit::predict.party()}} function.
+Calculated using the internal \code{\link[partykit:party-predict]{partykit::predict.party()}} function.
 \item \code{crank}: the expected mortality using \code{\link[mlr3proba:dot-surv_return]{mlr3proba::.surv_return()}}.
 }
 }
