### 2025-06-16 d4d4391
fix bibentries
