### 2025-06-16 3894747
fix notes
