+++ NAMESPACE
@@ -1,7 +1,9 @@
 # Generated by roxygen2: do not edit by hand
 
+S3method(marshal_model,fastai_model)
 S3method(marshal_model,tabpfn_model)
 S3method(marshal_model,xgboost_cox_model)
+S3method(unmarshal_model,fastai_model_marshaled)
 S3method(unmarshal_model,tabpfn_model_marshaled)
 S3method(unmarshal_model,xgboost_cox_model_marshaled)
 export(LearnerClassifAbess)
