### 2025-06-16 a9b8ca0
knn classif learner from old mlr2
