+++ NAMESPACE
@@ -28,6 +28,7 @@ export(LearnerClassifJ48)
 export(LearnerClassifJRip)
 export(LearnerClassifKSVM)
 export(LearnerClassifKStar)
+export(LearnerClassifKnn)
 export(LearnerClassifLMT)
 export(LearnerClassifLSSVM)
 export(LearnerClassifLiblineaR)
