+++ tests/testthat/test_mda_classif_mda.R
@@ -1,0 +1,7 @@
+test_that("autotest", {
+  learner = lrn("classif.mda", subclasses = 1)
+  expect_learner(learner)
+  # note that you can skip tests using the exclude argument
+  result = run_autotest(learner)
+  expect_true(result, info = result$error)
+})
