+++ NAMESPACE
@@ -28,11 +28,13 @@ export(LearnerClassifJ48)
 export(LearnerClassifJRip)
 export(LearnerClassifKSVM)
 export(LearnerClassifKStar)
+export(LearnerClassifKnn)
 export(LearnerClassifLMT)
 export(LearnerClassifLSSVM)
 export(LearnerClassifLiblineaR)
 export(LearnerClassifLightGBM)
 export(LearnerClassifLogistic)
+export(LearnerClassifMda)
 export(LearnerClassifMob)
 export(LearnerClassifMultilayerPerceptron)
 export(LearnerClassifNaiveBayesMultinomial)
