### 2025-06-17 aa30847
added Exhaustive Search (Classif) https://github.com/mlr-org/mlr3extralearners/issues/365#
