### 2025-06-23 7a6ce11
chore: require mlr3 1.0.0
