+++ NAMESPACE
@@ -43,6 +43,7 @@ export(LearnerClassifNaiveBayesWeka)
 export(LearnerClassifObliqueRandomForest)
 export(LearnerClassifOneR)
 export(LearnerClassifPART)
+export(LearnerClassifPamr)
 export(LearnerClassifPriorityLasso)
 export(LearnerClassifREPTree)
 export(LearnerClassifRandomForest)
@@ -152,6 +153,7 @@ importFrom(methods,formalArgs)
 importFrom(methods,getMethod)
 importFrom(mlr3,LearnerClassif)
 importFrom(mlr3,LearnerRegr)
+importFrom(mlr3,assert_validate)
 importFrom(mlr3,lrn)
 importFrom(mlr3,lrns)
 importFrom(mlr3,marshal_model)
