+++ R/learner_ExhaustiveSearch_classif_glm.R
@@ -1,32 +1,78 @@
+#' @title Logistic Regression with Exhaustive Search
+#' @author ngerman
+#' @name mlr_learners_classif.exhaustive_search
+#'
+#' @description
+#' Logistic Regression, where the features used in the model are selected by running Exhaustive Search.
+#'
+#' Calls [ExhaustiveSearch::ExhaustiveSearch()] from \CRANpkg{ExhaustiveSearch}.
+#'
+#' @section Initial parameter values:
+#' - `family`:
+#'   - Actual default: NULL
+#'   - Adjusted default: "binomial"
+#'   - Reason for change: To comply with mlr3 architecture, we differentiate between classification and regression learners.
+#' - `nThreads`:
+#'   - Actual default: NULL
+#'   - Adjusted default: 1
+#'   - Reason for change: Suppressing the automatic internal parallelization if
+#'     `cv.folds` > 0.
+#'
+#'
+#' @template learner
+#' @templateVar id classif.exhaustive_search
+#'
+#' @details
+#' Internal validation not yet implemented.
+#'
+#' @export
+#' @template seealso_learner
+#' @examples
+#'
+#' # define the learner
+#' learner = LearnerClassifExhaustiveSearch$new()
+#' learner$predict_type = "prob"
+#' learner$param_set$set_values(combsUpTo = 3)
+#'
+#' # define the task
+#' tsk_gc = tsk("german_credit")
+#'
+#' # train the learner
+#' learner$train(tsk_gc)
+#'
+#' # extract selected features
+#' learner$selected_features()
+#'
+#' # predict on training task
+#' learner$predict(tsk_gc)
+
 library(mlr3)
 library(paradox)
 library(R6)
-library(mlr3misc)
 
-LearnerClassifExhaustiveSearch = R6Class("LearnerRegrExhaustiveSearch",
+LearnerClassifExhaustiveSearch = R6Class("LearnerClassifExhaustiveSearch",
                                       inherit = LearnerClassif,
                                       public = list(
                                         initialize = function() {
                                           param_set = ps(
-                                            family = p_fct(c("gaussian", "binomial"), tags = "train"), # maybe delete option gaussian in classif case?
+                                            family = p_fct(c("gaussian", "binomial"), init = "binomial", tags = "train"),
                                             performanceMeasure = p_fct(c("MSE", "AIC"), tags = "train"), # interdepency to testSetIDs
                                             combsUpTo = p_int(1L, tags = "train"),
                                             nResults = p_int(1L, default = 5000L, tags = "train"),
-                                            nThreads = p_int(1L, tags = "train"), # side effects to parallel?
+                                            nThreads = p_int(1L, init = 1L, tags = "train"),
                                             testSetIDs = p_int(1L, tags = "train"), # use as internal validation?
-                                            errorVal = p_uty(default = -1, tags = "train"),
+                                            errorVal = p_uty(default = -1L, tags = "train"),
                                             quietly = p_lgl(default = FALSE, tags = "train"),
                                             checkLarge = p_lgl(default = TRUE, tags = "train")
                                           )
-                                          param_set$set_values(family = "binomial")
 
                                           super$initialize(
                                             id = "classif.exhaustive_search",
-                                            feature_types = c("logical", "integer", "numeric", "factor", "ordered"),
+                                            feature_types = c("logical", "integer", "numeric", "character", "factor", "ordered"),
                                             predict_types = c("response", "prob"),
                                             packages = "ExhaustiveSearch",
                                             param_set = param_set,
-                                            properties = c("twoclass", "missings", "selected_features"),
+                                            properties = c("twoclass", "selected_features"), # to add: "validation"
                                             label = "Exhaustive Search",
[…]
