### 2025-06-24 1dae555
added examples for regr.ES and classif.ES  https://github.com/mlr-org/mlr3extralearners/issues/365#
