+++ R/learner_ExhaustiveSearch_regr_lm.R
@@ -1,32 +1,71 @@
-library(mlr3)
-library(paradox)
-library(R6)
-library(mlr3misc)
-
+#' @title Linear Regression with Exhaustive Search
+#' @author ngerman
+#' @name mlr_learners_regr.exhaustive_search
+#'
+#' @description
+#' Linear Regression, where the features used in the model are selected by running Exhaustive Search.
+#'
+#' Calls [ExhaustiveSearch::ExhaustiveSearch()] from \CRANpkg{ExhaustiveSearch}.
+#'
+#' @section Initial parameter values:
+#' - `family`:
+#'   - Actual default: NULL
+#'   - Adjusted default: "gaussian"
+#'   - Reason for change: To comply with mlr3 architecture, we differentiate between classification and regression learners.
+#' - `nThreads`:
+#'   - Actual default: NULL
+#'   - Adjusted default: 1
+#'   - Reason for change: Suppressing the automatic internal parallelization if
+#'     `cv.folds` > 0.
+#'
+#'
+#' @template learner
+#' @templateVar id regr.exhaustive_search
+#'
+#' @details
+#' Internal validation not yet implemented.
+#'
+#' @export
+#' @template seealso_learner
+#' @examples
+#' # define learner
+#' learner = LearnerRegrExhaustiveSearch$new()
+#' learner$predict_type = "se"
+#'
+#' # define task
+#' tsk_cars = tsk("mtcars")
+#'
+#' # train learner
+#' learner$train(tsk_cars)
+#'
+#' # extract selected features
+#' learner$selected_features()
+#'
+#' # predict on training task
+#' learner$predict(tsk_cars)
 LearnerRegrExhaustiveSearch = R6Class("LearnerRegrExhaustiveSearch",
                                       inherit = LearnerRegr,
                                       public = list(
                                         initialize = function() {
                                           param_set = ps(
-                                            family = p_fct(c("gaussian", "binomial"), tags = "train"), # maybe delete option binomial in regr case?
+                                            family = p_fct(c("gaussian", "binomial"), init = "gaussian", tags = "train"),
                                             performanceMeasure = p_fct(c("MSE", "AIC"), tags = "train"), # interdepency to testSetIDs
                                             combsUpTo = p_int(1L, tags = "train"),
                                             nResults = p_int(1L, default = 5000L, tags = "train"),
-                                            nThreads = p_int(1L, tags = "train"), # side effects to parallel?
+                                            nThreads = p_int(1L, init = 1L, tags = "train"),
                                             testSetIDs = p_int(1L, tags = "train"), # use as internal validation?
                                             errorVal = p_uty(default = -1, tags = "train"),
                                             quietly = p_lgl(default = FALSE, tags = "train"),
                                             checkLarge = p_lgl(default = TRUE, tags = "train")
                                           )
-                                          param_set$set_values(family = "gaussian")
 
                                           super$initialize(
                                             id = "regr.exhaustive_search",
-                                            feature_types = c("logical", "integer", "numeric", "factor", "ordered"),
-                                            predict_types = "response",
+                                            feature_types = c("logical", "integer", "numeric", "factor", "character"),
+                                            predict_types = c("response", "se"),
                                             packages = "ExhaustiveSearch",
                                             param_set = param_set,
-                                            properties = c("missings", "selected_features"),
+                                            properties = c("selected_features"),
                                             label = "Exhaustive Search",
                                             man = "mlr3extralearners::mlr_learners_regr.exhaustive_search"
                                           )
@@ -50,34 +89,33 @@ LearnerRegrExhaustiveSearch = R6Class("LearnerRegrExhaustiveSearch",
                                             .args = pv
                                           )
                                           # extract selected features of best performing model
-                                          # make new task
[…]
