### 2025-06-30 510d0a9
build man/, added tests, changed default value of 'quietly' to TRUE
