### 2025-07-01 9e6edaa
fixed identation, NAMESPACE, man/ are now beeing generated without error
