+++ man/mlr_learners_classif.exhaustive_search.Rd
@@ -2,6 +2,7 @@
 % Please edit documentation in R/learner_ExhaustiveSearch_classif_glm.R
 \name{mlr_learners_classif.exhaustive_search}
 \alias{mlr_learners_classif.exhaustive_search}
+\alias{LearnerClassifExhaustiveSearch}
 \title{Logistic Regression with Exhaustive Search}
 \description{
 Logistic Regression, where the features used in the model are selected by running Exhaustive Search.
@@ -50,7 +51,7 @@ This \link[mlr3:Learner]{Learner} can be instantiated via \link[mlr3:mlr_sugar]{
 \item Task type: \dQuote{classif}
 \item Predict Types: \dQuote{response}, \dQuote{prob}
 \item Feature Types: \dQuote{logical}, \dQuote{integer}, \dQuote{numeric}, \dQuote{character}, \dQuote{factor}, \dQuote{ordered}
-\item Required Packages: \CRANpkg{mlr3}, \CRANpkg{ExhaustiveSearch}
+\item Required Packages: \CRANpkg{mlr3}, \CRANpkg{mlr3extralearners}, \CRANpkg{ExhaustiveSearch}
 }
 }
 
@@ -101,4 +102,69 @@ learner$predict(tsk_gc)
 }
 \author{
 ngerman
+}
+\section{Super classes}{
+\code{\link[mlr3:Learner]{mlr3::Learner}} -> \code{\link[mlr3:LearnerClassif]{mlr3::LearnerClassif}} -> \code{LearnerClassifExhaustiveSearch}
+}
+\section{Methods}{
+\subsection{Public methods}{
+\itemize{
+\item \href{#method-LearnerClassifExhaustiveSearch-new}{\code{LearnerClassifExhaustiveSearch$new()}}
+\item \href{#method-LearnerClassifExhaustiveSearch-selected_features}{\code{LearnerClassifExhaustiveSearch$selected_features()}}
+\item \href{#method-LearnerClassifExhaustiveSearch-clone}{\code{LearnerClassifExhaustiveSearch$clone()}}
+}
+}
+\if{html}{\out{
+<details><summary>Inherited methods</summary>
+<ul>
+<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="base_learner"><a href='../../mlr3/html/Learner.html#method-Learner-base_learner'><code>mlr3::Learner$base_learner()</code></a></span></li>
+<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="configure"><a href='../../mlr3/html/Learner.html#method-Learner-configure'><code>mlr3::Learner$configure()</code></a></span></li>
+<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="encapsulate"><a href='../../mlr3/html/Learner.html#method-Learner-encapsulate'><code>mlr3::Learner$encapsulate()</code></a></span></li>
+<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="format"><a href='../../mlr3/html/Learner.html#method-Learner-format'><code>mlr3::Learner$format()</code></a></span></li>
+<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="help"><a href='../../mlr3/html/Learner.html#method-Learner-help'><code>mlr3::Learner$help()</code></a></span></li>
+<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="predict"><a href='../../mlr3/html/Learner.html#method-Learner-predict'><code>mlr3::Learner$predict()</code></a></span></li>
[…]

[…]
