### 2025-07-01 b9fff44
deleted call to ExhaustiveSearch in test
