### 2025-07-02 42b9b94
add test, news, document
