+++ NEWS.md
@@ -1,5 +1,6 @@
 # dev
 
+* fix: lightgbm classifier now works with encapsulation (#437)
 * feat: Support offset in learners `regr|classif.mgcv`, `regr.glm` and `regr.lmer`.
 * feat: Added learners `LearnerRegrQGam` and `LearnerRegrMQGam`.
 * feat: Added learners `LearnerClassifTabPFN` and `LearnerRegrTabPFN`.
