### 2025-07-02 11da65a
fix: lightgbm prob matrix names (#438)
Co-authored-by: Sebastian Fischer <sebf.fischer@gmail.com>
