+++ man/mlr_learners_classif.pamr.Rd
@@ -1,0 +1,144 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/learner_pamr_classif_pamr.R
+\name{mlr_learners_classif.pamr}
+\alias{mlr_learners_classif.pamr}
+\alias{LearnerClassifPamr}
+\title{Classification Nearest Shrunken Centroid Learner}
+\description{
+Classification using nearest shrunken centroids.
+Calls \code{\link[pamr:pamr-package]{pamr::pamr()}} from \CRANpkg{pamr}.
+}
+\section{Dictionary}{
+
+This \link[mlr3:Learner]{Learner} can be instantiated via \link[mlr3:mlr_sugar]{lrn()}:
+
+\if{html}{\out{<div class="sourceCode">}}\preformatted{lrn("classif.pamr")
+}\if{html}{\out{</div>}}
+}
+
+\section{Meta Information}{
+
+\itemize{
+\item Task type: \dQuote{classif}
+\item Predict Types: \dQuote{response}, \dQuote{prob}
+\item Feature Types: \dQuote{numeric}
+\item Required Packages: \CRANpkg{mlr3}, \CRANpkg{pamr}
+}
+}
+
+\section{Parameters}{
+\tabular{lllll}{
+   Id \tab Type \tab Default \tab Levels \tab Range \cr
+   gene.subset \tab untyped \tab - \tab  \tab - \cr
+   hetero \tab untyped \tab - \tab  \tab - \cr
+   n.threshold \tab integer \tab 30 \tab  \tab \eqn{(-\infty, \infty)}{(-Inf, Inf)} \cr
+   ngroup.survival \tab integer \tab 2 \tab  \tab \eqn{(-\infty, \infty)}{(-Inf, Inf)} \cr
+   offset.percent \tab numeric \tab 50 \tab  \tab \eqn{[0, 100]}{[0, 100]} \cr
+   prior \tab numeric \tab - \tab  \tab \eqn{[0, 1]}{[0, 1]} \cr
+   remove.zeros \tab logical \tab TRUE \tab TRUE, FALSE \tab - \cr
+   sample.subset \tab untyped \tab - \tab  \tab - \cr
+   scale.sd \tab logical \tab TRUE \tab TRUE, FALSE \tab - \cr
+   se.scale \tab untyped \tab - \tab  \tab - \cr
+   sign.contrast \tab character \tab both \tab both, negative, positive \tab - \cr
+   threshold \tab numeric \tab - \tab  \tab \eqn{(-\infty, \infty)}{(-Inf, Inf)} \cr
+   threshold.scale \tab numeric \tab - \tab  \tab \eqn{(-\infty, \infty)}{(-Inf, Inf)} \cr
+}
+}
+
+\examples{
+\dontshow{if (mlr3misc::require_namespaces(lrn("classif.pamr")$packages, quietly = TRUE)) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
+# Define the Learner
+learner = mlr3::lrn("classif.pamr")
+print(learner)
+
+# Define a Task
+task = mlr3::tsk("sonar")
+
+# Create train and test set
+ids = mlr3::partition(task)
+
+# Train the learner on the training ids
+learner$train(task, row_ids = ids$train)
+
+print(learner$model)
+
+
+# Make predictions for the test rows
+predictions = learner$predict(task, row_ids = ids$test)
+
+# Score the predictions
+predictions$score()
+\dontshow{\}) # examplesIf}
+}
+\references{
+\verb{r format_bib("tibshirani2002diagnosis")}
+}
+\seealso{
+\itemize{
[…]
