+++ man/mlr_learners_classif.penalized.Rd
@@ -1,34 +1,23 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/learner_RWeka_regr_decision_stump.R
-\name{mlr_learners_regr.decision_stump}
-\alias{mlr_learners_regr.decision_stump}
-\alias{LearnerRegrDecisionStump}
-\title{Regression Decision Stump Learner}
+% Please edit documentation in R/learner_penalized_classif_penalized.R
+\name{mlr_learners_classif.penalized}
+\alias{mlr_learners_classif.penalized}
+\alias{LearnerClassifPenalized}
+\title{Classification L1 and L2 Penalized Learner}
 \description{
-Decision Stump Learner.
-Calls \code{\link[RWeka:DecisionStump]{RWeka::DecisionStump()}} from \CRANpkg{RWeka}.
+Penalized (L1 and L2) Logistic Regression.
+Calls \code{\link[penalized:penalized]{penalized::penalized()}} from: \CRANpkg{penalized}.
 }
-\section{Custom mlr3 parameters}{
+\details{
+The \code{penalized} and \code{unpenalized} arguments in the learner are implemented slightly
+differently than in \code{\link[penalized:penalized]{penalized::penalized()}}. Here, there is no parameter for \code{penalized} but
+instead it is assumed that every variable is penalized unless stated in the \code{unpenalized}
+parameter.
+}
+\section{Initial parameter values}{
 
 \itemize{
-\item \code{output_debug_info}:
-\itemize{
-\item original id: output-debug-info
-}
-\item \code{do_not_check_capabilities}:
-\itemize{
-\item original id: do-not-check-capabilities
-}
-\item \code{num_decimal_places}:
-\itemize{
-\item original id: num-decimal-places
-}
-\item \code{batch_size}:
-\itemize{
-\item original id: batch-size
-}
-\item Reason for change: This learner contains changed ids of the following control arguments
-since their ids contain irregular pattern
+\item \code{trace} is set to \code{"FALSE"} to disable printing output during model training.
 }
 }
 
@@ -36,41 +25,46 @@ since their ids contain irregular pattern
 
 This \link[mlr3:Learner]{Learner} can be instantiated via \link[mlr3:mlr_sugar]{lrn()}:
 
-\if{html}{\out{<div class="sourceCode">}}\preformatted{lrn("regr.decision_stump")
+\if{html}{\out{<div class="sourceCode">}}\preformatted{lrn("classif.penalized")
 }\if{html}{\out{</div>}}
 }
 
 \section{Meta Information}{
 
 \itemize{
-\item Task type: \dQuote{regr}
-\item Predict Types: \dQuote{response}
+\item Task type: \dQuote{classif}
+\item Predict Types: \dQuote{response}, \dQuote{prob}
 \item Feature Types: \dQuote{logical}, \dQuote{integer}, \dQuote{numeric}, \dQuote{factor}, \dQuote{ordered}
-\item Required Packages: \CRANpkg{mlr3}, \CRANpkg{RWeka}
+\item Required Packages: \CRANpkg{mlr3}, \CRANpkg{penalized}
 }
 }
 
 \section{Parameters}{
 \tabular{lllll}{
    Id \tab Type \tab Default \tab Levels \tab Range \cr
-   subset \tab untyped \tab - \tab  \tab - \cr
[…]
