+++ NAMESPACE
@@ -40,6 +40,7 @@ export(LearnerClassifNaiveBayesWeka)
 export(LearnerClassifObliqueRandomForest)
 export(LearnerClassifOneR)
 export(LearnerClassifPART)
+export(LearnerClassifPenalized)
 export(LearnerClassifPriorityLasso)
 export(LearnerClassifREPTree)
 export(LearnerClassifRandomForest)
