+++ NAMESPACE
@@ -43,7 +43,6 @@ export(LearnerClassifNaiveBayesWeka)
 export(LearnerClassifObliqueRandomForest)
 export(LearnerClassifOneR)
 export(LearnerClassifPART)
-export(LearnerClassifPamr)
 export(LearnerClassifPriorityLasso)
 export(LearnerClassifREPTree)
 export(LearnerClassifRandomForest)
