### 2025-07-04 48712e7
Fix parametric (#444)
Remove `discrete = FALSE` option, only return discrete survival
predictions, works with `mlr3proba` v0.8.1

---------

Co-authored-by: Sebastian Fischer <sebf.fischer@gmail.com>
