+++ R/learner_lightgbm_classif_lightgbm.R
@@ -404,10 +404,11 @@ LearnerClassifLightGBM = R6Class("LearnerClassifLightGBM",
 
       if ("multiclass" %in% task$properties) {
         pred_mat = pred
-        colnames(pred_mat) = self$state$train_task$levels()[[self$state$train_task$target_names]]
+        labels = self$state$train_task$levels()[[self$state$train_task$target_names]]
+        colnames(pred_mat) = labels
         if (self$predict_type == "response") {
           which = apply(pred_mat, 1, which.max)
-          response = self$state$labels[which]
+          response = labels[which]
           pred_mat = NULL
         }
       } else {
@@ -450,7 +451,6 @@ LearnerClassifLightGBM = R6Class("LearnerClassifLightGBM",
       # set number of classes if multiclass and save label ordering
       if (pars_train$objective %in% c("multiclass", "multiclassova")) {
         pars_train$num_class = length(task$class_names)
-        self$state$labels = unique(task$truth())
       }
 
       if (pars_train$objective %in% c("multiclass", "multiclassova")) {
