+++ tests/testthat/test_survivalmodels_surv_parametric.R
@@ -8,7 +8,7 @@ test_that("autotest aft", {
 
 test_that("autotest ph", {
   set.seed(1)
-  learner = lrn("surv.parametric", form = "ph", discrete = FALSE)
+  learner = lrn("surv.parametric", form = "ph")
   expect_learner(learner)
   result = run_autotest(learner, check_replicable = FALSE, exclude = "utf8_feature_names")
   expect_true(result, info = result$error)
@@ -45,167 +45,46 @@ test_that("time points for prediction", {
   expect_equal(length(times), 50)
 })
 
-test_that("manualtest - aft", {
-  learner = lrn("surv.parametric", dist = "weibull", form = "aft", discrete = FALSE)
-  expect_silent(learner$train(task))
-  p = learner$predict(task)
-  expect_prediction_surv(p)
-  expect_equal(-p$lp, predict(learner$model$model, type = "lp"))
-  expect_equal(p$distr[1]$survival(predict(
-    learner$model$model, type = "quantile", p = c(0.2, 0.8)
-  )[1, ]), c(0.8, 0.2))
-  expect_equal(p$distr[10]$cdf(predict(
-    learner$model$model, type = "quantile", p = seq.int(0, 1, 0.1)
-  )[10, ]),
-  seq.int(0, 1, 0.1))
-
-  learner = lrn("surv.parametric", dist = "lognormal", form = "aft", discrete = FALSE)$train(task)
-  p = learner$predict(task)
-  expect_equal(p$distr[15]$cdf(predict(
-    learner$model$model, type = "quantile", p = seq.int(0, 1, 0.1)
-  )[15, ]), seq.int(0, 1, 0.1))
-})
-
 test_that("missing", {
   learner = lrn("surv.parametric")
   learner$train(task)
   expect_error(learner$predict(lung_missings))
 })
 
-test_that("quantile type", {
-  learner = lrn("surv.parametric", dist = "weibull", form = "aft", discrete = FALSE)$train(task)
-  p = lrn("surv.parametric", dist = "weibull", form = "aft", discrete = FALSE)$train(task)$predict(task)
-  quantile = p$distr$quantile(c(0.2, 0.8))
-  expect_equal(matrix(t(quantile), ncol = 2),
-               predict(learner$model$model, type = "quantile", p = c(0.2, 0.8)))
-  quantile = p$distr$quantile(0.5)
-  expect_equal(unlist(p$distr$cdf(quantile), use.names = FALSE), rep(0.5, task$nrow))
+test_that("'form' affects survival prediction", {
+  # aft
+  p = lrn("surv.parametric", form = "aft")$train(task)$predict(task)
+  expect_true(inherits(p$distr, "Matdist"))
+  surv_aft = p$distr$survival(task$unique_times())
+  expect_matrix(surv_aft, nrows = length(task$unique_times()), ncols = task$nrow,
+                any.missing = FALSE) # [times x obs]
 
-  p = lrn("surv.parametric", dist = "weibull", form = "ph", discrete = FALSE)$train(task)$predict(task)
-  quantile = p$distr$quantile(0.5)
-  expect_equal(unlist(p$distr$cdf(quantile), use.names = FALSE), rep(0.5, task$nrow))
+  # tobit
+  p = lrn("surv.parametric", form = "tobit")$train(task)$predict(task)
+  expect_true(inherits(p$distr, "Matdist"))
+  surv_tobit = p$distr$survival(task$unique_times())
+  expect_matrix(surv_tobit, nrows = length(task$unique_times()), ncols = task$nrow,
+                any.missing = FALSE) # [times x obs]
 
[…]
