+++ NAMESPACE
@@ -92,6 +92,7 @@ export(LearnerRegrMars)
 export(LearnerRegrMob)
 export(LearnerRegrMultilayerPerceptron)
 export(LearnerRegrObliqueRandomForest)
+export(LearnerRegrPenalized)
 export(LearnerRegrPriorityLasso)
 export(LearnerRegrQGam)
 export(LearnerRegrREPTree)
