+++ tests/testthat/test_fastai_classif_fastai.R
@@ -1,215 +1,215 @@
-# reticulate::use_condaenv("r-reticulate")
-
-test_that("autotest", {
-  learner = lrn("classif.fastai", layers = c(200, 100))
-  expect_learner(learner, check_man = FALSE)
-  # results not replicable, because torch seed must be set in the python backend
-  result = run_autotest(learner, check_replicable = FALSE)
-  expect_true(result, info = result$error)
-})
-
-test_that("eval protocol", {
-  learner = lrn("classif.fastai")
-  task = tsk("sonar")
-  learner$train(task)
-  expect_true(is.list(learner$state$eval_protocol))
-})
-
-
-test_that("validation and inner tuning works", {
-  task = tsk("spam")
-
-  # with n_epoch and patience parameter
-  learner = lrn("classif.fastai",
-    n_epoch = 10,
-    patience = 1,
-    validate = 0.2
-  )
-
-  learner$train(task)
-  expect_named(learner$state$eval_protocol, c("epoch", "train_loss", "valid_loss", "accuracy"))
-  expect_list(learner$internal_valid_scores, types = "numeric")
-  expect_equal(names(learner$internal_valid_scores), "accuracy")
-  expect_list(learner$internal_tuned_values, types = "integerish")
-  expect_equal(names(learner$internal_tuned_values), "n_epoch")
-
-  # without validate parameter
-  learner$validate = NULL
-  expect_error(learner$train(task), "field 'validate'")
-
-  # with patience parameter
-  learner = lrn("classif.fastai",
-    n_epoch = 10,
-    validate = 0.2
-  )
-
-  learner$train(task)
-  expect_equal(learner$internal_tuned_values, NULL)
-  expect_named(learner$state$eval_protocol, c("epoch", "train_loss", "valid_loss", "accuracy"))
-  expect_list(learner$internal_valid_scores, types = "numeric")
-  expect_equal(names(learner$internal_valid_scores), "accuracy")
-
-  # internal tuning
-  learner = lrn("classif.fastai",
-    n_epoch = to_tune(upper = 20, internal = TRUE),
-    validate = 0.2
-  )
-  s = learner$param_set$search_space()
-  expect_error(learner$param_set$convert_internal_search_space(s), "patience")
-  learner$param_set$set_values(n_epoch = 10)
-  learner$param_set$disable_internal_tuning("n_epoch")
-  expect_equal(learner$param_set$values$n_epoch, NULL)
-
-  learner = lrn("classif.fastai",
-    n_epoch = 20,
-    patience = 5,
-    validate = 0.3
-  )
-  learner$train(task)
-  expect_equal(learner$internal_valid_scores$accuracy, learner$state$eval_protocol$accuracy[learner$internal_tuned_values$n_epoch])
-
-  # no validation and no internal tuning
-  learner = lrn("classif.fastai")
-  learner$train(task)
-  expect_null(learner$internal_valid_scores)
-  expect_null(learner$internal_tuned_values)
-
-  # no tuning without patience parameter
-  learner = lrn("classif.fastai", validate = 0.3, n_epoch = 10)
-  learner$train(task)
-  expect_equal(learner$internal_valid_scores$accuracy, learner$state$eval_protocol$accuracy[10L])
-  expect_null(learner$internal_tuned_values)
-})
-
-test_that("custom inner validation measure", {
-  # internal measure
-  task = tsk("sonar")
-
-  learner = lrn("classif.fastai",
-    n_epoch = 10,
-    validate = 0.2,
-    patience = 1,
-    eval_metric = fastai::error_rate()
-  )
-
-  learner$train(task)
-
[…]
