### 2025-07-04 a1e2818
update NEWS.md
