### 2025-07-04 239d872
quick param set fix
