### 2025-07-04 cd4cc15
add blockforest learner + doc
