+++ man/mlr_learners_surv.blockforest.Rd
@@ -1,0 +1,176 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/learner_blockForest_surv_blockforest.R
+\name{mlr_learners_surv.blockforest}
+\alias{mlr_learners_surv.blockforest}
+\alias{LearnerSurvBlockForest}
+\title{BlockForest Survival Learner}
+\description{
+Random survival forests for blocks of clinical and omics covariate data.
+Calls \code{\link[blockForest:blockfor]{blockForest::blockfor()}} from package \CRANpkg{blockForest}.
+}
+\section{Prediction types}{
+
+This learner returns two prediction types:
+\enumerate{
+\item \code{distr}: a survival matrix in two dimensions, where observations are
+represented in rows and (unique event) time points in columns.
+Calculated using the internal \code{blockForest:::predict.blockForest()} function.
+\item \code{crank}: the expected mortality using \code{\link[mlr3proba:dot-surv_return]{mlr3proba::.surv_return()}}.
+}
+}
+
+\section{Initial parameter values}{
+
+\itemize{
+\item \code{num.threads} is initialized to 1 to avoid conflicts with parallelization via \CRANpkg{future}.
+}
+}
+
+\section{Dictionary}{
+
+This \link[mlr3:Learner]{Learner} can be instantiated via \link[mlr3:mlr_sugar]{lrn()}:
+
+\if{html}{\out{<div class="sourceCode">}}\preformatted{lrn("surv.blockforest")
+}\if{html}{\out{</div>}}
+}
+
+\section{Meta Information}{
+
+\itemize{
+\item Task type: \dQuote{surv}
+\item Predict Types: \dQuote{crank}, \dQuote{distr}
+\item Feature Types: \dQuote{logical}, \dQuote{integer}, \dQuote{numeric}, \dQuote{character}, \dQuote{factor}, \dQuote{ordered}
+\item Required Packages: \CRANpkg{mlr3}, \CRANpkg{mlr3proba}, \CRANpkg{mlr3extralearners}, \CRANpkg{blockForest}
+}
+}
+
+\section{Parameters}{
+\tabular{lllll}{
+   Id \tab Type \tab Default \tab Levels \tab Range \cr
+   blocks \tab untyped \tab - \tab  \tab - \cr
+   block.method \tab character \tab BlockForest \tab BlockForest, RandomBlock, BlockVarSel, VarProb, SplitWeights \tab - \cr
+   num.trees \tab integer \tab 2000 \tab  \tab \eqn{[1, \infty)}{[1, Inf)} \cr
+   mtry \tab untyped \tab NULL \tab  \tab - \cr
+   nsets \tab integer \tab 300 \tab  \tab \eqn{[1, \infty)}{[1, Inf)} \cr
+   num.trees.pre \tab integer \tab 1500 \tab  \tab \eqn{[1, \infty)}{[1, Inf)} \cr
+   splitrule \tab character \tab extratrees \tab logrank, extratrees, C, maxstat \tab - \cr
+   always.select.block \tab integer \tab 0 \tab  \tab \eqn{[0, 1]}{[0, 1]} \cr
+   importance \tab character \tab - \tab none, impurity, impurity_corrected, permutation \tab - \cr
+   num.threads \tab integer \tab 1 \tab  \tab \eqn{[1, \infty)}{[1, Inf)} \cr
+}
+}
+
+\examples{
+\dontshow{if (mlr3misc::require_namespaces(c("mlr3proba"), quietly = TRUE)) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
+# Define a Task
+task = tsk("grace")
+# Create train and test set
+ids = partition(task)
+# check task's features
+task$feature_names
[…]
