### 2025-07-04 1296b8c
update news
