+++ DESCRIPTION
@@ -18,7 +18,7 @@ Authors@R: c(
     person("Byron", "Jaeger", , "bjaeger@wakehealth.edu", role = "ctb"),
     person("Sebastian", "Fischer", , "sebf.fischer@gmail.com", role = c("cre", "aut")),
     person("Zezhi", "Wang", ,"homura@mail.ustc.edu.cn", role = "ctb"),
-    person("John", "Zobolas", ,"bblodfon@gmail.com", role = "ctb",
+    person("John", "Zobolas", ,"bblodfon@gmail.com", role = "aut",
            comment = c(ORCID = "0000-0002-3609-8674")),
     person("Lukas", "Burk", , "burk@leibniz-bips.de", role = "ctb",
            comment = c(ORCID = "0000-0001-7528-3795")),
