+++ tests/testthat/test_prioritylasso_classif_priority_lasso.R
@@ -1,7 +1,5 @@
 test_that("autotest", {
-  # These shenanigans are necessary because we have to dynanically set the blocks, depending on the task
   set.seed(1)
-  # blocks gets changed later anyway but is required
   learner = lrn("classif.priority_lasso", type.measure = "class", blocks = "PLACEHOLDER")
   on.exit({
     assignInNamespace(".__LearnerClassifPriorityLasso__.train", train_old, ns = "mlr3extralearners")
