### 2025-07-05 c0c1fba
add back the arrows in the tests
