+++ tests/testthat/test_prioritylasso_surv_priority_lasso.R
@@ -10,8 +10,8 @@ test_that("autotest", {
 
   src = as.list(body(train_old))
   new_lines = list(
-    quote(s = seq_along(task$feature_names)),
-    quote(pars$blocks = set_names(list(s), "bp1"))
+    quote(s <- seq_along(task$feature_names)),
+    quote(pars$blocks <- set_names(list(s), "bp1"))
   )
   src = c(src[1:2], new_lines, src[3:length(src)])
   new_body = as.call(src)
