### 2025-07-05 ffc34ba
simplify test with the use of configure_learner arg
