### 2025-07-05 b9ecd6a
refactor priority lasso tests using configure_learner
