+++ tests/testthat/test_prioritylasso_classif_priority_lasso.R
@@ -1,22 +1,13 @@
 test_that("autotest", {
   set.seed(1)
-  learner = lrn("classif.priority_lasso", type.measure = "class", blocks = "PLACEHOLDER")
-  on.exit({
-    assignInNamespace(".__LearnerClassifPriorityLasso__.train", train_old, ns = "mlr3extralearners")
-  }, add = TRUE)
-  train_old = mlr3extralearners:::.__LearnerClassifPriorityLasso__.train # nolint
+  learner = lrn("classif.priority_lasso", type.measure = "class")
 
-  src = as.list(body(train_old))
-  new_lines = list(
-    quote(s <- seq_along(task$feature_names)),
-    quote(pars$blocks <- set_names(list(s), "bp1"))
-  )
-  src = c(src[1:2], new_lines, src[3:length(src)])
-  new_body = as.call(src)
-  train = train_old
-  body(train) = new_body
-  assignInNamespace(".__LearnerClassifPriorityLasso__.train", train, ns = "mlr3extralearners")
+  # Dynamically define a single block containing all features for the given task
+  cfg_lrn = function(learner, task) {
+    s = seq_along(task$feature_names)
+    learner$param_set$set_values(blocks = set_names(list(s), "bp1"))
+  }
 
-  result = run_autotest(learner, exclude = "feat_single")
+  result = run_autotest(learner, exclude = "feat_single", configure_learner = cfg_lrn)
   expect_true(result, info = result$error)
 })
