+++ tests/testthat/test_prioritylasso_surv_priority_lasso.R
@@ -2,7 +2,7 @@ test_that("autotest", {
   set.seed(1)
   learner = lrn("surv.priority_lasso")
 
-  # dynamically set the blocks, depending on the task
+  # Dynamically define a single block containing all features for the given task
   cfg_lrn = function(learner, task) {
     s = seq_along(task$feature_names)
     learner$param_set$set_values(blocks = set_names(list(s), "bp1"))
