### 2025-07-05 6a8706d
add seed and verbose learner params
