+++ R/learner_blockForest_surv_blockforest.R
@@ -61,8 +61,11 @@ LearnerSurvBlockForest = R6::R6Class("LearnerSurvBlockForest",
         num.trees.pre       = p_int(1L, default = 1500L, tags = "train"),
         splitrule           = p_fct(c("logrank", "extratrees", "C", "maxstat"), default = "extratrees", tags = "train"),
         always.select.block = p_int(0L, 1L, default = 0L, tags = "train"),
+        # adding few key parameters from blockForest::blockForest()
         importance          = p_fct(c("none", "impurity", "impurity_corrected", "permutation"), tags = "train"),
-        num.threads         = p_int(1L, default = 1L, tags = c("train", "predict", "threads"))
+        num.threads         = p_int(1L, default = 1L, tags = c("train", "predict", "threads")),
+        seed                = p_int(default = NULL, special_vals = list(NULL), tags = c("train", "predict")),
+        verbose             = p_lgl(default = TRUE, tags = c("train", "predict"))
       )
 
       param_set$values = list(num.threads = 1)
