### 2025-07-05 51d508e
add tests for surv.blockforest
