### 2025-07-06 466c53e
remove character type features as they are not supported
