+++ R/learner_blockForest_surv_blockforest.R
@@ -74,7 +74,7 @@ LearnerSurvBlockForest = R6::R6Class("LearnerSurvBlockForest",
         id = "surv.blockforest",
         param_set = param_set,
         predict_types = c("crank", "distr"),
-        feature_types = c("logical", "integer", "numeric", "character", "factor", "ordered"),
+        feature_types = c("logical", "integer", "numeric", "factor", "ordered"),
         properties = c("weights", "importance"),
         packages = c("mlr3extralearners", "blockForest"),
         label = "Block Forests: Random Forests for Blocks of Clinical and Omics Covariate Data"
