+++ tests/testthat/test_blockForest_surv_blockforest.R
@@ -1,6 +1,7 @@
 test_that("autotest", {
   set.seed(1)
-  learner = lrn("surv.blockforest", nsets = 5, num.trees = 10, num.trees.pre = 5)
+  learner = lrn("surv.blockforest", nsets = 5, num.trees = 10,
+                num.trees.pre = 5, importance = "permutation")
 
   # Dynamically define a single block containing all features for the given task
   cfg_lrn = function(learner, task) {
@@ -8,6 +9,7 @@ test_that("autotest", {
     learner$param_set$set_values(blocks = set_names(list(s), "bp1"))
   }
 
-  result = run_autotest(learner, exclude = "feat_single", configure_learner = cfg_lrn)
-  expect_true(result, info = result$error)
+  # 1 feature task for this learner doesn't make sense (1 block => at least 2 features)
+  result = run_autotest(learner, exclude = "feat_single|utf8_feature_names", configure_learner = cfg_lrn)
+  expect_true(result)
 })
