### 2025-07-06 6a6ecc0
refactor tests: all learners initialized with lrn()
