+++ tests/testthat/test_survivalmodels_surv_akritas.R
@@ -1,10 +1,12 @@
 test_that("autotest", {
-  learner = LearnerSurvAkritas$new()
-  expect_learner(learner)
-  result = run_autotest(learner, check_replicable = FALSE, exclude = "sanity")
-  # sanity test disabled, because akritas is similar to the Kaplan-Meier in that dependent on parameter it always predicts 0.5
-  # whereas the sanity check requires >= 0.6
-  expect_true(result, info = result$error)
+  with_seed(1, {
+    learner = lrn("surv.akritas", lambda = 0.5)
+    expect_learner(learner)
+    result = run_autotest(learner, check_replicable = FALSE)
+    # sanity test disabled => TO CHECK: akritas default uses lambda = 0.5, which
+    # can better than KM!
+    expect_true(result, info = result$error)
+  })
 })
 
 test_that("time points for prediction", {
