+++ tests/testthat/test_plugdensity_dens_plug.R
@@ -1,8 +1,8 @@
 test_that("autotest", {
   set.seed(1)
-  learner = LearnerDensPlugin$new()
+  learner = lrn("dens.plug")
   expect_learner(learner)
-  # mrl3proba's generate_task does not respect feature types
-  result = run_autotest(learner, exclude = "integer")
+  # mlr3proba's generate_task does not respect feature types
+  result = run_autotest(learner) # exclude = "integer" (CHECK IF IT WORKS)
   expect_true(result, info = result$error)
 })
