+++ tests/testthat/test_randomForestSRC_surv_rfsrc.R
@@ -12,14 +12,14 @@ test_that("autotest", {
 test_that("importance/selected", {
   set.seed(1)
   task = tsk("rats")
-  learn = LearnerSurvRandomForestSRC$new()
-  learn$param_set$values = list(estimator = "kaplan")
-  learn$train(task)
-  expect_error(learn$importance(), "Set 'importance'")
-  expect_error(learn$selected_features(), "Set 'var.used'")
-  learn$param_set$values = list(estimator = "nelson", var.used = "all.trees", importance = "random")
-  learn$train(task)
-  expect_silent(learn$selected_features())
-  expect_silent(learn$importance())
-  expect_silent(learn$oob_error())
+  learner = lrn("surv.rfsrc", estimator = "kaplan")
+  learner$train(task)
+  expect_error(learner$importance(), "Set 'importance'")
+  expect_error(learner$selected_features(), "Set 'var.used'")
+  learner$param_set$values = list(
+    estimator = "nelson", var.used = "all.trees", importance = "random"
+  )
+  learner$train(task)
+  expect_silent(learner$selected_features())
+  expect_silent(learner$importance())
 })
