### 2025-07-06 2e0690b
update learner parameter doc
