### 2025-07-06 f80b294
remove comments from two tests
