+++ tests/testthat/test_plugdensity_dens_plug.R
@@ -2,7 +2,6 @@ test_that("autotest", {
   set.seed(1)
   learner = lrn("dens.plug")
   expect_learner(learner)
-  # mlr3proba's generate_task does not respect feature types
-  result = run_autotest(learner) # exclude = "integer" (CHECK IF IT WORKS)
+  result = run_autotest(learner)
   expect_true(result, info = result$error)
 })
