+++ NEWS.md
@@ -1,3 +1,5 @@
+# mlr3extralearners (development version)
+
 # mlr3extralearners 1.1.0
 
 New Features:
