### 2025-07-07 9ab68d8
remove default tag from num.threads
