+++ R/learner_blockForest_surv_blockforest.R
@@ -75,7 +75,7 @@ LearnerSurvBlockForest = R6::R6Class("LearnerSurvBlockForest",
         feature_types = c("logical", "integer", "numeric", "factor", "ordered"),
         properties = c("weights", "importance"),
         packages = c("mlr3extralearners", "blockForest"),
-        label = "Block Forests: Random Forests for Blocks of Clinical and Omics Covariate Data"
+        label = "Block Forests: Random Forests for Block-wise Data"
       )
     },
 
