### 2025-07-07 6423d95
need mlr3 dev for configure_learner() in the tests
