+++ NAMESPACE
@@ -108,6 +108,7 @@ export(LearnerRegrTabPFN)
 export(LearnerSurvAkritas)
 export(LearnerSurvAorsf)
 export(LearnerSurvBlackBoost)
+export(LearnerSurvBlockForest)
 export(LearnerSurvCForest)
 export(LearnerSurvCTree)
 export(LearnerSurvCVCoxboost)
