### 2025-07-08 0693270
add Competing risks RSF paper
