### 2025-07-08 86c919d
reformat learner parameters
