+++ R/learner_randomForestSRC_surv_rfsrc.R
@@ -66,83 +66,54 @@ LearnerSurvRandomForestSRC = R6Class("LearnerSurvRandomForestSRC",
     #' @description
     #' Creates a new instance of this [R6][R6::R6Class] class.
     initialize = function() {
-      ps = ps(
-        ntree = p_int(default = 500L, lower = 1L, tags = "train"),
-        mtry = p_int(lower = 1L, tags = "train"),
-        mtry.ratio = p_dbl(lower = 0, upper = 1, tags = "train"),
-        nodesize = p_int(default = 15L, lower = 1L, tags = "train"),
-        nodedepth = p_int(lower = 1L, tags = "train"),
-        splitrule = p_fct(
-          levels = c("logrank", "bs.gradient", "logrankscore"),
-          default = "logrank", tags = "train"),
-        nsplit = p_int(lower = 0, default = 10, tags = "train"),
-        importance = p_fct(
-          default = "FALSE",
-          levels = c("FALSE", "TRUE", "none", "permute", "random", "anti"),
-          tags = c("train", "predict")),
-        block.size = p_int(default = 10L, lower = 1L, tags = c("train", "predict")),
-        bootstrap = p_fct(
-          default = "by.root",
-          levels = c("by.root", "by.node", "none", "by.user"), tags = "train"),
-        samptype = p_fct(
-          default = "swor", levels = c("swor", "swr"),
-          tags = "train"),
-        samp = p_uty(tags = "train"),
-        membership = p_lgl(default = FALSE, tags = c("train", "predict")),
-        sampsize = p_uty(tags = "train"),
+      param_set = ps(
+        ntree          = p_int(default = 500L, lower = 1L, tags = "train"),
+        mtry           = p_int(lower = 1L, tags = "train"),
+        mtry.ratio     = p_dbl(lower = 0, upper = 1, tags = "train"),
+        nodesize       = p_int(default = 15L, lower = 1L, tags = "train"),
+        nodedepth      = p_int(lower = 1L, tags = "train"),
+        splitrule      = p_fct(levels = c("logrank", "bs.gradient", "logrankscore"), default = "logrank", tags = "train"),
+        nsplit         = p_int(lower = 0, default = 10, tags = "train"),
+        importance     = p_fct(default = "FALSE", levels = c("FALSE", "TRUE", "none", "anti", "permute", "random"), tags = c("train", "predict")), #nolint
+        block.size     = p_int(default = 10L, lower = 1L, tags = c("train", "predict")),
+        bootstrap      = p_fct(default = "by.root", levels = c("by.root", "by.node", "none", "by.user"), tags = "train"), #nolint
+        samptype       = p_fct(default = "swor", levels = c("swor", "swr"), tags = "train"),
+        samp           = p_uty(tags = "train"),
+        membership     = p_lgl(default = FALSE, tags = c("train", "predict")),
+        sampsize       = p_uty(tags = "train"),
         sampsize.ratio = p_dbl(0, 1, tags = "train"),
-        na.action = p_fct(
-          default = "na.omit", levels = c("na.omit", "na.impute"),
-          tags = c("train", "predict")),
-        nimpute = p_int(lower = 1L, default = 1L, special_vals = list(NULL), tags = "train"),
-        ntime = p_int(lower = 0L, default = 150L, special_vals = list(NULL), tags = "train"),
-        cause = p_int(lower = 1L, tags = "train"),
-        proximity = p_fct(
-          default = "FALSE",
-          levels = c("FALSE", "TRUE", "inbag", "oob", "all"),
-          tags = c("train", "predict")),
-        distance = p_fct(
-          default = "FALSE",
-          levels = c("FALSE", "TRUE", "inbag", "oob", "all"),
-          tags = c("train", "predict")),
[…]
