### 2025-07-08 3af9aeb
refactor test, use with_seed
