+++ tests/testthat/test_randomForestSRC_surv_rfsrc.R
@@ -1,17 +1,15 @@
 test_that("autotest", {
-  learner = lrn("surv.rfsrc")
-  learner$param_set$values = insert_named(
-    learner$param_set$values,
-    list(importance = "random", na.action = "na.impute"))
-  expect_learner(learner)
-  set.seed(1)
-  result = run_autotest(learner, check_replicable = FALSE, N = 100)
-  expect_true(result, info = result$error)
+  with_seed(1, {
+    learner = lrn("surv.rfsrc", importance = "random", na.action = "na.impute")
+    expect_learner(learner)
+    result = run_autotest(learner, check_replicable = FALSE, N = 100)
+    expect_true(result, info = result$error)
+  })
 })
 
 test_that("importance/selected", {
-  set.seed(1)
-  task = tsk("rats")
+  with_seed(1, {
+    task = tsk("rats")
   learner = lrn("surv.rfsrc", estimator = "kaplan")
   learner$train(task)
   expect_error(learner$importance(), "Set 'importance'")
@@ -22,4 +20,5 @@ test_that("importance/selected", {
   learner$train(task)
   expect_silent(learner$selected_features())
   expect_silent(learner$importance())
+  })
 })
