+++ R/learner_randomForestSRC_surv_rfsrc.R
@@ -158,7 +158,7 @@ LearnerSurvRandomForestSRC = R6Class("LearnerSurvRandomForestSRC",
       pv = convert_ratio(pv, "sampsize", "sampsize.ratio", task$nrow)
       cores = pv$cores %??% 1L
       pv$cores = NULL
-      pv$case.wt = private$.get_weights(task) # nolint
+      pv$case.wt = private$.get_weights(task)
 
       invoke(randomForestSRC::rfsrc,
         formula = task$formula(), data = task$data(),
