### 2025-07-08 cfbf996
fix: add L to integer params, add NULL to random_state special_vals
