+++ DESCRIPTION
@@ -28,7 +28,7 @@ Authors@R: c(
 Description: Extra learners for use in mlr3.
 License: LGPL-3
 Depends:
-    R (>= 3.1.0),
+    R (>= 3.3.0),
     mlr3 (>= 1.0.0)
 Imports:
     checkmate,
