+++ R/learner_randomForestSRC_classif_imbalanced_rfsrc.R
@@ -83,7 +83,6 @@ LearnerClassifImbalancedRandomForestSRC = R6Class("LearnerClassifImbalancedRando
             tags = c("train", "predict")),
         nimpute = p_int(default = 1L, lower = 1L, tags = "train"),
         ntime = p_int(lower = 1L, tags = "train"),
-        cause = p_int(lower = 1L, tags = "train"),
         proximity = p_fct(
           default = "FALSE",
           levels = c("FALSE", "TRUE", "inbag", "oob", "all"),
