+++ R/learner_randomForestSRC_classif_rfsrc.R
@@ -82,7 +82,6 @@ LearnerClassifRandomForestSRC = R6Class("LearnerClassifRandomForestSRC",
           tags = c("train", "predict")),
         nimpute = p_int(default = 1L, lower = 1L, tags = "train"),
         ntime = p_int(lower = 1L, tags = "train"),
-        cause = p_int(lower = 1L, tags = "train"),
         proximity = p_fct(
           default = "FALSE",
           levels = c("FALSE", "TRUE", "inbag", "oob", "all"),
