+++ tests/testthat/test_paramtest_randomForestSRC_regr_rfsrc.R
@@ -9,7 +9,8 @@ test_that("paramtest regr.rfsrc train", {
     "case.wt", # handled by task weights
     "cores", # set as option(rf.cores)
     "sampsize.ratio", # alternative to sampsize
-    "mtry.ratio" # alternative to mtry
+    "mtry.ratio", # alternative to mtry
+    "cause" # for competing risk forests only
   )
 
   paramtest = run_paramtest(learner, fun_list, exclude, tag = "train")
