### 2025-07-09 013baf4
fix: case.depth is also a train parameter!
