+++ tests/testthat/test_paramtest_randomForestSRC_classif_rfsrc.R
@@ -17,7 +17,6 @@ test_that("paramtest classif.rfsrc train", {
   expect_paramtest(paramtest)
 })
 
-
 test_that("paramtest classif.rfsrc predict", {
   learner = lrn("classif.rfsrc")
   fun_list = list(randomForestSRC:::predict.rfsrc) # nolint
