### 2025-07-10 33670e4
feat: Classification learners for rda and sda from old mlr2 (#427)
Adding:
- Classification learner for regularized discriminant analysis from
package [klaR](https://cran.r-project.org/web/packages/klaR/index.html).
- Classification learner for shrinkage discriminant analysis from
package [sda](https://cran.r-project.org/web/packages/sda/index.html).

Both learners were previously implemented in
[mlr2](https://github.com/mlr-org/mlr).

---------

Co-authored-by: Sebastian Fischer <sebf.fischer@gmail.com>
Co-authored-by: Marc Becker <33069354+be-marc@users.noreply.github.com>
