+++ DESCRIPTION
@@ -63,6 +63,7 @@ Suggests:
     jsonlite,
     keras (>= 2.3.0),
     kernlab,
+    klaR,
     knitr,
     ks,
     LiblineaR,
@@ -101,6 +102,7 @@ Suggests:
     RWeka,
     sandwich,
     set6,
+    sda,
     sm,
     stats,
     survival,
