### 2025-07-10 a48366d
logrankscore is a compering risks alternative, not survival
