+++ man/mlr_learners_cmprsk.rfsrc.Rd
@@ -1,39 +1,28 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/learner_randomForestSRC_surv_rfsrc.R
-\name{mlr_learners_surv.rfsrc}
-\alias{mlr_learners_surv.rfsrc}
-\alias{LearnerSurvRandomForestSRC}
-\title{Survival Random Forest SRC Learner}
+% Please edit documentation in R/learner_randomForestSRC_cmprsk_rfsrc.R
+\name{mlr_learners_cmprsk.rfsrc}
+\alias{mlr_learners_cmprsk.rfsrc}
+\alias{LearnerCompRisksRandomForestSRC}
+\title{Survival Random Forest Competing Risks Learner}
 \description{
-Random survival forest.
+Random survival forests for competing risks.
 Calls \code{\link[randomForestSRC:rfsrc]{randomForestSRC::rfsrc()}} from \CRANpkg{randomForestSRC}.
 }
-\section{Prediction types}{
-
-This learner returns two prediction types:
-\enumerate{
-\item \code{distr}: a survival matrix in two dimensions, where observations are
-represented in rows and (unique event) time points in columns.
-Calculated using the internal \code{\link[randomForestSRC:predict.rfsrc]{randomForestSRC::predict.rfsrc()}} function.
-\item \code{crank}: the expected mortality using \code{\link[mlr3proba:dot-surv_return]{mlr3proba::.surv_return()}}.
-}
-}
-
 \section{Dictionary}{
 
 This \link[mlr3:Learner]{Learner} can be instantiated via \link[mlr3:mlr_sugar]{lrn()}:
 
-\if{html}{\out{<div class="sourceCode">}}\preformatted{lrn("surv.rfsrc")
+\if{html}{\out{<div class="sourceCode">}}\preformatted{lrn("cmprsk.rfsrc")
 }\if{html}{\out{</div>}}
 }
 
 \section{Meta Information}{
 
 \itemize{
-\item Task type: \dQuote{surv}
-\item Predict Types: \dQuote{crank}, \dQuote{distr}
+\item Task type: \dQuote{cmprsk}
+\item Predict Types: \dQuote{cif}
 \item Feature Types: \dQuote{logical}, \dQuote{integer}, \dQuote{numeric}, \dQuote{factor}
-\item Required Packages: \CRANpkg{mlr3}, \CRANpkg{mlr3proba}, \CRANpkg{mlr3extralearners}, \CRANpkg{randomForestSRC}, \CRANpkg{pracma}
+\item Required Packages: \CRANpkg{mlr3}, \CRANpkg{mlr3proba}, \CRANpkg{mlr3extralearners}, \CRANpkg{randomForestSRC}
 }
 }
 
@@ -45,7 +34,7 @@ This \link[mlr3:Learner]{Learner} can be instantiated via \link[mlr3:mlr_sugar]{
    mtry.ratio \tab numeric \tab - \tab  \tab \eqn{[0, 1]}{[0, 1]} \cr
    nodesize \tab integer \tab 15 \tab  \tab \eqn{[1, \infty)}{[1, Inf)} \cr
    nodedepth \tab integer \tab - \tab  \tab \eqn{[1, \infty)}{[1, Inf)} \cr
-   splitrule \tab character \tab logrank \tab logrank, bs.gradient \tab - \cr
+   splitrule \tab character \tab logrankCR \tab logrankCR, logrank \tab - \cr
    nsplit \tab integer \tab 10 \tab  \tab \eqn{[0, \infty)}{[0, Inf)} \cr
    importance \tab character \tab FALSE \tab FALSE, TRUE, none, anti, permute, random \tab - \cr
    block.size \tab integer \tab 10 \tab  \tab \eqn{[1, \infty)}{[1, Inf)} \cr
[…]
