### 2025-07-10 69513e0
check for named numeric vector for selected features and importance scores
