### 2025-07-10 81cbb91
return character for
