### 2025-07-10 22b8feb
fix: when var.used='by.tree' selected_features() was returning empty vector
