+++ tests/testthat/test_randomForestSRC_classif_rfsrc.R
@@ -1,5 +1,4 @@
 test_that("importance/selected", {
-  set.seed(1)
   task = tsk("iris")
   learner = lrn("classif.rfsrc")
   learner$train(task)
@@ -8,14 +7,13 @@ test_that("importance/selected", {
 })
 
 test_that("autotest", {
-  learner = lrn("classif.rfsrc")
-  learner$param_set$values = list(
-    importance = "random", na.action = "na.impute",
-    do.trace = TRUE)
-  expect_learner(learner)
-  set.seed(1)
-  result = run_autotest(learner, exclude = "uf8_feature_names")
-  expect_true(result, info = result$error)
+  with_seed(1, {
+    learner = lrn("classif.rfsrc", ntree = 20, importance = "random", na.action = "na.impute")
+    expect_learner(learner)
+
+    result = run_autotest(learner, exclude = "uf8_feature_names")
+    expect_true(result, info = result$error)
+  })
 })
 
 test_that("convert_ratio", {
