### 2025-07-10 e10ac81
remove by.tree option
