+++ R/learner_randomForestSRC_cmprsk_rfsrc.R
@@ -1,0 +1,214 @@
+#' @title Survival Random Forest Competing Risks Learner
+#' @author bblodfon
+#' @name mlr_learners_cmprsk.rfsrc
+#'
+#' @description
+#' Random survival forests for competing risks.
+#' Calls [randomForestSRC::rfsrc()] from \CRANpkg{randomForestSRC}.
+#'
+#' @template learner
+#' @templateVar id cmprsk.rfsrc
+#'
+#' @inheritSection mlr_learners_classif.rfsrc Custom mlr3 parameters
+#'
+#' @section Initial parameter values:
+#' - `ntime`: Number of time points to coerce the observed event times for use
+#' in the estimated cumulative incidence functions during prediction.
+#' We changed the default value of `150` to `0`, meaning we now use all the
+#' **unique event times from the train set** across all competing causes.
+#'
+#' @references
+#' `r format_bib("ishwaran2014rsf")`
+#'
+#' @template seealso_learner
+#' @examplesIf requireNamespace("randomForestSRC", quietly = TRUE)
+#' # Define the Learner
+#' learner = mlr3::lrn("cmprsk.rfsrc", importance = "TRUE")
+#' print(learner)
+#'
+#' # Define a Task
+#' task = mlr3::tsk("pbc")
+#'
+#' # Stratification based on event
+#' task$set_col_roles(cols = "status", add_to = "stratum")
+#'
+#' # Create train and test set
+#' ids = mlr3::partition(task)
+#'
+#' # Train the learner on the training ids
+#' learner$train(task, row_ids = ids$train)
+#'
+#' print(learner$model)
+#' print(learner$importance(cause = 1)) # VIMP for cause = 1
+#' print(learner$importance(cause = 2)) # VIMP for cause = 2
+#'
+#' # Make predictions for the test rows
+#' predictions = learner$predict(task, row_ids = ids$test)
+#'
+#' # Score the predictions
+#' predictions$score()
+#' @export
+LearnerCompRisksRandomForestSRC = R6Class("LearnerCompRisksRandomForestSRC",
+  inherit = mlr3proba::LearnerCompRisks,
+  public = list(
+    #' @description
+    #' Creates a new instance of this [R6][R6::R6Class] class.
+    initialize = function() {
+      param_set = ps(
+        ntree          = p_int(default = 500L, lower = 1L, tags = "train"),
+        mtry           = p_int(lower = 1L, tags = "train"),
+        mtry.ratio     = p_dbl(lower = 0, upper = 1, tags = "train"),
+        nodesize       = p_int(default = 15L, lower = 1L, tags = "train"),
+        nodedepth      = p_int(lower = 1L, tags = "train"),
+        splitrule      = p_fct(levels = c("logrankCR", "logrank"), default = "logrankCR", tags = "train"),
+        nsplit         = p_int(lower = 0, default = 10, tags = "train"),
+        importance     = p_fct(default = "FALSE", levels = c("FALSE", "TRUE", "none", "anti", "permute", "random"), tags = c("train", "predict")), #nolint
+        block.size     = p_int(default = 10L, lower = 1L, tags = c("train", "predict")),
+        bootstrap      = p_fct(default = "by.root", levels = c("by.root", "by.node", "none", "by.user"), tags = "train"), #nolint
+        samptype       = p_fct(default = "swor", levels = c("swor", "swr"), tags = "train"),
+        samp           = p_uty(tags = "train"),
+        membership     = p_lgl(default = FALSE, tags = c("train", "predict")),
+        sampsize       = p_uty(tags = "train"),
+        sampsize.ratio = p_dbl(0, 1, tags = "train"),
+        na.action      = p_fct(default = "na.omit", levels = c("na.omit", "na.impute"), tags = c("train", "predict")),
[…]
