+++ NAMESPACE
@@ -52,6 +52,7 @@ export(LearnerClassifSMO)
 export(LearnerClassifSimpleLogistic)
 export(LearnerClassifTabPFN)
 export(LearnerClassifVotedPerceptron)
+export(LearnerCompRisksRandomForestSRC)
 export(LearnerDensKDEks)
 export(LearnerDensLocfit)
 export(LearnerDensLogspline)
