+++ man/mlr_learners_classif.rfsrc.Rd
@@ -53,7 +53,7 @@ This \link[mlr3:Learner]{Learner} can be instantiated via \link[mlr3:mlr_sugar]{
    xvar.wt \tab untyped \tab - \tab  \tab - \cr
    split.wt \tab untyped \tab - \tab  \tab - \cr
    forest \tab logical \tab TRUE \tab TRUE, FALSE \tab - \cr
-   var.used \tab character \tab FALSE \tab FALSE, all.trees, by.tree \tab - \cr
+   var.used \tab character \tab FALSE \tab FALSE, all.trees \tab - \cr
    split.depth \tab character \tab FALSE \tab FALSE, all.trees, by.tree \tab - \cr
    seed \tab integer \tab - \tab  \tab \eqn{(-\infty, -1]}{(-Inf, -1]} \cr
    do.trace \tab logical \tab FALSE \tab TRUE, FALSE \tab - \cr
