### 2025-07-10 3c72d10
skip competing risk tests if riskRegression is not installed
