+++ man/mlr_learners_cmprsk.rfsrc.Rd
@@ -95,7 +95,7 @@ set to 1 by default.
 }
 
 \examples{
-\dontshow{if (requireNamespace("randomForestSRC", quietly = TRUE)) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
+\dontshow{if (mlr3misc::require_namespaces(c("randomForestSRC", "riskRegression"), quietly = TRUE)) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
 # Define the Learner
 learner = mlr3::lrn("cmprsk.rfsrc", importance = "TRUE")
 print(learner)
