### 2025-07-10 0ff11ad
fix bug in oob_error
