### 2025-07-10 cef5599
fix NOTEs
