+++ R/learner_tabpfn_regr_tabpfn.R
@@ -92,14 +92,14 @@ LearnerRegrTabPFN = R6Class("LearnerRegrTabPFN",
     #' @description
     #' Marshal the learner's model.
     #' @param ... (any)\cr
-    #'   Additional arguments passed to [`marshal_model()`].
+    #'   Additional arguments passed to [`marshal_model()`][mlr3::marshaling()].
     marshal = function(...) {
       mlr3::learner_marshal(.learner = self, ...)
     },
     #' @description
     #' Unmarshal the learner's model.
     #' @param ... (any)\cr
-    #'   Additional arguments passed to [`unmarshal_model()`].
+    #'   Additional arguments passed to [`unmarshal_model()`][mlr3::marshaling()].
     unmarshal = function(...) {
       mlr3::learner_unmarshal(.learner = self, ...)
     }
