+++ R/learner_xgboost_surv_xgboost_cox.R
@@ -152,14 +152,14 @@ LearnerSurvXgboostCox = R6Class("LearnerSurvXgboostCox",
     #' @description
     #' Marshal the learner's model.
     #' @param ... (any)\cr
-    #'   Additional arguments passed to [`marshal_model()`].
+    #'   Additional arguments passed to [`marshal_model()`][mlr3::marshaling()].
     marshal = function(...) {
       mlr3::learner_marshal(.learner = self, ...)
     },
     #' @description
     #' Unmarshal the learner's model.
     #' @param ... (any)\cr
-    #'   Additional arguments passed to [`unmarshal_model()`].
+    #'   Additional arguments passed to [`unmarshal_model()`][mlr3::marshaling()].
     unmarshal = function(...) {
       mlr3::learner_unmarshal(.learner = self, ...)
     }
