### 2025-07-10 3d028ea
fix stylying notes
