### 2025-07-19 6d946c0
manually install riskRegression for competing risk learners autotest
