### 2025-07-19 5ae60bc
add separate riskRegression install step
