+++ .github/workflows/devcmdcheck.yml
@@ -63,11 +63,14 @@ jobs:
           reticulate::py_install("tensorflow-cpu", pip = TRUE)
           survivalmodels::install_pycox(pip = TRUE, install_torch = FALSE)
           survivalmodels::install_keras(pip = TRUE, install_tensorflow = FALSE)
-          pak::pkg_install("riskRegression")
         shell: Rscript {0}
 
       - name: Install dev versions
         run: pak::pkg_install(c('${{ matrix.config.dev-package }}'))
+        shell: Rscript {0}
+
+      - name: Install riskRegression
+        run: pak::pkg_install("riskRegression")
         shell: Rscript {0}
 
       - uses: r-lib/actions/check-r-package@v2
