+++ tests/testthat/test_tabpfn_classif_tabpfn.R
@@ -118,7 +118,7 @@ test_that("checks for memory saving mode work", {
 
 test_that("random_state works", {
   task = tsk("iris")
-  
+
   # different seeds => slightly different predictions
   learner1 = lrn("classif.tabpfn", predict_type = "prob")
   learner1$param_set$set_values(random_state = "None")
