### 2025-07-20 83dfba1
chore: install tabpfn in workflow
