+++ .github/workflows/devcmdcheck.yml
@@ -63,6 +63,7 @@ jobs:
           reticulate::py_install("tensorflow-cpu", pip = TRUE)
           survivalmodels::install_pycox(pip = TRUE, install_torch = FALSE)
           survivalmodels::install_keras(pip = TRUE, install_tensorflow = FALSE)
+          reticulate::py_install("tabpfn")
         shell: Rscript {0}
 
       - name: Install dev versions
