### 2025-07-20 bc19af3
test: force cpu
