### 2025-07-20 e92137f
fix: revert device change, use reticulate hotfix
See https://github.com/rstudio/reticulate/issues/1811
