### 2025-07-20 2cd458e
test: remove gpu-specific "autocast"
