### 2025-07-21 e8fd6a4
Adding learner from package sparsediscrim (#430)
Adding three new classification learners from the package
[sparsediscrim](https://cran.r-project.org/web/packages/sparsediscrim/index.html),
that are also used by
[tidymodels](https://parsnip.tidymodels.org/reference/details_discrim_linear_sparsediscrim.html)
for linear discriminant analysis:
*  Diagonal Linear Discriminant Analysis classif learner (dlda)
* Minimum Distance Empirical Bayesian classif learner (mdeb)
* Shrinkage-based Diagonal Linear Discriminant Analysis classif learner
(sdlda)


Thank you for contributing a learner to the mlr3 e
[…]
