+++ man/mlr_learners_classif.sdlda.Rd
@@ -1,76 +1,49 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/learner_RWeka_regr_decision_stump.R
-\name{mlr_learners_regr.decision_stump}
-\alias{mlr_learners_regr.decision_stump}
-\alias{LearnerRegrDecisionStump}
-\title{Regression Decision Stump Learner}
+% Please edit documentation in R/learner_sparsediscrim_classif_sdlda.R
+\name{mlr_learners_classif.sdlda}
+\alias{mlr_learners_classif.sdlda}
+\alias{LearnerClassifSdlda}
+\title{Classification Linear Discriminant Analysis Learner}
 \description{
-Decision Stump Learner.
-Calls \code{\link[RWeka:DecisionStump]{RWeka::DecisionStump()}} from \CRANpkg{RWeka}.
+Shrinkage-based Diagonal Linear Discriminant Analysis classfier.
+Type of Naive Bayes classifiers that improves the estimation of the pooled variances by
+using a shrinkage-based estimator of the pooled covariance matrix.
+Calls \code{\link[sparsediscrim:lda_shrink_cov]{sparsediscrim::lda_shrink_cov()}} \CRANpkg{sparsediscrim}.
 }
-\section{Custom mlr3 parameters}{
-
-\itemize{
-\item \code{output_debug_info}:
-\itemize{
-\item original id: output-debug-info
-}
-\item \code{do_not_check_capabilities}:
-\itemize{
-\item original id: do-not-check-capabilities
-}
-\item \code{num_decimal_places}:
-\itemize{
-\item original id: num-decimal-places
-}
-\item \code{batch_size}:
-\itemize{
-\item original id: batch-size
-}
-\item Reason for change: This learner contains changed ids of the following control arguments
-since their ids contain irregular pattern
-}
-}
-
 \section{Dictionary}{
 
 This \link[mlr3:Learner]{Learner} can be instantiated via \link[mlr3:mlr_sugar]{lrn()}:
 
-\if{html}{\out{<div class="sourceCode">}}\preformatted{lrn("regr.decision_stump")
+\if{html}{\out{<div class="sourceCode">}}\preformatted{lrn("classif.sdlda")
 }\if{html}{\out{</div>}}
 }
 
 \section{Meta Information}{
 
 \itemize{
-\item Task type: \dQuote{regr}
-\item Predict Types: \dQuote{response}
-\item Feature Types: \dQuote{logical}, \dQuote{integer}, \dQuote{numeric}, \dQuote{factor}, \dQuote{ordered}
-\item Required Packages: \CRANpkg{mlr3}, \CRANpkg{RWeka}
+\item Task type: \dQuote{classif}
+\item Predict Types: \dQuote{response}, \dQuote{prob}
+\item Feature Types: \dQuote{integer}, \dQuote{numeric}
+\item Required Packages: \CRANpkg{mlr3}, \CRANpkg{sparsediscrim}
 }
 }
 
 \section{Parameters}{
-\tabular{lllll}{
-   Id \tab Type \tab Default \tab Levels \tab Range \cr
-   subset \tab untyped \tab - \tab  \tab - \cr
-   na.action \tab untyped \tab - \tab  \tab - \cr
-   output_debug_info \tab logical \tab FALSE \tab TRUE, FALSE \tab - \cr
-   do_not_check_capabilities \tab logical \tab FALSE \tab TRUE, FALSE \tab - \cr
[…]
