+++ NAMESPACE
@@ -14,6 +14,7 @@ export(LearnerClassifCTree)
 export(LearnerClassifCatboost)
 export(LearnerClassifDecisionStump)
 export(LearnerClassifDecisionTable)
+export(LearnerClassifDiagLda)
 export(LearnerClassifEarth)
 export(LearnerClassifFNN)
 export(LearnerClassifGAMBoost)
@@ -33,6 +34,7 @@ export(LearnerClassifLSSVM)
 export(LearnerClassifLiblineaR)
 export(LearnerClassifLightGBM)
 export(LearnerClassifLogistic)
+export(LearnerClassifMdeb)
 export(LearnerClassifMob)
 export(LearnerClassifMultilayerPerceptron)
 export(LearnerClassifNaiveBayesMultinomial)
@@ -51,6 +53,7 @@ export(LearnerClassifRda)
 export(LearnerClassifSGD)
 export(LearnerClassifSMO)
 export(LearnerClassifSda)
+export(LearnerClassifSdlda)
 export(LearnerClassifSimpleLogistic)
 export(LearnerClassifTabPFN)
 export(LearnerClassifVotedPerceptron)
