### 2025-07-21 8ec6de1
feat: evtree learner from old mlr2 (#433)
Adding a classification and regression learner from package
[evtree](https://cran.r-project.org/web/packages/evtree/index.html).

Both learners have been previously implemented in
[mlr2](https://github.com/mlr-org/mlr)

- [x] The added learner(s) are sufficiently tested
- [x] All the CI tests are passing (including the CodeFactor)
- [x] You ran `devtools::document()`
- [x] You updated the `NEWS.md` field to include the addition of the
learner
- [x] You did not modify anything **not** related to the new learner
- [x] You are listed as a contributor in t
[…]
