+++ NAMESPACE
@@ -16,6 +16,7 @@ export(LearnerClassifDecisionStump)
 export(LearnerClassifDecisionTable)
 export(LearnerClassifDiagLda)
 export(LearnerClassifEarth)
+export(LearnerClassifEvtree)
 export(LearnerClassifFNN)
 export(LearnerClassifGAMBoost)
 export(LearnerClassifGBM)
@@ -74,6 +75,7 @@ export(LearnerRegrCubist)
 export(LearnerRegrDecisionStump)
 export(LearnerRegrDecisionTable)
 export(LearnerRegrEarth)
+export(LearnerRegrEvtree)
 export(LearnerRegrFNN)
 export(LearnerRegrGAMBoost)
 export(LearnerRegrGBM)
