### 2025-07-21 fab1516
feat: Adding ada learner from old mlr2 (#435)
Adding classification learner from package
[ada](https://cran.r-project.org/web/packages/ada/index.html).

This learner has been previously implemented in
[mlr2](https://github.com/mlr-org/mlr).

- [x] The added learner(s) are sufficiently tested
- [x] All the CI tests are passing (including the CodeFactor)
- [x] You ran `devtools::document()`
- [x] You updated the `NEWS.md` field to include the addition of the
learner
- [x] You did not modify anything **not** related to the new learner
- [x] You are listed as a contributor in the `DESCRIPTION` of 
[…]
