+++ DESCRIPTION
@@ -40,6 +40,7 @@ Imports:
     R6
 Suggests:
     abess,
+    ada,
     aorsf (>= 0.1.4),
     actuar,
     apcluster,
