+++ NAMESPACE
@@ -6,6 +6,7 @@ S3method(unmarshal_model,tabpfn_model_marshaled)
 S3method(unmarshal_model,xgboost_cox_model_marshaled)
 export(LearnerClassifAbess)
 export(LearnerClassifAdaBoostM1)
+export(LearnerClassifAdaBoosting)
 export(LearnerClassifBart)
 export(LearnerClassifBayesNet)
 export(LearnerClassifC50)
