### 2025-07-21 e342575
feat: bst regression and classification learners from old mlr2 (#441)
Thank you for contributing a learner to the mlr3 ecosystem.
Please make sure that:

- [x] The added learner(s) are sufficiently tested
- [x] All the CI tests are passing (including the CodeFactor)
- [x] You ran `devtools::document()`
- [x] You updated the `NEWS.md` field to include the addition of the
learner
- [x] You did not modify anything **not** related to the new learner
- [x] You are listed as a contributor in the `DESCRIPTION` of the R
package

Co-authored-by: be-marc <marcbecker@posteo.de>
