+++ NEWS.md
@@ -2,11 +2,9 @@
 
 * fix: remove `discrete` parameter from `surv.parametric`, so that it is impossible to return `distr6::VectorDistribution` survival predictions (softly deprecated in `mlr3proba@v0.8.1`)
 * feat: Added learner `LearnerSurvBlockForest`.
-<<<<<<< HEAD
+* feat: Added learners `LearnerClassifBst` and `LearnerRegrBst`.
 * feat: Added learner `LearnerClassifAdaBoosting`.
-=======
 * feat: Added learners `LearnerClassifEvtree` and `LearnerRegrEvtree`.
->>>>>>> main
 
 # mlr3extralearners 1.1.0
 
@@ -21,6 +19,7 @@ New Features:
 Bugfixes:
 
 * lightgbm classifier now works with encapsulation (#437)
+
 
 # mlr3extralearners 1.0.0
 
