+++ NAMESPACE
@@ -9,6 +9,7 @@ export(LearnerClassifAdaBoostM1)
 export(LearnerClassifAdaBoosting)
 export(LearnerClassifBart)
 export(LearnerClassifBayesNet)
+export(LearnerClassifBst)
 export(LearnerClassifC50)
 export(LearnerClassifCForest)
 export(LearnerClassifCTree)
@@ -69,6 +70,7 @@ export(LearnerDensPlugin)
 export(LearnerDensSpline)
 export(LearnerRegrAbess)
 export(LearnerRegrBart)
+export(LearnerRegrBst)
 export(LearnerRegrCForest)
 export(LearnerRegrCTree)
 export(LearnerRegrCatboost)
