### 2025-07-21 c230147
correct pytorch-cpu installation
