+++ R/bibentries.R
@@ -799,6 +799,16 @@ bibentries = c( # nolint start
     booktitle = "International Conference on Learning Representations 2023",
     year = "2023"
   ),
+  knn2002 = bibentry("book",
+    title = "Modern Applied Statistics with S",
+    author = "W. N. Venables and B. D. Ripley",
+    publisher = "Springer",
+    edition = "Fourth",
+    address = "New York",
+    year = "2002",
+    note = "ISBN 0-387-95457-0",
+    url = "https://www.stats.ox.ac.uk/pub/MASS4/"
+  ),
   hornung2019blockforest = bibentry("article",
     title = "Block Forests: Random forests for blocks of clinical and omics covariate data",
     author = "Hornung, R. and Wright, M. N.",
