+++ man/mlr_learners_classif.knn.Rd
@@ -1,0 +1,138 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/learner_class_classif_knn.R
+\name{mlr_learners_classif.knn}
+\alias{mlr_learners_classif.knn}
+\alias{LearnerClassifKnn}
+\title{Classification K-nearest Neighbour Learner}
+\description{
+K-nearest neighbor classification.
+Calls \code{\link[class:knn]{class::knn()}} from \CRANpkg{class}.
+}
+\section{Dictionary}{
+
+This \link[mlr3:Learner]{Learner} can be instantiated via \link[mlr3:mlr_sugar]{lrn()}:
+
+\if{html}{\out{<div class="sourceCode">}}\preformatted{lrn("classif.knn")
+}\if{html}{\out{</div>}}
+}
+
+\section{Meta Information}{
+
+\itemize{
+\item Task type: \dQuote{classif}
+\item Predict Types: \dQuote{response}
+\item Feature Types: \dQuote{numeric}
+\item Required Packages: \CRANpkg{mlr3}, \CRANpkg{class}
+}
+}
+
+\section{Parameters}{
+\tabular{lllll}{
+   Id \tab Type \tab Default \tab Levels \tab Range \cr
+   k \tab integer \tab 1 \tab  \tab \eqn{[1, \infty)}{[1, Inf)} \cr
+   l \tab numeric \tab 0 \tab  \tab \eqn{[0, \infty)}{[0, Inf)} \cr
+   prob \tab logical \tab FALSE \tab TRUE, FALSE \tab - \cr
+   use.all \tab logical \tab TRUE \tab TRUE, FALSE \tab - \cr
+}
+}
+
+\examples{
+\dontshow{if (mlr3misc::require_namespaces(lrn("classif.knn")$packages, quietly = TRUE)) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
+# Define the Learner
+learner = mlr3::lrn("classif.knn")
+print(learner)
+
+# Define a Task
+task = mlr3::tsk("sonar")
+
+# Create train and test set
+ids = mlr3::partition(task)
+
+# Train the learner on the training ids
+learner$train(task, row_ids = ids$train)
+
+print(learner$model)
+
+
+# Make predictions for the test rows
+predictions = learner$predict(task, row_ids = ids$test)
+
+# Score the predictions
+predictions$score()
+\dontshow{\}) # examplesIf}
+}
+\references{
+Venables WN, Ripley BD (2002).
+\emph{Modern Applied Statistics with S}, Fourth edition.
+Springer, New York.
+ISBN 0-387-95457-0, \url{https://www.stats.ox.ac.uk/pub/MASS4/}.
+}
+\seealso{
+\itemize{
+\item \link[mlr3misc:Dictionary]{Dictionary} of \link[mlr3:Learner]{Learners}: \link[mlr3:mlr_learners]{mlr3::mlr_learners}.
+\item \code{as.data.table(mlr_learners)} for a table of available \link[mlr3:Learner]{Learners} in the running session (depending on the loaded packages).
+\item Chapter in the \href{https://mlr3book.mlr-org.com/}{mlr3book}: \url{https://mlr3book.mlr-org.com/basics.html#learners}
+\item \CRANpkg{mlr3learners} for a selection of recommended learners.
+\item \CRANpkg{mlr3cluster} for unsupervised clustering learners.
+\item \CRANpkg{mlr3pipelines} to combine learners with pre- and postprocessing steps.
+\item \CRANpkg{mlr3tuning} for tuning of hyperparameters, \CRANpkg{mlr3tuningspaces} for established default tuning spaces.
+}
+}
+\author{
+annanzrv
+}
[…]
