+++ DESCRIPTION
@@ -24,7 +24,7 @@ Authors@R: c(
            comment = c(ORCID = "0000-0001-7528-3795")),
     person("Lona", "Koers", , "lona.koers@gmail.com", role = "ctb"),
     person("Baisu", "Zhou", , "baisu.zhou@outlook.com", role = "ctb"),
-    person("Anna", "Nazarova", , "anna.nazarova@outlook.com", role = "ctb")
+    person("Nazarova", "Anna", , "anna.nazarova@outlook.com", role = "ctb")
   )
 Description: Extra learners for use in mlr3.
 License: LGPL-3
@@ -49,6 +49,7 @@ Suggests:
     blockForest (>= 0.2.6),
     C50,
     catboost,
+    class,
     coin,
     CoxBoost,
     Cubist,
