+++ NEWS.md
@@ -2,6 +2,7 @@
 
 * fix: remove `discrete` parameter from `surv.parametric`, so that it is impossible to return `distr6::VectorDistribution` survival predictions (softly deprecated in `mlr3proba@v0.8.1`)
 * feat: Added learner `LearnerSurvBlockForest`.
+* feat: Added learners `LearnerRegrPenalized` and `LearnerClassifPenalized`.
 * feat: Added learners `LearnerClassifBst` and `LearnerRegrBst`.
 * feat: Added learner `LearnerClassifAdaBoosting`.
 * feat: Added learners `LearnerClassifEvtree` and `LearnerRegrEvtree`.
