+++ NAMESPACE
@@ -46,6 +46,7 @@ export(LearnerClassifNaiveBayesWeka)
 export(LearnerClassifObliqueRandomForest)
 export(LearnerClassifOneR)
 export(LearnerClassifPART)
+export(LearnerClassifPenalized)
 export(LearnerClassifPriorityLasso)
 export(LearnerClassifREPTree)
 export(LearnerClassifRandomForest)
@@ -102,6 +103,7 @@ export(LearnerRegrMars)
 export(LearnerRegrMob)
 export(LearnerRegrMultilayerPerceptron)
 export(LearnerRegrObliqueRandomForest)
+export(LearnerRegrPenalized)
 export(LearnerRegrPriorityLasso)
 export(LearnerRegrQGam)
 export(LearnerRegrREPTree)
