### 2025-07-21 d52e5b7
feat: fastai tabular learner (#419)
closes #389 

Added a classification learner and regression learner based on `fastai`,
to use neural networks on tabular data.

---------

Co-authored-by: Sebastian Fischer <sebf.fischer@gmail.com>
Co-authored-by: be-marc <marcbecker@posteo.de>
