+++ attic/scratch_to_be_deleted.R
@@ -1,0 +1,274 @@
+# nolint start
+#reticulate::use_condaenv('r-reticulate',required = TRUE)
+
+
+task = tsk("boston_housing")
+numeric_features = c(task$feature_types[type == "numeric", id], task$target_names)
+df = task$data()
+df = df[, ..numeric_features]
+
+# Step 1: Load data into fastai-compatible format:
+# We  need to divide column names into Categorical columns and Continuous columns
+# We can also define several data preprocessors `procs = list(FillMissing(),Categorify(),Normalize())` and pass them to `TabularDataTable`
+
+task = tsk("sonar")
+df = task$data()
+df_fai = fastai::TabularDataTable(df, #cat_names = cat_cols, cont_names = num_cols,
+                                  y_names = task$target_names, splits=NULL)
+dl = fastai::dataloaders(df_fai)
+# dl$train$y
+#
+# target_factor <- factor(dl$train$y)
+# weights <- c(0.7, 0.3)
+# weight_mapping <- setNames(weights, levels(target_factor))
+# weights_for_factor <- weight_mapping[target_factor]
+# class_weights <- unname(weights_for_factor)
+#
+# dl$train$wgts <- class_weights / sum(class_weights)
+
+# These params are not documented by fastai-r but only by fastai-py:
+# ps_dataloaders = ps(
+#   bs          = p_int(defualt = NULL, tags="train"), # how many samples per batch to load
+#   num_workers = p_int(default = 0L, tags = "train"),  # how many subprocesses to use for data loading
+#   pin_memory  = p_lgl(default = TRUE, tags = "train"),  # If True, the data loader will copy Tensors into CUDA pinned memory before returning them.
+#   shuffle     = p_lgl(default = FALSE, tags = "train"),  # If True, then data is shuffled every time dataloader is fully read/iterated.
+#   drop_last   = p_lgl(default = FALSE, tags = "train")  #  If True, then the last incomplete batch is dropped.
+# )
+
+
+# Step 2: Define a set of configs for the learner
+config = fastai::tabular_config(embed_p = 0.3, use_bn = FALSE)
+# ps_config = ps(
+#   ps       = p_uty(default = NULL, tags = "train"),  # Sequence of dropout probabilities
+#   embed_p  = p_dbl(lower = 0L, upper = 1L, default = 0L, tags = "train"),  # Dropout probability for Embedding layer
+#   y_range  = p_uty(default = NULL, tags = "train"),  # Low and high for SigmoidRange activation (see below)
+#   use_bn   = p_lgl(default = TRUE, tags = "train"),  # Use BatchNorm1d in LinBnDrop layers
+#   bn_final = p_lgl(default = FALSE, tags = "train"),  # Use BatchNorm1d on final layer
+#   bn_cont  = p_lgl(default = TRUE, tags = "train"),  # Use BatchNorm1d on continuous variables
+#   act_cls  = p_uty(tags = "train")  # Activation type for LinBnDrop layers, e.g., fastai::nn()$ReLU(inplace = TRUE)
+# )
+
+# SigmoidRange: torch.sigmoid(x) * (high - low) + low
+#
+# LinBnDrop is a module grouping BatchNorm1d, Dropout and Linear layers:
+#
+# lin_first is not set in fastai-r and defaults to True
+# class LinBnDrop(nn.Sequential):
+#   def __init__(self, n_in, n_out, bn=True, p=0., act=None, lin_first=False):
+#   layers = [BatchNorm(n_out if lin_first else n_in, ndim=1)] if bn else []
[…]
