+++ man/mlr_learners_regr.mqgam.Rd
@@ -55,32 +55,6 @@ This \link[mlr3:Learner]{Learner} can be instantiated via \link[mlr3:mlr_sugar]{
 }
 }
 
-\examples{
-\dontshow{if (requireNamespace("qgam", quietly = TRUE)) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
-# Define the Learner
-learner = mlr3::lrn("regr.mqgam")
-learner$param_set$values$form = mpg ~ cyl + am + s(disp) + s(hp)
-learner$quantiles = c(0.05, 0.5, 0.95)
-learner$quantile_response = 0.5
-print(learner)
-
-# Define a Task
-task = mlr3::tsk("mtcars")
-task$select(c("cyl", "am", "disp", "hp"))
-
-# Create train and test set
-ids = mlr3::partition(task)
-
-# Train the learner on the training ids
-learner$train(task, row_ids = ids$train)
-
-# Make predictions for the test rows
-predictions = learner$predict(task, row_ids = ids$test)
-
-# Score the predictions
-predictions$score()
-\dontshow{\}) # examplesIf}
-}
 \references{
 Fasiolo, Matteo, Wood, N. S, Zaffran, Margaux, Nedellec, Raphael, Goude, Yannig (2017).
 \dQuote{Fast Calibrated Additive Quantile Regression.}
