+++ NAMESPACE
@@ -1,7 +1,9 @@
 # Generated by roxygen2: do not edit by hand
 
+S3method(marshal_model,fastai_model)
 S3method(marshal_model,tabpfn_model)
 S3method(marshal_model,xgboost_cox_model)
+S3method(unmarshal_model,fastai_model_marshaled)
 S3method(unmarshal_model,tabpfn_model_marshaled)
 S3method(unmarshal_model,xgboost_cox_model_marshaled)
 export(LearnerClassifAbess)
@@ -20,6 +22,7 @@ export(LearnerClassifDiagLda)
 export(LearnerClassifEarth)
 export(LearnerClassifEvtree)
 export(LearnerClassifFNN)
+export(LearnerClassifFastai)
 export(LearnerClassifGAMBoost)
 export(LearnerClassifGBM)
 export(LearnerClassifGLMBoost)
@@ -163,6 +166,7 @@ importFrom(methods,formalArgs)
 importFrom(methods,getMethod)
 importFrom(mlr3,LearnerClassif)
 importFrom(mlr3,LearnerRegr)
+importFrom(mlr3,assert_validate)
 importFrom(mlr3,lrn)
 importFrom(mlr3,lrns)
 importFrom(mlr3,marshal_model)

[…]
