### 2025-07-21 6d28cb0
ci: remove tappfn example
