### 2025-07-21 2133c01
remove examples from fastai
