+++ man/mlr3extralearners-package.Rd
@@ -30,9 +30,8 @@ Other contributors:
   \item Zezhi Wang \email{homura@mail.ustc.edu.cn} [contributor]
   \item Lukas Burk \email{burk@leibniz-bips.de} (\href{https://orcid.org/0000-0001-7528-3795}{ORCID}) [contributor]
   \item Lona Koers \email{lona.koers@gmail.com} [contributor]
-  \item Anna Nazarova \email{anna.nazarova@outlook.com} [contributor]
   \item Baisu Zhou \email{baisu.zhou@outlook.com} [contributor]
-  \item Anna Nazarova \email{anna.nazarova@outlook.com} [contributor]
+  \item Nazarova Anna \email{anna.nazarova@outlook.com} [contributor]
 }
 
 }
