### 2025-07-21 5183571
ci: remove tappfn example (#453)
