+++ man/mlr_learners_classif.gamboost.Rd
@@ -1,154 +1,0 @@
-% Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/learner_mboost_classif_gamboost.R
-\name{mlr_learners_classif.gamboost}
-\alias{mlr_learners_classif.gamboost}
-\alias{LearnerClassifGAMBoost}
-\title{Boosted Generalized Additive Classification Learner}
-\description{
-Fit a generalized additive classification model using a boosting algorithm.
-Calls \code{\link[mboost:gamboost]{mboost::gamboost()}}  from \CRANpkg{mboost}.
-}
-\section{Dictionary}{
-
-This \link[mlr3:Learner]{Learner} can be instantiated via \link[mlr3:mlr_sugar]{lrn()}:
-
-\if{html}{\out{<div class="sourceCode">}}\preformatted{lrn("classif.gamboost")
-}\if{html}{\out{</div>}}
-}
-
-\section{Meta Information}{
-
-\itemize{
-\item Task type: \dQuote{classif}
-\item Predict Types: \dQuote{response}, \dQuote{prob}
-\item Feature Types: \dQuote{integer}, \dQuote{numeric}, \dQuote{factor}, \dQuote{ordered}
-\item Required Packages: \CRANpkg{mlr3}, \CRANpkg{mlr3extralearners}, \CRANpkg{mboost}
-}
-}
-
-\section{Parameters}{
-\tabular{lllll}{
-   Id \tab Type \tab Default \tab Levels \tab Range \cr
-   baselearner \tab character \tab bbs \tab bbs, bols, btree \tab - \cr
-   dfbase \tab integer \tab 4 \tab  \tab \eqn{(-\infty, \infty)}{(-Inf, Inf)} \cr
-   family \tab character \tab Binomial \tab Binomial, AdaExp, AUC, custom \tab - \cr
-   custom.family \tab untyped \tab - \tab  \tab - \cr
-   link \tab character \tab logit \tab logit, probit \tab - \cr
-   type \tab character \tab adaboost \tab glm, adaboost \tab - \cr
-   mstop \tab integer \tab 100 \tab  \tab \eqn{(-\infty, \infty)}{(-Inf, Inf)} \cr
-   nu \tab numeric \tab 0.1 \tab  \tab \eqn{(-\infty, \infty)}{(-Inf, Inf)} \cr
-   risk \tab character \tab inbag \tab inbag, oobag, none \tab - \cr
-   oobweights \tab untyped \tab NULL \tab  \tab - \cr
-   trace \tab logical \tab FALSE \tab TRUE, FALSE \tab - \cr
-   stopintern \tab untyped \tab FALSE \tab  \tab - \cr
-   na.action \tab untyped \tab stats::na.omit \tab  \tab - \cr
-}
-}
-
-\section{Offset}{
-
-If a \code{Task} contains a column with the \code{offset} role, it is automatically
-incorporated via the \code{offset} argument in \code{mboost}'s training function.
-No offset is applied during prediction for this learner.
-}
-
-\examples{
-\dontshow{if (mlr3misc::require_namespaces(lrn("classif.gamboost")$packages, quietly = TRUE)) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
-# Define the Learner
-learner = mlr3::lrn("classif.gamboost")
-print(learner)
-
-# Define a Task
-task = mlr3::tsk("sonar")
-
-# Create train and test set
-ids = mlr3::partition(task)
-
-# Train the learner on the training ids
-learner$train(task, row_ids = ids$train)
-
-print(learner$model)
-
-
-# Make predictions for the test rows
-predictions = learner$predict(task, row_ids = ids$test)
-
-# Score the predictions
[…]
