+++ man/mlr_learners_regr.penalized.Rd
@@ -1,160 +1,0 @@
-% Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/learner_penalized_regr_penalized.R
-\name{mlr_learners_regr.penalized}
-\alias{mlr_learners_regr.penalized}
-\alias{LearnerRegrPenalized}
-\title{Regression L1 and L2 Penalized Learner}
-\description{
-Penalized (L1 and L2) Logistic Regression.
-Calls \code{\link[penalized:penalized]{penalized::penalized()}} from: \CRANpkg{penalized}.
-}
-\details{
-The \code{penalized} and \code{unpenalized} arguments in the learner are implemented slightly
-differently than in \code{\link[penalized:penalized]{penalized::penalized()}}. Here, there is no parameter for \code{penalized} but
-instead it is assumed that every variable is penalized unless stated in the \code{unpenalized}
-parameter.
-}
-\section{Initial parameter values}{
-
-\itemize{
-\item \code{trace} is set to \code{"FALSE"} to disable printing output during model training.
-}
-}
-
-\section{Dictionary}{
-
-This \link[mlr3:Learner]{Learner} can be instantiated via \link[mlr3:mlr_sugar]{lrn()}:
-
-\if{html}{\out{<div class="sourceCode">}}\preformatted{lrn("regr.penalized")
-}\if{html}{\out{</div>}}
-}
-
-\section{Meta Information}{
-
-\itemize{
-\item Task type: \dQuote{regr}
-\item Predict Types: \dQuote{response}
-\item Feature Types: \dQuote{logical}, \dQuote{integer}, \dQuote{numeric}, \dQuote{factor}, \dQuote{ordered}
-\item Required Packages: \CRANpkg{mlr3}, \CRANpkg{penalized}
-}
-}
-
-\section{Parameters}{
-\tabular{lllll}{
-   Id \tab Type \tab Default \tab Levels \tab Range \cr
-   epsilon \tab numeric \tab 1e-10 \tab  \tab \eqn{[0, \infty)}{[0, Inf)} \cr
-   fusedl \tab logical \tab FALSE \tab TRUE, FALSE \tab - \cr
-   lambda1 \tab numeric \tab 0 \tab  \tab \eqn{[0, \infty)}{[0, Inf)} \cr
-   lambda2 \tab numeric \tab 0 \tab  \tab \eqn{[0, \infty)}{[0, Inf)} \cr
-   maxiter \tab integer \tab - \tab  \tab \eqn{[1, \infty)}{[1, Inf)} \cr
-   model \tab character \tab linear \tab linear, poisson \tab - \cr
-   positive \tab untyped \tab FALSE \tab  \tab - \cr
-   standardize \tab logical \tab FALSE \tab TRUE, FALSE \tab - \cr
-   startbeta \tab untyped \tab - \tab  \tab - \cr
-   startgamma \tab untyped \tab - \tab  \tab - \cr
-   steps \tab untyped \tab 1L \tab  \tab - \cr
-   trace \tab logical \tab TRUE \tab TRUE, FALSE \tab - \cr
-   unpenalized \tab untyped \tab - \tab  \tab - \cr
-}
-}
-
-\examples{
-\dontshow{if (mlr3misc::require_namespaces(lrn("regr.penalized")$packages, quietly = TRUE)) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
-# Define the Learner
-learner = mlr3::lrn("regr.penalized")
-print(learner)
-
-# Define a Task
-task = mlr3::tsk("mtcars")
-
-# Create train and test set
-ids = mlr3::partition(task)
-
-# Train the learner on the training ids
-learner$train(task, row_ids = ids$train)
-
-print(learner$model)
-
-
[…]

[…]
