+++ .github/workflows/rcmdcheck.yml
@@ -30,7 +30,6 @@ jobs:
       matrix:
         config:
           - {os: ubuntu-latest, r: 'release'}
-          - {os: ubuntu-latest, r: 'devel'}
 
     env:
       GITHUB_PAT: ${{ secrets.GITHUB_TOKEN }}
@@ -53,18 +52,5 @@ jobs:
         if: ${{ github.event_name == 'workflow_dispatch' && inputs.debug_enabled }}
         with:
           limit-access-to-actor: true
-
-      - name: Install Python packages
-        run: |
-          pak::pkg_install('reticulate')
-          reticulate::install_miniconda(force = TRUE)
-          reticulate::use_condaenv('r-reticulate')
-          pak::pkg_install(c('eagerai/fastai', 'keras'))
-          fastai::install_fastai(gpu = FALSE, cuda_version = '11.6', overwrite = FALSE)
-          reticulate::py_install("pytorch-cpu")
-          reticulate::py_install("tensorflow-cpu", pip = TRUE)
-          survivalmodels::install_pycox(pip = TRUE, install_torch = FALSE)
-          survivalmodels::install_keras(pip = TRUE, install_tensorflow = FALSE)
-        shell: Rscript {0}
 
       - uses: r-lib/actions/check-r-package@v2
