+++ attic/DESCRIPTION
@@ -39,10 +39,98 @@ Imports:
     paradox (>= 1.0.1),
     R6
 Suggests:
+    abess,
+    actuar,
+    ada,
+    aorsf (>= 0.1.4),
+    apcluster,
+    BART (>= 2.9.4),
+    blockForest (>= 0.2.6),
+    bst,
+    C50,
+    catboost,
+    class,
+    coin,
+    CoxBoost,
+    Cubist,
+    curl,
+    dbarts,
+    distr6,
+    earth,
+    evtree,
     fastai,
-    testthat
+    flexsurv (>= 2.3),
+    FNN,
+    formattable,
+    future,
+    gbm (>= 2.2.2),
+    glmnet (>= 4.1-6),
+    gss,
+    jsonlite,
+    keras (>= 2.3.0),
+    kernlab,
+    klaR,
+    knitr,
+    ks,
+    LiblineaR,
+    lightgbm (>= 4.5.0),
+    lme4,
+    locfit,
+    logspline,
+    mboost (>= 2.9-10),
+    mda,
+    mgcv,
+    mlr3cluster,
+    mlr3learners (>= 0.12.0),
+    mlr3pipelines,
+    mlr3proba (>= 0.8.1),
+    mvtnorm,
+    nnet,
+    np,
+    param6,
+    partykit (>= 1.2-21),
+    penalized (>= 0.9-52),
+    pendensity,
+    plugdensity,
+    pracma,
+    prioritylasso (>= 0.3.1),
+    pseudo,
+    qgam,
+    randomForest,
+    randomForestSRC (>= 3.3.0),
+    randomPlantedForest,
+    ranger (>= 0.17.0),
+    remotes,
+    reticulate (>= 1.16),
+    rpart,
+    rsm,
+    rvest,
+    RWeka,
+    sandwich,
+    sda,
+    set6,
+    sm,
+    sparsediscrim,
+    stats,
+    survival,
+    survivalmodels (>= 0.1.19),
+    survivalsvm (>= 0.0.5),
+    tensorflow (>= 2.0.0),
+    testthat,
+    xgboost (>= 1.7.8.1)
 VignetteBuilder:
     knitr
+Remotes:
+    binderh/CoxBoost,
+    catboost/catboost/catboost/R-package,
+    mlr-org/mlr3,
+    mlr-org/mlr3proba,
+    PlantedML/randomPlantedForest,
+    RaphaelS1/survivalmodels,
+    ropensci/aorsf,
+    xoopR/distr6,
+    xoopR/param6,
+    xoopR/set6
 Config/Needs/website: rmarkdown
 Config/testthat/edition: 3
 Encoding: UTF-8
