+++ man/mlr_learners_surv.coxtime.Rd
@@ -1,166 +1,0 @@
-% Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/learner_survivalmodels_surv_coxtime.R
-\name{mlr_learners_surv.coxtime}
-\alias{mlr_learners_surv.coxtime}
-\alias{LearnerSurvCoxtime}
-\title{Survival Cox-Time Learner}
-\description{
-Cox-Time survival model.
-Calls \code{\link[survivalmodels:coxtime]{survivalmodels::coxtime()}} from package 'survivalmodels'.
-}
-\section{Prediction types}{
-
-This learner returns two prediction types:
-\enumerate{
-\item \code{distr}: a survival matrix in two dimensions, where observations are
-represented in rows and time points in columns.
-Calculated using the internal \code{\link[survivalmodels:predict.pycox]{survivalmodels::predict.pycox()}} function.
-\item \code{crank}: the expected mortality using \code{\link[survivalmodels:surv_to_risk]{survivalmodels::surv_to_risk()}}.
-}
-}
-
-\section{Dictionary}{
-
-This \link[mlr3:Learner]{Learner} can be instantiated via \link[mlr3:mlr_sugar]{lrn()}:
-
-\if{html}{\out{<div class="sourceCode">}}\preformatted{lrn("surv.coxtime")
-}\if{html}{\out{</div>}}
-}
-
-\section{Meta Information}{
-
-\itemize{
-\item Task type: \dQuote{surv}
-\item Predict Types: \dQuote{crank}, \dQuote{distr}
-\item Feature Types: \dQuote{integer}, \dQuote{numeric}
-\item Required Packages: \CRANpkg{mlr3}, \CRANpkg{mlr3proba}, \CRANpkg{mlr3extralearners}, \CRANpkg{survivalmodels}, \CRANpkg{distr6}, \CRANpkg{reticulate}
-}
-}
-
-\section{Parameters}{
-\tabular{lllll}{
-   Id \tab Type \tab Default \tab Levels \tab Range \cr
-   frac \tab numeric \tab 0 \tab  \tab \eqn{[0, 1]}{[0, 1]} \cr
-   standardize_time \tab logical \tab FALSE \tab TRUE, FALSE \tab - \cr
-   log_duration \tab logical \tab FALSE \tab TRUE, FALSE \tab - \cr
-   with_mean \tab logical \tab TRUE \tab TRUE, FALSE \tab - \cr
-   with_std \tab logical \tab TRUE \tab TRUE, FALSE \tab - \cr
-   num_nodes \tab untyped \tab c(32L, 32L) \tab  \tab - \cr
-   batch_norm \tab logical \tab TRUE \tab TRUE, FALSE \tab - \cr
-   dropout \tab numeric \tab - \tab  \tab \eqn{[0, 1]}{[0, 1]} \cr
-   activation \tab character \tab relu \tab celu, elu, gelu, glu, hardshrink, hardsigmoid, hardswish, hardtanh, relu6, leakyrelu, \link{...} \tab - \cr
-   device \tab untyped \tab - \tab  \tab - \cr
-   shrink \tab numeric \tab 0 \tab  \tab \eqn{[0, \infty)}{[0, Inf)} \cr
-   optimizer \tab character \tab adam \tab adadelta, adagrad, adam, adamax, adamw, asgd, rmsprop, rprop, sgd, sparse_adam \tab - \cr
-   rho \tab numeric \tab 0.9 \tab  \tab \eqn{(-\infty, \infty)}{(-Inf, Inf)} \cr
-   eps \tab numeric \tab 1e-08 \tab  \tab \eqn{(-\infty, \infty)}{(-Inf, Inf)} \cr
[…]
