+++ man/mlr_learners_regr.randomForest.Rd
@@ -1,179 +1,0 @@
-% Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/learner_randomForest_regr_randomForest.R
-\name{mlr_learners_regr.randomForest}
-\alias{mlr_learners_regr.randomForest}
-\alias{LearnerRegrRandomForest}
-\title{Regression Random Forest Learner}
-\description{
-Random forest for regression.
-Calls \code{\link[randomForest:randomForest]{randomForest::randomForest()}} from \CRANpkg{randomForest}.
-}
-\section{Dictionary}{
-
-This \link[mlr3:Learner]{Learner} can be instantiated via \link[mlr3:mlr_sugar]{lrn()}:
-
-\if{html}{\out{<div class="sourceCode">}}\preformatted{lrn("regr.randomForest")
-}\if{html}{\out{</div>}}
-}
-
-\section{Meta Information}{
-
-\itemize{
-\item Task type: \dQuote{regr}
-\item Predict Types: \dQuote{response}
-\item Feature Types: \dQuote{logical}, \dQuote{integer}, \dQuote{numeric}, \dQuote{factor}, \dQuote{ordered}
-\item Required Packages: \CRANpkg{mlr3}, \CRANpkg{mlr3extralearners}, \CRANpkg{randomForest}
-}
-}
-
-\section{Parameters}{
-\tabular{lllll}{
-   Id \tab Type \tab Default \tab Levels \tab Range \cr
-   ntree \tab integer \tab 500 \tab  \tab \eqn{[1, \infty)}{[1, Inf)} \cr
-   mtry \tab integer \tab - \tab  \tab \eqn{[1, \infty)}{[1, Inf)} \cr
-   replace \tab logical \tab TRUE \tab TRUE, FALSE \tab - \cr
-   strata \tab untyped \tab - \tab  \tab - \cr
-   sampsize \tab untyped \tab - \tab  \tab - \cr
-   nodesize \tab integer \tab 5 \tab  \tab \eqn{[1, \infty)}{[1, Inf)} \cr
-   maxnodes \tab integer \tab - \tab  \tab \eqn{[1, \infty)}{[1, Inf)} \cr
-   importance \tab character \tab FALSE \tab mse, nudepurity, none \tab - \cr
-   localImp \tab logical \tab FALSE \tab TRUE, FALSE \tab - \cr
-   proximity \tab logical \tab FALSE \tab TRUE, FALSE \tab - \cr
-   oob.prox \tab logical \tab - \tab TRUE, FALSE \tab - \cr
-   norm.votes \tab logical \tab TRUE \tab TRUE, FALSE \tab - \cr
-   do.trace \tab logical \tab FALSE \tab TRUE, FALSE \tab - \cr
-   keep.forest \tab logical \tab TRUE \tab TRUE, FALSE \tab - \cr
-   keep.inbag \tab logical \tab FALSE \tab TRUE, FALSE \tab - \cr
-   predict.all \tab logical \tab FALSE \tab TRUE, FALSE \tab - \cr
-   nodes \tab logical \tab FALSE \tab TRUE, FALSE \tab - \cr
-}
-}
-
-\examples{
-\dontshow{if (requireNamespace("randomForest", quietly = TRUE)) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
-# Define the Learner
-learner = mlr3::lrn("regr.randomForest", importance = "mse")
-print(learner)
-
-# Define a Task
-task = mlr3::tsk("mtcars")
-# Create train and test set
-ids = mlr3::partition(task)
-
-# Train the learner on the training ids
-learner$train(task, row_ids = ids$train)
-
-print(learner$model)
-print(learner$importance())
-
[…]

[…]
