### 2025-07-21 1b1cc59
verify python installation
