+++ .github/workflows/devcmdcheck.yml
@@ -72,7 +72,7 @@ jobs:
         shell: Rscript {0}
         run: |
           pak::pkg_install('reticulate')
-          # reticulate::use_condaenv('r-reticulate', required = TRUE)
+          reticulate::use_condaenv('r-reticulate', required = TRUE)
 
           fastai::install_fastai(gpu = FALSE, cuda_version = '11.6', overwrite = FALSE)
           survivalmodels::install_pycox(pip = TRUE, install_torch = FALSE)
@@ -93,7 +93,6 @@ jobs:
       - name: Set RETICULATE_PYTHON for tests
         shell: bash
         run: |
-          # Ensure reticulate picks up the correct Python
           PYTHON_PATH=$(conda run -n r-reticulate which python)
           echo "RETICULATE_PYTHON=$PYTHON_PATH" >> $GITHUB_ENV
 
