+++ DESCRIPTION
@@ -40,7 +40,10 @@ Imports:
     R6
 Suggests:
     fastai,
+    mlr3proba,
     testthat
+Remotes:
+    mlr-org/mlr3proba
 Config/Needs/website: rmarkdown
 Config/testthat/edition: 3
 Encoding: UTF-8
