+++ .github/workflows/rcmdcheck.yml
@@ -83,10 +83,12 @@ jobs:
       - name: Verify Python dependencies
         shell: bash
         run: |
+           set -e
            python --version
            pip list
            conda --version
-           conda run -n r-reticulate python -c "import fastai, torch, ..."
+           conda env list
+           conda run -n r-reticulate which python
 
       - name: Set RETICULATE_PYTHON for tests
         shell: bash
