+++ .github/workflows/rcmdcheck.yml
@@ -43,6 +43,15 @@ jobs:
         with:
           r-version: ${{ matrix.config.r }}
 
+      - name: Setup Miniconda
+        uses: conda-incubator/setup-miniconda@v2
+        with:
+          python-version: '3.10'
+          auto-update-conda: true
+          use-mamba: true
+          activate-environment: r-reticulate
+          add-pip-as-python-dependency: true
+
       - uses: r-lib/actions/setup-r-dependencies@v2
         with:
           extra-packages: any::rcmdcheck
