+++ R/learner_fastai_classif_fastai.R
@@ -151,8 +151,6 @@ LearnerClassifFastai = R6Class("LearnerClassifFastai",
       measure = pars$eval_metric
       patience = pars$patience
 
-      browser()
-
       if (length(cat_cols) && is.null(pars$procs)) pars$procs = list(fastai::Categorify)
       if (is.null(measure)) measure = fastai::accuracy
 
