### 2025-07-21 766351b
add test that verifies reticulate env
