+++ tests/testthat/test_fastai_classif_fastai.R
@@ -1,7 +1,9 @@
 # reticulate::use_condaenv("r-reticulate")
 
 test_that("Python config is valid", {
-  print(reticulate::py_config())
+  cnfg = reticulate::py_config()
+  env = basename(cnfg$pythonhome)
+  expect_equal(env, "r-reticulate")
 })
 
 test_that("autotest", {
