### 2025-07-21 7d0a13f
test: skip if py module unavailable
