### 2025-07-21 0b147d4
test: try init the conda env
