### 2025-07-21 b0f2f66
try init conda env in fastai tests
