+++ tests/testthat/test_fastai_classif_fastai.R
@@ -1,4 +1,9 @@
 # reticulate::use_condaenv("r-reticulate")
+tryCatch({
+  reticulate::use_condaenv("r-reticulate", required = TRUE)
+}, error = function(e) {
+  skip("r-reticulate conda environment not found.")
+})
 
 test_that("autotest", {
   learner = lrn("classif.fastai", layers = c(200, 100))
