### 2025-07-21 f61b5ea
quick fix in fastai
