### 2025-07-22 e714b64
test: move py module availability checks into test functions
