+++ tests/testthat/test_tabpfn_classif_tabpfn.R
@@ -1,4 +1,14 @@
+test_that("Python config is valid", {
+  skip_if_no_tabpfn()
+
+  cnfg = reticulate::py_config()
+  env = basename(cnfg$pythonhome)
+  expect_equal(env, "r-reticulate")
+})
+
 test_that("autotest", {
+  skip_if_no_tabpfn()
+
   learner = lrn("classif.tabpfn")
   expect_learner(learner)
   # reproducibility is not guaranteed, hence check_replicable = FALSE
@@ -7,6 +17,8 @@ test_that("autotest", {
 })
 
 test_that("marshaling works for classif.tabpfn", {
+  skip_if_no_tabpfn()
+
   learner = lrn("classif.tabpfn")
   task = tsk("iris")
   # expect_marshalable_learner(learner, task)
@@ -33,6 +45,8 @@ test_that("marshaling works for classif.tabpfn", {
 })
 
 test_that("categorical feature columns are encoded correctly", {
+  skip_if_no_tabpfn()
+
   n = 6
   task = as_task_classif(
     data.frame(
@@ -51,9 +65,8 @@ test_that("categorical feature columns are encoded correctly", {
 })
 
 test_that("device selection works", {
-  if (!reticulate::py_module_available("torch")) {
-    skip("torch not available for testing.")
-  }
+  skip_if_no_tabpfn()
+
   torch = reticulate::import("torch")
 
   learner = lrn("classif.tabpfn", device = "cpu")
@@ -66,7 +79,33 @@ test_that("device selection works", {
   expect_identical(learner$model$fitted$device, "auto")
 })
 
+test_that("inference_precision works", {
+  skip_if_no_tabpfn()
+
+  torch = reticulate::import("torch")
+  task = tsk("iris")
+
+  # No test for the "autocast" option, because it is not supported on cpu.
+  # But we test all possible torch dtypes.
+  dtypes = c(
+    "float32", "float",
+    "float64", "double",
+    "float16", "half",
+    "bfloat16"
+  )
+  lapply(dtypes, function(dtype) {
+    learner = lrn("classif.tabpfn", inference_precision = paste0("torch.", dtype))
+    expect_invisible(learner$train(task))
+    actual_dtype = learner$model$fitted$inference_precision
+    expected_dtype = reticulate::py_get_attr(torch, dtype)
+    # actual and expected should be the same torch.dtype object
+    expect_identical(reticulate::py_id(actual_dtype), reticulate::py_id(expected_dtype))
+  })
+})
+
 test_that("checks for memory saving mode work", {
+  skip_if_no_tabpfn()
+
   learner = lrn("classif.tabpfn")
   task = tsk("iris")
 
@@ -84,4 +123,35 @@ test_that("checks for memory saving mode work", {
   learner$param_set$set_values(memory_saving_mode = 50)
   learner$train(task)
   expect_identical(learner$model$fitted$memory_saving_mode, 50)
+})
+
+test_that("random_state works", {
+  skip_if_no_tabpfn()
+
+  task = tsk("iris")
+
+  # different seeds => slightly different predictions
+  learner1 = lrn("classif.tabpfn", predict_type = "prob")
+  learner1$param_set$set_values(random_state = "None")
+  learner1$train(task)
+  learner2 = lrn("classif.tabpfn", predict_type = "prob")
+  learner2$param_set$set_values(random_state = "None")
+  learner2$train(task)
+  expect_character(all.equal(learner1$predict(task), learner2$predict(task)))
+
+  # same seed (default: 0) => (almost) same predictions
+  learner1 = lrn("classif.tabpfn", predict_type = "prob")
+  learner1$train(task)
[…]
