### 2025-07-22 a4ac499
chore: no init conda env in workflows
