### 2025-07-22 5c81cf2
try only py_require
