+++ .github/workflows/rcmdcheck.yml
@@ -57,13 +57,8 @@ jobs:
       - name: Install Python packages
         run: |
           pak::pkg_install('reticulate')
-          reticulate::install_miniconda(force = TRUE)
-          pak::pkg_install(c('eagerai/fastai', 'keras'))
-          fastai::install_fastai(gpu = FALSE, cuda_version = '11.6', overwrite = FALSE)
-          reticulate::py_install("pytorch-cpu")
-          reticulate::py_install("tensorflow-cpu", pip = TRUE)
-          survivalmodels::install_pycox(pip = TRUE, install_torch = FALSE)
-          survivalmodels::install_keras(pip = TRUE, install_tensorflow = FALSE)
+          reticulate::py_require(c("torch", "torchvision", "torchaudio", "fastai", "tabpfn", "pycox", "tensorflow-cpu", "keras"))
+          reticulate::py_config()
         shell: Rscript {0}
 
       - uses: r-lib/actions/check-r-package@v2
