### 2025-07-22 90bd940
try install torch cpu version manually
