### 2025-07-22 0e56bca
try force imports right after installation
