+++ .github/workflows/rcmdcheck.yml
@@ -59,6 +59,10 @@ jobs:
           pak::pkg_install('reticulate')
           reticulate::py_install(c("torch", "torchvision", "torchaudio"), pip = TRUE, pip_options = "--index-url https://download.pytorch.org/whl/cpu")
           reticulate::py_require(c("torch", "torchvision", "torchaudio", "fastai", "tabpfn", "pycox", "tensorflow-cpu", "keras"))
+          reticulate::import("torch")
+          reticulate::import("fastai")
+          reticulate::import("tabpfn")
+          reticulate::import("pycox")
           reticulate::py_config()
         shell: Rscript {0}
 
