### 2025-07-22 7ba50b8
try list packages
