### 2025-07-22 f5f9359
add py_require to test helper file
