+++ tests/testthat/test_fastai_classif_fastai.R
@@ -1,4 +1,4 @@
-print(reticulate::py_list_packages()[, c("name", "version")])
+print(reticulate::py_list_packages()[, c("package", "version")])
 
 test_that("Python config is valid", {
   skip_if_no_fastai()
