+++ tests/testthat/test_tabpfn_classif_tabpfn.R
@@ -1,4 +1,4 @@
-print(reticulate::py_list_packages())[, c("name", "version")]
+print(reticulate::py_list_packages())[, c("package", "version")]
 
 test_that("Python config is valid", {
   skip_if_no_tabpfn()
