+++ tests/testthat/test_survivalmodels_surv_coxtime.R
@@ -1,5 +1,4 @@
 skip_on_os("windows")
-print(reticulate::py_list_packages())[, c("package", "version")]
 
 test_that("autotest", {
   skip_if_no_pycox()
