+++ man/mlr_learners_classif.fastai.Rd
@@ -70,31 +70,6 @@ This \link[mlr3:Learner]{Learner} can be instantiated via \link[mlr3:mlr_sugar]{
 }
 }
 
-\examples{
-\dontshow{if (mlr3misc::require_namespaces(lrn("classif.fastai")$packages, quietly = TRUE)) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
-# Define the Learner
-learner = mlr3::lrn("classif.fastai")
-print(learner)
-
-# Define a Task
-task = mlr3::tsk("sonar")
-
-# Create train and test set
-ids = mlr3::partition(task)
-
-# Train the learner on the training ids
-learner$train(task, row_ids = ids$train)
-
-print(learner$model)
-
-
-# Make predictions for the test rows
-predictions = learner$predict(task, row_ids = ids$test)
-
-# Score the predictions
-predictions$score()
-\dontshow{\}) # examplesIf}
-}
 \references{
 Howard, Jeremy, Gugger, Sylvain (2020).
 \dQuote{Fastai: A Layered API for Deep Learning.}
