+++ .github/workflows/rcmdcheck.yml
@@ -43,27 +43,11 @@ jobs:
         with:
           r-version: ${{ matrix.config.r }}
 
-      - name: Setup Miniconda
-        uses: conda-incubator/setup-miniconda@v2
-        with:
-          python-version: '3.10'
-          auto-update-conda: true
-          use-mamba: true
-          activate-environment: r-reticulate
-          add-pip-as-python-dependency: true
-
-
-      - name: Install & configure reticulate-based packages
-        shell: Rscript {0}
-        run: |
-          install.packages('reticulate')
-          reticulate::use_condaenv('r-reticulate', required = TRUE)
-          fastai::install_fastai(gpu = FALSE, cuda_version = '11.6', overwrite = FALSE)
-
       - uses: r-lib/actions/setup-r-dependencies@v2
         with:
           extra-packages: any::rcmdcheck
           needs: check
+          cache: "always"
 
       - uses: mxschmitt/action-tmate@v3
         if: ${{ github.event_name == 'workflow_dispatch' && inputs.debug_enabled }}
