+++ NAMESPACE
@@ -1,8 +1,11 @@
 # Generated by roxygen2: do not edit by hand
 
 S3method(marshal_model,fastai_model)
+S3method(marshal_model,tabpfn_model)
 S3method(unmarshal_model,fastai_model_marshaled)
+S3method(unmarshal_model,tabpfn_model_marshaled)
 export(LearnerClassifFastai)
+export(LearnerClassifTabPFN)
 export(create_learner)
 export(lrn)
 export(lrns)
