+++ NAMESPACE
@@ -6,6 +6,7 @@ S3method(unmarshal_model,fastai_model_marshaled)
 S3method(unmarshal_model,tabpfn_model_marshaled)
 export(LearnerClassifFastai)
 export(LearnerClassifTabPFN)
+export(LearnerSurvCoxtime)
 export(create_learner)
 export(lrn)
 export(lrns)
