+++ R/learner_survivalmodels_surv_dnnsurv.R
@@ -137,6 +137,7 @@ LearnerSurvDNNSurv = R6Class("LearnerSurvDNNSurv",
     },
 
     .predict = function(task) {
+      reticulate::py_require(c("keras", "tensorflow"))
 
       pars = self$param_set$get_values(tags = "predict")
       newdata = ordered_features(task, self)
