+++ DESCRIPTION
@@ -44,6 +44,7 @@ Suggests:
     fastai,
     mlr3cluster,
     mlr3proba,
+    keras (>= 2.3.0),
     remotes,
     survival,
     survivalmodels,
