+++ man/mlr_learners_surv.dnnsurv.Rd
@@ -1,0 +1,186 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/learner_survivalmodels_surv_dnnsurv.R
+\name{mlr_learners_surv.dnnsurv}
+\alias{mlr_learners_surv.dnnsurv}
+\alias{LearnerSurvDNNSurv}
+\title{Survival DNNSurv Learner}
+\description{
+Fits a neural network based on pseudo-conditional survival probabilities.
+Calls \code{\link[survivalmodels:dnnsurv]{survivalmodels::dnnsurv()}} from package 'survivalmodels'.
+}
+\details{
+Custom nets can be used in this learner either using the
+\link[survivalmodels:build_keras_net]{survivalmodels::build_keras_net} utility function or using \CRANpkg{keras}.
+The number of output channels should be of length \code{1} and number of input channels is
+the number of features plus number of cuts.
+}
+\section{Prediction types}{
+
+This learner returns two prediction types:
+\enumerate{
+\item \code{distr}: a survival matrix in two dimensions, where observations are
+represented in rows and time points in columns.
+Calculated using the internal \code{\link[survivalmodels:predict.dnnsurv]{survivalmodels::predict.dnnsurv()}} function.
+\item \code{crank}: the expected mortality using \code{\link[survivalmodels:surv_to_risk]{survivalmodels::surv_to_risk()}}.
+}
+}
+
+\section{Dictionary}{
+
+This \link[mlr3:Learner]{Learner} can be instantiated via \link[mlr3:mlr_sugar]{lrn()}:
+
+\if{html}{\out{<div class="sourceCode">}}\preformatted{lrn("surv.dnnsurv")
+}\if{html}{\out{</div>}}
+}
+
+\section{Meta Information}{
+
+\itemize{
+\item Task type: \dQuote{surv}
+\item Predict Types: \dQuote{crank}, \dQuote{distr}
+\item Feature Types: \dQuote{integer}, \dQuote{numeric}
+\item Required Packages: \CRANpkg{mlr3}, \CRANpkg{mlr3proba}, \CRANpkg{mlr3extralearners}, \CRANpkg{survivalmodels}, \CRANpkg{keras}, \CRANpkg{pseudo}, \CRANpkg{tensorflow}, \CRANpkg{distr6}
+}
+}
+
+\section{Parameters}{
+\tabular{lllll}{
+   Id \tab Type \tab Default \tab Levels \tab Range \cr
+   cuts \tab integer \tab 5 \tab  \tab \eqn{[1, \infty)}{[1, Inf)} \cr
+   cutpoints \tab untyped \tab - \tab  \tab - \cr
+   custom_model \tab untyped \tab - \tab  \tab - \cr
+   optimizer \tab character \tab adam \tab adadelta, adagrad, adamax, adam, nadam, rmsprop, sgd \tab - \cr
+   lr \tab numeric \tab 0.02 \tab  \tab \eqn{[0, \infty)}{[0, Inf)} \cr
+   beta_1 \tab numeric \tab 0.9 \tab  \tab \eqn{[0, 1]}{[0, 1]} \cr
+   beta_2 \tab numeric \tab 0.999 \tab  \tab \eqn{[0, 1]}{[0, 1]} \cr
+   epsilon \tab numeric \tab - \tab  \tab \eqn{[0, \infty)}{[0, Inf)} \cr
+   decay \tab numeric \tab 0 \tab  \tab \eqn{[0, \infty)}{[0, Inf)} \cr
[…]
