+++ NAMESPACE
@@ -7,6 +7,7 @@ S3method(unmarshal_model,tabpfn_model_marshaled)
 export(LearnerClassifFastai)
 export(LearnerClassifTabPFN)
 export(LearnerSurvCoxtime)
+export(LearnerSurvDNNSurv)
 export(create_learner)
 export(lrn)
 export(lrns)
