+++ DESCRIPTION
@@ -45,6 +45,7 @@ Suggests:
     mlr3cluster,
     mlr3proba,
     keras (>= 2.3.0),
+    pseudo,
     remotes,
     survival,
     survivalmodels,
