+++ R/learner_survivalmodels_surv_pchazard.R
@@ -1,10 +1,11 @@
-#' @title Survival Logistic-Hazard Learner
+#' @title Survival PC-Hazard Learner
 #' @author RaphaelS1
-#' @name mlr_learners_surv.loghaz
+#' @name mlr_learners_surv.pchazard
 #'
 #' @description
-#' Survival logistic hazard learner.
-#' Calls [survivalmodels::loghaz()] from package 'survivalmodels'.
+#' Logistic-Hazard fits a discrete neural network based on a cross-entropy loss and
+#' predictions of a discrete hazard function, also known as Nnet-Survival.
+#' Calls [survivalmodels::pchazard()] from package 'survivalmodels'.
 #'
 #' @section Prediction types:
 #' This learner returns two prediction types:
@@ -13,8 +14,8 @@
 #' Calculated using the internal [survivalmodels::predict.pycox()] function.
 #' 2. `crank`: the expected mortality using [survivalmodels::surv_to_risk()].
 #'
-#' @templateVar id surv.loghaz
 #' @template learner
+#' @templateVar id surv.pchazard
 #'
 #' @template install_survivalmodels
 #'
@@ -25,12 +26,12 @@
 #'  the parameters `cuts` or `cutpoints`.
 #'
 #' @references
-#' `r format_bib("gensheimer2018simple", "kvamme2019time")`
+#' `r format_bib("kvamme2019continuous")`
 #'
 #' @template seealso_learner
 #' @template simple_example
 #' @export
-LearnerSurvLogisticHazard = R6Class("LearnerSurvLogisticHazard",
+LearnerSurvPCHazard = R6Class("LearnerSurvPCHazard",
   inherit = mlr3proba::LearnerSurv,
 
   public = list(
@@ -38,7 +39,7 @@ LearnerSurvLogisticHazard = R6Class("LearnerSurvLogisticHazard",
     #' Creates a new instance of this [R6][R6::R6Class] class.
     initialize = function() {
 
-      param_set = ps(
+      ps = ps(
         frac = p_dbl(default = 0, lower = 0, upper = 1, tags = "train"),
         cuts = p_int(default = 10L, lower = 1L, tags = "train"),
         cutpoints = p_uty(tags = "train"),
@@ -47,6 +48,8 @@ LearnerSurvLogisticHazard = R6Class("LearnerSurvLogisticHazard",
         cut_min = p_dbl(default = 0, lower = 0, tags = "train"),
         num_nodes = p_uty(default = c(32L, 32L), tags = "train"),
         batch_norm = p_lgl(default = TRUE, tags = "train"),
+        reduction = p_fct(default = "mean", levels = c("mean", "none", "sum"),
+          tags = "train"),
         dropout = p_dbl(lower = 0, upper = 1, tags = "train"),
         activation = p_fct(default = "relu",
           levels = c("celu", "elu", "gelu", "glu", "hardshrink", "hardsigmoid", "hardswish",
@@ -86,57 +89,55 @@ LearnerSurvLogisticHazard = R6Class("LearnerSurvLogisticHazard",
         min_delta = p_dbl(default = 0, tags = "train"),
         patience = p_int(default = 10, tags = "train"),
         interpolate = p_lgl(default = FALSE, tags = "predict"),
-        inter_scheme = p_fct(default = "const_hazard",
-          levels = c("const_hazard", "const_pdf"), tags = "predict"),
         sub = p_int(default = 10L, lower = 1L, tags = "predict")
       )
 
-      param_set$add_dep("rho", "optimizer", CondEqual$new("adadelta"))
[…]
