+++ R/learner_survivalmodels_surv_deephit.R
@@ -1,10 +1,10 @@
-#' @title Survival DeepSurv Learner
+#' @title Survival DeepHit Learner
 #' @author RaphaelS1
-#' @name mlr_learners_surv.deepsurv
+#' @name mlr_learners_surv.deephit
 #'
 #' @description
-#' DeepSurv fits a neural network based on the partial likelihood from a Cox PH.
-#' Calls [survivalmodels::deepsurv()] from package 'survivalmodels'.
+#' Neural network 'Deephit' for survival analysis.
+#' Calls [survivalmodels::deephit()] from pacakge 'survivalmodels'.
 #'
 #' @section Prediction types:
 #' This learner returns two prediction types:
@@ -14,17 +14,23 @@
 #' 2. `crank`: the expected mortality using [survivalmodels::surv_to_risk()].
 #'
 #' @template learner
-#' @templateVar id surv.deepsurv
+#' @templateVar id surv.deephit
 #'
 #' @template install_survivalmodels
 #'
+#' @details
+#' Custom nets can be used in this learner either using the
+#'  [survivalmodels::build_pytorch_net] utility function or using `torch` via \CRANpkg{reticulate}.
+#'  The number of output channels depends on the number of discretised time-points, i.e.
+#'  the parameters `cuts` or `cutpoints`.
+#'
 #' @references
-#' `r format_bib("katzman2018deepsurv")`
+#' `r format_bib("lee2018deephit")`
 #'
 #' @template seealso_learner
 #' @template simple_example
 #' @export
-LearnerSurvDeepsurv = R6Class("LearnerSurvDeepsurv",
+LearnerSurvDeephit = R6Class("LearnerSurvDeephit",
   inherit = mlr3proba::LearnerSurv,
 
   public = list(
@@ -34,6 +40,11 @@ LearnerSurvDeepsurv = R6Class("LearnerSurvDeepsurv",
 
       ps = ps(
         frac = p_dbl(default = 0, lower = 0, upper = 1, tags = "train"),
+        cuts = p_int(default = 10L, lower = 1L, tags = "train"),
+        cutpoints = p_uty(tags = "train"),
+        scheme = p_fct(default = "equidistant", levels = c("equidistant", "quantiles"),
+          tags = "train"),
+        cut_min = p_dbl(default = 0, lower = 0, tags = "train"),
         num_nodes = p_uty(default = c(32L, 32L), tags = "train"),
         batch_norm = p_lgl(default = TRUE, tags = "train"),
         dropout = p_dbl(lower = 0, upper = 1, tags = "train"),
@@ -43,7 +54,10 @@ LearnerSurvDeepsurv = R6Class("LearnerSurvDeepsurv",
             "rrelu", "relu", "selu", "sigmoid", "softmax", "softmax2d", "softmin",
             "softplus", "softshrink", "softsign", "tanh", "tanhshrink", "threshold"),
           tags = "train"),
+        custom_net = p_uty(tags = "train"),
         device = p_uty(tags = "train"),
+        mod_alpha = p_dbl(default = 0.2, lower = 0, upper = 1, tags = "train"),
+        sigma = p_dbl(default = 0.1, lower = 0, tags = "train"),
         optimizer = p_fct(default = "adam",
           levels = c("adadelta", "adagrad", "adam", "adamax", "adamw", "asgd", "rmsprop", "rprop",
             "sgd", "sparse_adam"), tags = "train"),
@@ -72,7 +86,11 @@ LearnerSurvDeepsurv = R6Class("LearnerSurvDeepsurv",
         best_weights = p_lgl(default = FALSE, tags = "train"),
         early_stopping = p_lgl(default = FALSE, tags = "train"),
[…]
