+++ NAMESPACE
@@ -6,8 +6,14 @@ S3method(unmarshal_model,fastai_model_marshaled)
 S3method(unmarshal_model,tabpfn_model_marshaled)
 export(LearnerClassifFastai)
 export(LearnerClassifTabPFN)
+export(LearnerSurvAkritas)
 export(LearnerSurvCoxtime)
 export(LearnerSurvDNNSurv)
+export(LearnerSurvDeephit)
+export(LearnerSurvDeepsurv)
+export(LearnerSurvLogisticHazard)
+export(LearnerSurvPCHazard)
+export(LearnerSurvParametric)
 export(create_learner)
 export(lrn)
 export(lrns)

[…]
