+++ man/mlr_learners_regr.tabpfn.Rd
@@ -1,0 +1,213 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/learner_tabpfn_regr_tabpfn.R
+\name{mlr_learners_regr.tabpfn}
+\alias{mlr_learners_regr.tabpfn}
+\alias{LearnerRegrTabPFN}
+\title{TabPFN Regression Learner}
+\description{
+Foundation model for tabular data.
+Uses \CRANpkg{reticulate} to interface with the \href{https://github.com/PriorLabs/TabPFN}{\code{tabpfn}} Python package.
+}
+\section{Installation}{
+
+While the Python dependencies are handled via \code{reticulate::py_require()}, you can
+manually specify a virtual environment by calling \code{reticulate::use_virtualenv()}
+prior to calling the \verb{$train()} function.
+In this virtual environment, the \code{tabpfn} package and its dependencies must be installed.
+}
+
+\section{Saving a Learner}{
+
+In order to save a \code{LearnerRegrTabPFN} for later usage,
+it is necessary to call the \verb{$marshal()} method on the \code{Learner}
+before writing it to disk, as the object will otherwise not be saved correctly.
+After loading a marshaled \code{LearnerRegrTabPFN} into R again,
+you then need to call \verb{$unmarshal()} to transform it into a useable state.
+}
+
+\section{Initial parameter values}{
+
+\itemize{
+\item \code{n_jobs} is initialized to 1 to avoid threading conflicts with \CRANpkg{future}.
+}
+}
+
+\section{Custom mlr3 parameters}{
+
+\itemize{
+\item \code{output_type} corresponds to the same argument of the \code{.predict()} method of the \code{TabPFNRegressor} class,
+but only supports the options \code{"mean"}, \code{"median"} and \code{"mode"}.
+The point predictions are stored as \verb{$response} of the prediction object.
+\item \code{categorical_feature_indices} uses R indexing instead of zero-based Python indexing.
+\item \code{device} must be a string.
+If set to \code{"auto"}, the behavior is the same as original.
+Otherwise, the string is passed as argument to \code{torch.device()} to create a device.
+\item \code{inference_precision} must be \code{"auto"} or \code{"autocast"}.
+Passing \code{torch.dtype} is currently not supported.
+\item \code{inference_config} is currently not supported.
+}
+}
+
+\section{Dictionary}{
+
+This \link[mlr3:Learner]{Learner} can be instantiated via \link[mlr3:mlr_sugar]{lrn()}:
+
+\if{html}{\out{<div class="sourceCode">}}\preformatted{lrn("regr.tabpfn")
+}\if{html}{\out{</div>}}
+}
+
+\section{Meta Information}{
+
+\itemize{
+\item Task type: \dQuote{regr}
+\item Predict Types: \dQuote{response}, \dQuote{quantiles}
+\item Feature Types: \dQuote{logical}, \dQuote{integer}, \dQuote{numeric}
+\item Required Packages: \CRANpkg{mlr3}
+}
+}
+
+\section{Parameters}{
+\tabular{lllll}{
+   Id \tab Type \tab Default \tab Levels \tab Range \cr
+   output_type \tab character \tab mean \tab mean, median, mode \tab - \cr
[…]
