+++ .github/workflows/rcmdcheck.yml
@@ -56,4 +56,4 @@ jobs:
 
       - uses: r-lib/actions/check-r-package@v2
         with:
-          error-on: "error"
+          error-on: '"error"'
