+++ .codefactor.yml
@@ -1,0 +1,2 @@
+exclude:
+  - attic/*
