+++ man/mlr_learners_regr.qgam.Rd
@@ -1,0 +1,133 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/learner_qgam_regr_qgam.R
+\name{mlr_learners_regr.qgam}
+\alias{mlr_learners_regr.qgam}
+\alias{LearnerRegrQGam}
+\title{Generalized Additive Quantile Regression Model}
+\description{
+Quantile Regression with generalized additive models.
+Calls \code{\link[qgam:qgam]{qgam::qgam()}} from package \CRANpkg{qgam}.
+}
+\section{Form}{
+
+For the \code{form} parameter, a gam formula specific to the \link[mlr3:Task]{Task} is required (see example and \code{?mgcv::formula.gam}).
+If no formula is provided, a fallback formula using all features in the task is used that will make the Learner behave like Linear Quantile Regression.
+The features specified in the formula need to be the same as columns with col_roles "feature" in the task.
+}
+
+\section{Quantile}{
+
+The quantile for the Learner, i.e. \code{qu} parameter from \code{\link[qgam:qgam]{qgam::qgam()}}, is set using the value specified in \code{learner$quantiles}.
+}
+
+\section{Dictionary}{
+
+This \link[mlr3:Learner]{Learner} can be instantiated via \link[mlr3:mlr_sugar]{lrn()}:
+
+\if{html}{\out{<div class="sourceCode">}}\preformatted{lrn("regr.qgam")
+}\if{html}{\out{</div>}}
+}
+
+\section{Meta Information}{
+
+\itemize{
+\item Task type: \dQuote{regr}
+\item Predict Types: \dQuote{response}, \dQuote{se}, \dQuote{quantiles}
+\item Feature Types: \dQuote{logical}, \dQuote{integer}, \dQuote{numeric}, \dQuote{factor}
+\item Required Packages: \CRANpkg{mlr3}, \CRANpkg{qgam}
+}
+}
+
+\section{Parameters}{
+\tabular{lllll}{
+   Id \tab Type \tab Default \tab Levels \tab Range \cr
+   form \tab untyped \tab - \tab  \tab - \cr
+   lsig \tab numeric \tab - \tab  \tab \eqn{(-\infty, \infty)}{(-Inf, Inf)} \cr
+   err \tab numeric \tab - \tab  \tab \eqn{[0, 1]}{[0, 1]} \cr
+   cluster \tab untyped \tab NULL \tab  \tab - \cr
+   multicore \tab logical \tab - \tab TRUE, FALSE \tab - \cr
+   ncores \tab numeric \tab - \tab  \tab \eqn{(-\infty, \infty)}{(-Inf, Inf)} \cr
+   paropts \tab untyped \tab list() \tab  \tab - \cr
+   link \tab untyped \tab "identity" \tab  \tab - \cr
+   argGam \tab untyped \tab - \tab  \tab - \cr
+   discrete \tab logical \tab FALSE \tab TRUE, FALSE \tab - \cr
+   block.size \tab integer \tab 1000 \tab  \tab \eqn{(-\infty, \infty)}{(-Inf, Inf)} \cr
+   unconditional \tab logical \tab FALSE \tab TRUE, FALSE \tab - \cr
+}
+}
+
+\references{
+Fasiolo, Matteo, Wood, N. S, Zaffran, Margaux, Nedellec, Raphael, Goude, Yannig (2017).
+\dQuote{Fast Calibrated Additive Quantile Regression.}
+\emph{Journal of the American Statistical Association}, \bold{116}, 1402--1412.
+\doi{10.1080/01621459.2020.1725521}.
+}
+\seealso{
+\itemize{
[…]
