+++ NAMESPACE
@@ -7,6 +7,7 @@ S3method(unmarshal_model,tabpfn_model_marshaled)
 export(LearnerClassifFastai)
 export(LearnerClassifTabPFN)
 export(LearnerRegrMQGam)
+export(LearnerRegrQGam)
 export(LearnerRegrTabPFN)
 export(LearnerSurvAkritas)
 export(LearnerSurvCoxtime)
