+++ .codefactor.yml
@@ -1,2 +1,0 @@
-exclude:
-  - attic/*
