+++ DESCRIPTION
@@ -45,6 +45,7 @@ Suggests:
     mlr3cluster,
     mlr3proba,
     keras (>= 2.3.0),
+    pracma,
     pseudo,
     qgam,
     remotes,
@@ -59,6 +60,7 @@ Remotes:
     rstudio/reticulate
 Config/Needs/website: rmarkdown
 Config/testthat/edition: 3
+Config/testthat/parallel: false
 Encoding: UTF-8
 NeedsCompilation: no
 Roxygen: list(markdown = TRUE, r6 = TRUE)
