+++ attic/R/learner_fastai_classif_fastai.R
@@ -1,356 +1,0 @@
-#' @title Classification Neural Network Learner
-#' @author annanzrv
-#' @name mlr_learners_classif.fastai
-#'
-#' @description
-#' Simple and fast neural nets for tabular data classification.
-#' Calls [fastai::tabular_learner()] from package \CRANpkg{fastai}.
-#'
-#' @section Initial parameter values:
-#' - `n_epoch`:
-#'   Needs to be set for [fastai::fit()] to work.
-#'   If no value is given, it is set to 5.
-#' - `eval_metric`:
-#'   Needs to be set for [fastai::predict()] to work.
-#'   If no value is given, it is set to `fastai::accuracy()`.
-#'
-#' @templateVar id classif.fastai
-#' @template learner
-#'
-#' @references
-#' `r format_bib("howard_2020")`
-#'
-#' @template seealso_learner
-#' @template example
-#' @export
-LearnerClassifFastai = R6Class("LearnerClassifFastai",
-  inherit = LearnerClassif,
-  public = list(
-
-    #' @description
-    #' Creates a new instance of this [R6][R6::R6Class] class.
-    initialize = function() {
-
-      p_n_epoch = p_int(1L,
-        tags = c("train", "hotstart", "internal_tuning"),
-        init = 5L,
-        aggr = crate(function(x) as.integer(ceiling(mean(unlist(x)))), .parent = topenv()),
-        in_tune_fn = crate(function(domain, param_vals) {
-          if (is.null(param_vals$patience)) {
-            stop("Parameter 'patience' must be set to use internal tuning.")
-          }
-          assert_integerish(domain$upper, len = 1L, any.missing = FALSE)
-        }, .parent = topenv()),
-        disable_in_tune = list(n_epoch = NULL)
-      )
-
-      param_set = ps(
-        act_cls     = p_uty(tags = "train"),  # Activation type for LinBnDrop layers, e.g., fastai::nn()$ReLU(inplace = TRUE)
-        bn_cont     = p_lgl(default = TRUE, tags = "train"),  # Use BatchNorm1d on continuous variables
-        bn_final    = p_lgl(default = FALSE, tags = "train"),  # Use BatchNorm1d on final layer
-        drop_last   = p_lgl(default = FALSE, tags = "train"),  #  If True, then the last incomplete batch is dropped.
-        embed_p     = p_dbl(lower = 0L, upper = 1L, default = 0L, tags = "train"),  # Dropout probability for Embedding layer
-        emb_szs     = p_uty(default = NULL, tags = "train"),  # Sequence of (num_embeddings, embedding_dim) for each categorical variable
-        n_epoch     = p_n_epoch,
-        eval_metric = p_uty(tags = "train", custom_check = crate({
-          function(x) check_true(is.function(x) || inherits(x, "Measure"))
-        })),
-        layers      = p_uty(tags = "train"),  # Sequence of ints used to specify the input and output size of each LinBnDrop layer
-        loss_func   = p_uty(tags = "train"),  # Defaults to fastai::CrossEntropyLossFlat()
-        lr          = p_dbl(lower = 0, default = 0.001, tags = "train"),  # Learning rate
-        metrics     = p_uty(tags = "train"),  # optional list of metrics, e.g, fastai::Precision() or fastai::accuracy()
-        n_out       = p_int(tags = "train"),  # ?
-        num_workers = p_int(default = 0L, tags = "train"),  # how many subprocesses to use for data loading
-        opt_func    = p_uty(tags = "train"),  # Optimizer created when Learner.fit is called. E.g., fastai::Adam()
-        patience    = p_int(1L, default = 1, tags = "train"),  # number of epochs to wait when training has not improved model. add `depends = quote(early_stopping == TRUE`)`
[…]
