+++ attic/R/learner_survivalmodels_surv_dnnsurv.R
@@ -1,157 +1,0 @@
-#' @title Survival DNNSurv Learner
-#' @author RaphaelS1
-#' @name mlr_learners_surv.dnnsurv
-#'
-#' @description
-#' Fits a neural network based on pseudo-conditional survival probabilities.
-#' Calls [survivalmodels::dnnsurv()] from package 'survivalmodels'.
-#'
-#' @section Prediction types:
-#' This learner returns two prediction types:
-#' 1. `distr`: a survival matrix in two dimensions, where observations are
-#' represented in rows and time points in columns.
-#' Calculated using the internal [survivalmodels::predict.dnnsurv()] function.
-#' 2. `crank`: the expected mortality using [survivalmodels::surv_to_risk()].
-#'
-#' @template learner
-#' @templateVar id surv.dnnsurv
-#'
-#' @template install_survivalmodels
-#'
-#' @details
-#' Custom nets can be used in this learner either using the
-#'  [survivalmodels::build_keras_net] utility function or using \CRANpkg{keras}.
-#'  The number of output channels should be of length `1` and number of input channels is
-#'  the number of features plus number of cuts.
-#'
-#' @section Initial parameter values:
-#' - `verbose` is initialized to 0.
-#'
-#' @references
-#' `r format_bib("zhao2019dnnsurv")`
-#'
-#' @template seealso_learner
-#' @template simple_example
-#' @export
-LearnerSurvDNNSurv = R6Class("LearnerSurvDNNSurv",
-  inherit = mlr3proba::LearnerSurv,
-
-  public = list(
-    #' @description
-    #' Creates a new instance of this [R6][R6::R6Class] class.
-    initialize = function() {
-
-      ps = ps(
-        cuts = p_int(default = 5, lower = 1, tags = "train"),
-        cutpoints = p_uty(tags = "train"),
-        custom_model = p_uty(tags = "train"),
-        optimizer = p_fct(default = "adam",
-          levels = c("adadelta", "adagrad", "adamax", "adam", "nadam", "rmsprop", "sgd"),
-          tags = "train"),
-        lr = p_dbl(default = 0.02, lower = 0, tags = "train"),
-        beta_1 = p_dbl(default = 0.9, lower = 0, upper = 1, tags = "train"),
-        beta_2 = p_dbl(default = 0.999, lower = 0, upper = 1, tags = "train"),
-        epsilon = p_dbl(lower = 0, tags = "train"),
-        decay = p_dbl(default = 0, lower = 0, tags = "train"),
-        clipnorm = p_dbl(tags = "train"),
-        clipvalue = p_dbl(tags = "train"),
-        momentum = p_dbl(default = 0, lower = 0, tags = "train"),
-        nesterov = p_lgl(default = FALSE, tags = "train"),
-        loss_weights = p_uty(tags = "train"),
-        weighted_metrics = p_uty(tags = "train"),
-        early_stopping = p_lgl(default = FALSE, tags = "train"),
-        min_delta = p_dbl(default = 0, lower = 0, tags = "train"),
-        patience = p_int(default = 0L, lower = 0, tags = "train"),
-        verbose = p_int(default = 0L, lower = 0, upper = 2, tags = c("train", "predict")),
-        baseline = p_dbl(tags = "train"),
-        restore_best_weights = p_lgl(default = FALSE, tags = "train"),
-        batch_size = p_int(default = 32L, lower = 1, tags = c("train", "predict")),
-        epochs = p_int(default = 10L, lower = 1, tags = "train"),
-        validation_split = p_dbl(default = 0, lower = 0, upper = 1, tags = "train"),
[…]

[…]
