### 2025-07-22 d9875cc
print config
