+++ tests/testthat/test_tabpfn_classif_tabpfn.R
@@ -1,3 +1,6 @@
+print(reticulate::py_config())
+print(reticulate::py_list_packages()[, c("package", "version")])
+
 test_that("autotest", {
   learner = lrn("classif.tabpfn")
   expect_learner(learner)
