### 2025-07-22 b6d8f46
try omit python installation in workflow
