+++ R/learner_dbarts_classif_bart.R
@@ -94,13 +94,15 @@ LearnerClassifBart = R6Class("LearnerClassifBart",
 
       pars$weights = private$.get_weights(task)
 
-      invoke(
+      capture.output({model = invoke(
         dbarts::bart,
         x.train = x_train,
         y.train = y_train,
         keeptrees = TRUE,
         .args = pars
-      )
+      )})
+
+      model
     },
 
     .predict = function(task) {
