+++ tests/testthat/test_kernlab_classif_gausspr.R
@@ -1,7 +1,7 @@
 test_that("autotest", {
   learner = lrn("classif.gausspr")
   expect_learner(learner)
-  result = run_autotest(learner)
+  capture.output({result = run_autotest(learner)})
   expect_true(result, info = result$error)
 })
 
