+++ tests/testthat/test_kernlab_classif_lssvm.R
@@ -2,7 +2,7 @@ test_that("autotest", {
   set.seed(1)
   learner = lrn("classif.lssvm")
   expect_learner(learner)
-  result = run_autotest(learner, exclude = "single|utf8_feature_names_multiclass", N = 50)
+  capture.output({result = run_autotest(learner, exclude = "single|utf8_feature_names_multiclass", N = 50)})
   expect_true(result, info = result$error)
 })
 
