+++ tests/testthat/test_kernlab_classif_ksvm.R
@@ -1,6 +1,6 @@
 test_that("autotest", {
   learner = lrn("classif.ksvm")
   expect_learner(learner)
-  result = run_autotest(learner)
+  capture.output({result = run_autotest(learner)})
   expect_true(result, info = result$error)
 })
