+++ R/learner_dbarts_regr_bart.R
@@ -99,12 +99,10 @@ LearnerRegrBart = R6Class("LearnerRegrBart",
       pars$weights = private$.get_weights(task)
 
       # y.train should either be a binary factor or have values {0, 1}
-      capture.output({model = invoke(dbarts::bart,
+      invoke(dbarts::bart,
         x.train = data, y.train = outcome,
         .args = pars
-      )})
-
-      model
+      )
     },
 
     .predict = function(task) {
