+++ tests/testthat/test_dbarts_classif_bart.R
@@ -7,6 +7,6 @@ test_that("autotest", {
   )
 
   expect_learner(learner)
-  result = run_autotest(learner, exclude = "sanity|utf8_feature_names")
+  capture.output({result = run_autotest(learner, exclude = "sanity|utf8_feature_names")})
   expect_true(result, info = result$error)
 })
