+++ tests/testthat/test_pendensity_dens_pen.R
@@ -1,6 +1,6 @@
 test_that("autotest", {
   learner = lrn("dens.pen")
   expect_learner(learner)
-  result = run_autotest(learner)
+  capture.output({result = run_autotest(learner)})
   expect_true(result, info = result$error)
 })
