+++ tests/testthat/test_survivalmodels_surv_parametric.R
@@ -53,28 +53,36 @@ test_that("missing", {
 
 test_that("'form' affects survival prediction", {
   # aft
-  p = lrn("surv.parametric", form = "aft")$train(task)$predict(task)
+  learner = lrn("surv.parametric", form = "aft")
+  learner$train(task)
+  p = learner$predict(task)
   expect_true(inherits(p$distr, "Matdist"))
   surv_aft = p$distr$survival(task$unique_times())
   expect_matrix(surv_aft, nrows = length(task$unique_times()), ncols = task$nrow,
                 any.missing = FALSE) # [times x obs]
 
   # tobit
-  p = lrn("surv.parametric", form = "tobit")$train(task)$predict(task)
+  learner = lrn("surv.parametric", form = "tobit")
+  learner$train(task)
+  p = learner$predict(task)
   expect_true(inherits(p$distr, "Matdist"))
   surv_tobit = p$distr$survival(task$unique_times())
   expect_matrix(surv_tobit, nrows = length(task$unique_times()), ncols = task$nrow,
                 any.missing = FALSE) # [times x obs]
 
   # ph
-  p = lrn("surv.parametric", form = "ph")$train(task)$predict(task)
+  learner = lrn("surv.parametric", form = "ph")
+  learner$train(task)
+  p = learner$predict(task)
   expect_true(inherits(p$distr, "Matdist"))
   surv_ph = p$distr$survival(task$unique_times())
   expect_matrix(surv_ph, nrows = length(task$unique_times()), ncols = task$nrow,
                 any.missing = FALSE) # [times x obs]
 
   # po
-  p = lrn("surv.parametric", form = "po")$train(task)$predict(task)
+  learner = lrn("surv.parametric", form = "po")
+  learner$train(task)
+  p = learner$predict(task)
   expect_true(inherits(p$distr, "Matdist"))
   surv_po = p$distr$survival(task$unique_times())
   expect_matrix(surv_po, nrows = length(task$unique_times()), ncols = task$nrow,
