+++ tests/testthat/test_randomForestSRC_classif_rfsrc.R
@@ -2,7 +2,7 @@ test_that("importance/selected", {
   set.seed(1)
   task = tsk("iris")
   learner = lrn("classif.rfsrc")
-  learner$train(task)
+  capture.output({learner$train(task)})
   expect_error(learner$importance(), "Set 'importance'")
   expect_error(learner$selected_features(), "Set 'var.used'")
 })
@@ -14,24 +14,24 @@ test_that("autotest", {
     do.trace = TRUE)
   expect_learner(learner)
   set.seed(1)
-  result = run_autotest(learner, exclude = "uf8_feature_names")
+  capture.output({result = run_autotest(learner, exclude = "uf8_feature_names")})
   expect_true(result, info = result$error)
 })
 
 test_that("convert_ratio", {
   task = tsk("sonar")
   learner = lrn("classif.rfsrc", ntree = 5, mtry.ratio = .5)
-  expect_equal(learner$train(task)$model$mtry, 30)
+  capture.output({expect_equal(learner$train(task)$model$mtry, 30)})
 
   learner$param_set$values$mtry.ratio = 0
-  expect_equal(learner$train(task)$model$mtry, 1)
+  capture.output({expect_equal(learner$train(task)$model$mtry, 1)})
 
   learner$param_set$values$mtry.ratio = 1
-  expect_equal(learner$train(task)$model$mtry, 60)
+  capture.output({expect_equal(learner$train(task)$model$mtry, 60)})
 
   learner$param_set$values$mtry = 10
   expect_error(learner$train(task), "exclusive")
 
   learner$param_set$values$mtry.ratio = NULL
-  expect_equal(learner$train(task)$model$mtry, 10)
+  capture.output({expect_equal(learner$train(task)$model$mtry, 10)})
 })
