+++ tests/testthat/test_survivalmodels_surv_akritas.R
@@ -10,18 +10,21 @@ test_that("autotest", {
 test_that("time points for prediction", {
   task = tsk("lung")
   learner = lrn("surv.akritas")
-  p = learner$train(task)$predict(task)
+  learner$train(task)
+  p = learner$predict(task)
   times = as.integer(colnames(p$data$distr))
   expect_equal(times, task$unique_times()) # unique train time points are used
 
   # use many more time points than the ones in the train set
   learner$param_set$set_values(.values = list(ntime = 9999))
-  p = learner$train(task)$predict(task)
+  learner$train(task)
+  p = learner$predict(task)
   times = as.integer(colnames(p$data$distr))
   expect_equal(times, task$unique_times()) # all unique train time points are still used
 
   learner$param_set$set_values(.values = list(ntime = 50))
-  p = learner$train(task)$predict(task)
+  learner$train(task)
+  p = learner$predict(task)
   times = as.integer(colnames(p$data$distr))
   # min and max times are retained after coersion
   expect_equal(min(times), min(task$unique_times()))
