+++ tests/testthat/test_randomForestSRC_regr_rfsrc.R
@@ -2,7 +2,7 @@ test_that("importance/selected", {
   set.seed(1)
   task = tsk("boston_housing")
   learner = lrn("regr.rfsrc")
-  learner$train(task)
+  capture.output({learner$train(task)})
   expect_error(learner$importance(), "Set 'importance'")
   expect_error(learner$selected_features(), "Set 'var.used'")
 })
@@ -14,6 +14,6 @@ test_that("autotest", {
     do.trace = TRUE)
   expect_learner(learner)
   set.seed(1)
-  result = run_autotest(learner, exclude = "single")
+  capture.output({result = run_autotest(learner, exclude = "single")})
   expect_true(result, info = result$error)
 })
