+++ tests/testthat/test_dbarts_regr_bart.R
@@ -6,6 +6,8 @@ test_that("autotest", {
     ntree = 5L)
 
   expect_learner(learner)
-  capture.output({result = run_autotest(learner, exclude = "utf8_feature_names")})
+  capture.output({
+    result = run_autotest(learner, exclude = "utf8_feature_names")
+  })
   expect_true(result, info = result$error)
 })
