+++ tests/testthat/test_kernlab_regr_gausspr.R
@@ -1,7 +1,9 @@
 test_that("autotest", {
   learner = lrn("regr.gausspr")
   expect_learner(learner)
-  capture.output({result = run_autotest(learner)})
+  capture.output({
+    result = run_autotest(learner)
+  })
   expect_true(result, info = result$error)
 })
 
