+++ tests/testthat/test_kernlab_regr_rvm.R
@@ -11,10 +11,14 @@ test_that("autotest", {
 test_that("regr.rvm sigma", {
   learner = lrn("regr.rvm", kpar = list(sigma = 0.2))
   t = tsk("mtcars")
-  capture.output({learner$train(t)})
+  capture.output({
+    learner$train(t)
+  })
   expect_equal(learner$model@kernelf@kpar$sigma, 0.2)
 
   learner = lrn("regr.rvm", sigma = 0.2)
-  capture.output({learner$train(t)})
+  capture.output({
+    learner$train(t)
+  })
   expect_equal(learner$model@kernelf@kpar$sigma, 0.2)
 })
