### 2025-07-22 608ac06
ci: fix python deps (#455)
