+++ R/learner_survivalmodels_surv_loghaz.R
@@ -133,7 +133,7 @@ LearnerSurvLogisticHazard = R6Class("LearnerSurvLogisticHazard",
 
   private = list(
     .train = function(task) {
-
+      reticulate::py_require(c("torch", "pycox", "numpy"))
       pars = self$param_set$get_values(tags = "train")
       invoke(
         survivalmodels::loghaz,
@@ -146,7 +146,7 @@ LearnerSurvLogisticHazard = R6Class("LearnerSurvLogisticHazard",
     },
 
     .predict = function(task) {
-
+      reticulate::py_require(c("torch", "pycox", "numpy"))
       pars = self$param_set$get_values(tags = "predict")
       newdata = ordered_features(task, self)
 
