+++ tests/testthat/test_survivalmodels_surv_coxtime.R
@@ -1,10 +1,5 @@
 skip_on_os("windows")
 
-if (!reticulate::py_module_available("torch") || !reticulate::py_module_available("pycox") ||
-  !reticulate::py_module_available("numpy")) {
-  skip("One of torch, numpy, pycox not available for testing.")
-}
-
 np = reticulate::import("numpy")
 torch = reticulate::import("torch")
 set.seed(1)
