+++ tests/testthat/test_sda_classif_sda.R
@@ -2,6 +2,8 @@ test_that("autotest", {
   learner = lrn("classif.sda")
   expect_learner(learner)
   # note that you can skip tests using the exclude argument
-  result = run_autotest(learner)
+  capture.output({
+    result = run_autotest(learner)
+  })
   expect_true(result, info = result$error)
 })
