### 2025-07-22 33d0120
ci: activate dev check
