+++ attic/DESCRIPTION
@@ -1,139 +1,0 @@
-Package: mlr3extralearners
-Title: Extra Learners For mlr3
-Version: 1.1.0.9000
-Authors@R: c(
-    person("Raphael", "Sonabend", , "raphaelsonabend@gmail.com", role = "aut",
-           comment = c(ORCID = "0000-0001-9225-4654")),
-    person("Patrick", "Schratz", , "patrick.schratz@gmail.com", role = "aut",
-           comment = c(ORCID = "0000-0003-0748-6624")),
-    person("Lorenz A.", "Kapsner", , "lorenz.kapsner@gmail.com", role = "ctb",
-           comment = c(ORCID = "0000-0003-1866-860X")),
-    person("Lennart", "Schneider", , "lennart.sch@web.de", role = "ctb",
-           comment = c(ORCID = "0000-0003-4152-5308")),
-    person("Stephen A", "Lauer", , "stephenalauer@gmail.com", role = "ctb",
-           comment = c(ORCID = "0000-0003-2948-630X")),
-    person("Pierre", "Camilleri", , "camilleri_pierre@hotmail.fr", role = "ctb"),
-    person("Javier", "García", , "geshter@hotmail.com", role = "ctb"),
-    person("Keenan", "Ganz", , "ganzkeenan1@gmail.com", role = "ctb"),
-    person("Byron", "Jaeger", , "bjaeger@wakehealth.edu", role = "ctb"),
-    person("Sebastian", "Fischer", , "sebf.fischer@gmail.com", role = c("cre", "aut")),
-    person("Zezhi", "Wang", , "homura@mail.ustc.edu.cn", role = "ctb"),
-    person("John", "Zobolas", , "bblodfon@gmail.com", role = "aut",
-           comment = c(ORCID = "0000-0002-3609-8674")),
-    person("Lukas", "Burk", , "burk@leibniz-bips.de", role = "ctb",
-           comment = c(ORCID = "0000-0001-7528-3795")),
-    person("Lona", "Koers", , "lona.koers@gmail.com", role = "ctb"),
-    person("Anna", "Nazarova", , "anna.nazarova@outlook.com", role = "ctb"),
-    person("Baisu", "Zhou", , "baisu.zhou@outlook.com", role = "ctb")
-  )
-Description: Extra learners for use in mlr3.
-License: LGPL-3
-Depends:
-    mlr3 (>= 1.0.0),
-    R (>= 3.3.0)
-Imports:
-    checkmate,
-    data.table,
-    methods,
-    mlr3misc (>= 0.9.4),
-    paradox (>= 1.0.1),
-    R6
-Suggests:
-    abess,
-    actuar,
-    ada,
-    aorsf (>= 0.1.4),
-    apcluster,
-    BART (>= 2.9.4),
-    blockForest (>= 0.2.6),
-    bst,
-    C50,
-    catboost,
-    class,
-    coin,
-    CoxBoost,
-    Cubist,
-    curl,
-    dbarts,
-    distr6,
-    earth,
-    evtree,
-    fastai,
-    flexsurv (>= 2.3),
-    FNN,
-    formattable,
-    future,
-    gbm (>= 2.2.2),
-    glmnet (>= 4.1-6),
-    gss,
-    jsonlite,
-    keras (>= 2.3.0),
-    kernlab,
-    klaR,
-    knitr,
-    ks,
-    LiblineaR,
-    lightgbm (>= 4.5.0),
-    lme4,
-    locfit,
-    logspline,
-    mboost (>= 2.9-10),
-    mda,
-    mgcv,
-    mlr3cluster,
[…]
