### 2025-07-23 83a74fa
chore: add vignette and pkgdown
