+++ DESCRIPTION
@@ -117,6 +117,8 @@ Suggests:
     tensorflow (>= 2.0.0),
     testthat,
     xgboost (>= 1.7.8.1)
+VignetteBuilder:
+    knitr
 Remotes:
     binderh/CoxBoost,
     catboost/catboost/catboost/R-package,
