+++ DESCRIPTION
@@ -24,7 +24,9 @@ Authors@R: c(
            comment = c(ORCID = "0000-0001-7528-3795")),
     person("Lona", "Koers", , "lona.koers@gmail.com", role = "ctb"),
     person("Anna", "Nazarova", , "anna.nazarova@outlook.com", role = "ctb"),
-    person("Baisu", "Zhou", , "baisu.zhou@outlook.com", role = "ctb")
+    person("Baisu", "Zhou", , "baisu.zhou@outlook.com", role = "ctb"),
+    person("Marc", "Becker", , "marcbecker@posteo.de", role = "ctb",
+           comment = c(ORCID = "0000-0002-8115-0400"))
   )
 Description: Extra learners for use in mlr3.
 License: LGPL-3
