### 2025-07-25 081a78a
feat: assert_python_packages
