+++ R/learner_survivalmodels_surv_parametric.R
@@ -110,7 +110,9 @@ LearnerSurvParametric = R6Class("LearnerSurvParametric",
 
   private = list(
     .train = function(task) {
-      reticulate::py_require(c("torch", "pycox", "numpy"))
+      python_packages = c("torch", "pycox", "numpy")
+      reticulate::py_require(python_packages)
+      assert_python_packages(python_packages)
 
       pv = self$param_set$get_values(tags = "train")
       pv$weights = private$.get_weights(task)
@@ -125,7 +127,9 @@ LearnerSurvParametric = R6Class("LearnerSurvParametric",
     },
 
     .predict = function(task) {
-      reticulate::py_require(c("torch", "pycox", "numpy"))
+      python_packages = c("torch", "pycox", "numpy")
+      reticulate::py_require(python_packages)
+      assert_python_packages(python_packages)
 
       pv = self$param_set$get_values(tags = "predict")
       newdata = as.data.frame(ordered_features(task, self))
