+++ R/learner_survivalmodels_surv_akritas.R
@@ -56,7 +56,9 @@ LearnerSurvAkritas = R6Class("LearnerSurvAkritas",
 
   private = list(
     .train = function(task) {
-      reticulate::py_require(c("torch", "pycox", "numpy"))
+      python_packages = c("torch", "pycox", "numpy")
+      reticulate::py_require(python_packages)
+      assert_python_packages(python_packages)
 
       pv = self$param_set$get_values(tags = "train")
       invoke(
@@ -69,7 +71,9 @@ LearnerSurvAkritas = R6Class("LearnerSurvAkritas",
     },
 
     .predict = function(task) {
-      reticulate::py_require(c("torch", "pycox", "numpy"))
+      python_packages = c("torch", "pycox", "numpy")
+      reticulate::py_require(python_packages)
+      assert_python_packages(python_packages)
 
       pv = self$param_set$get_values(tags = "predict")
       newdata = ordered_features(task, self)
