+++ R/learner_survivalmodels_surv_pchazard.R
@@ -133,7 +133,9 @@ LearnerSurvPCHazard = R6Class("LearnerSurvPCHazard",
 
   private = list(
     .train = function(task) {
-      reticulate::py_require(c("torch", "pycox", "numpy"))
+      python_packages = c("torch", "pycox", "numpy")
+      reticulate::py_require(python_packages)
+      assert_python_packages(python_packages)
 
       pars = self$param_set$get_values(tags = "train")
       invoke(
@@ -147,7 +149,9 @@ LearnerSurvPCHazard = R6Class("LearnerSurvPCHazard",
     },
 
     .predict = function(task) {
-      reticulate::py_require(c("torch", "pycox", "numpy"))
+      python_packages = c("torch", "pycox", "numpy")
+      reticulate::py_require(python_packages)
+      assert_python_packages(python_packages)
       pars = self$param_set$get_values(tags = "predict")
       newdata = ordered_features(task, self)
 
