+++ R/learner_survivalmodels_surv_coxtime.R
@@ -121,7 +121,9 @@ LearnerSurvCoxtime = R6Class("LearnerSurvCoxtime",
 
   private = list(
     .train = function(task) {
-      reticulate::py_require(c("torch", "pycox", "numpy"))
+      python_packages = c("torch", "pycox", "numpy")
+      reticulate::py_require(python_packages)
+      assert_python_packages(python_packages)
 
       pars = self$param_set$get_values(tags = "train")
       invoke(
@@ -135,7 +137,9 @@ LearnerSurvCoxtime = R6Class("LearnerSurvCoxtime",
     },
 
     .predict = function(task) {
-      reticulate::py_require(c("torch", "pycox", "numpy"))
+      python_packages = c("torch", "pycox", "numpy")
+      reticulate::py_require(python_packages)
+      assert_python_packages(python_packages)
 
       pars = self$param_set$get_values(tags = "predict")
       newdata = ordered_features(task, self)
