+++ R/learner_fastai_classif_fastai.R
@@ -13,12 +13,8 @@
 #' In this virtual environment, the `fastai` package and its dependencies must be installed.
 #'
 #' @section Saving a Learner:
-#' In order to save a `<%= class %>` for later usage, it is necessary to call the `$marshal()` method on the `Learner`  before writing it to disk, as the object will otherwise not be saved correctly.
-#' After loading a marshaled `<%= class %>` into R again,
-#' you then need to call `$unmarshal()` to transform it into a useable state.
-#'
-#' @section Initial parameter values:
-#' - `n_jobs` is initialized to 1 to avoid threading conflicts with \CRANpkg{future}.
+#' In order to save a `lrn("classif.fastai")` for later usage, it is necessary to call the `$marshal()` method on the `Learner`  before writing it to disk, as the object will otherwise not be saved correctly.
+#' After loading a marshaled `lrn("classif.fastai")` into R again, you then need to call `$unmarshal()` to transform it into a useable state.
 #'
 #' @section Initial parameter values:
 #' - `n_epoch`:
