+++ R/learner_survivalmodels_surv_deephit.R
@@ -135,9 +135,7 @@ LearnerSurvDeephit = R6Class("LearnerSurvDeephit",
 
   private = list(
     .train = function(task) {
-      python_packages = c("torch", "pycox", "numpy")
-      reticulate::py_require(python_packages)
-      assert_python_packages(python_packages)
+      assert_python_packages(c("torch", "pycox", "numpy"))
 
       pars = self$param_set$get_values(tags = "train")
       invoke(
@@ -151,9 +149,8 @@ LearnerSurvDeephit = R6Class("LearnerSurvDeephit",
     },
 
     .predict = function(task) {
-      python_packages = c("torch", "pycox", "numpy")
-      reticulate::py_require(python_packages)
-      assert_python_packages(python_packages)
+      assert_python_packages(c("torch", "pycox", "numpy"))
+
       pars = self$param_set$get_values(tags = "predict")
       newdata = ordered_features(task, self)
 
