+++ R/learner_survivalmodels_surv_coxtime.R
@@ -121,9 +121,7 @@ LearnerSurvCoxtime = R6Class("LearnerSurvCoxtime",
 
   private = list(
     .train = function(task) {
-      python_packages = c("torch", "pycox", "numpy")
-      reticulate::py_require(python_packages)
-      assert_python_packages(python_packages)
+      assert_python_packages(c("torch", "pycox", "numpy"))
 
       pars = self$param_set$get_values(tags = "train")
       invoke(
@@ -137,9 +135,7 @@ LearnerSurvCoxtime = R6Class("LearnerSurvCoxtime",
     },
 
     .predict = function(task) {
-      python_packages = c("torch", "pycox", "numpy")
-      reticulate::py_require(python_packages)
-      assert_python_packages(python_packages)
+      assert_python_packages(c("torch", "pycox", "numpy"))
 
       pars = self$param_set$get_values(tags = "predict")
       newdata = ordered_features(task, self)
