+++ R/assertions.R
@@ -8,6 +8,7 @@
 #'
 #' @return (`character()`)\cr
 assert_python_packages = function(packages) {
+  reticulate::py_require(packages)
   available = map_lgl(packages, reticulate::py_module_available)
   if (any(!available)) {
     stopf("Package %s not available.", packages[!available])
