+++ R/learner_survivalmodels_surv_loghaz.R
@@ -133,9 +133,7 @@ LearnerSurvLogisticHazard = R6Class("LearnerSurvLogisticHazard",
 
   private = list(
     .train = function(task) {
-      python_packages = c("torch", "pycox", "numpy")
-      reticulate::py_require(python_packages)
-      assert_python_packages(python_packages)
+      assert_python_packages(c("torch", "pycox", "numpy"))
       pars = self$param_set$get_values(tags = "train")
       invoke(
         survivalmodels::loghaz,
@@ -148,9 +146,7 @@ LearnerSurvLogisticHazard = R6Class("LearnerSurvLogisticHazard",
     },
 
     .predict = function(task) {
-      python_packages = c("torch", "pycox", "numpy")
-      reticulate::py_require(python_packages)
-      assert_python_packages(python_packages)
+      assert_python_packages(c("torch", "pycox", "numpy"))
       pars = self$param_set$get_values(tags = "predict")
       newdata = ordered_features(task, self)
 
