+++ R/learner_survivalmodels_surv_pchazard.R
@@ -133,9 +133,7 @@ LearnerSurvPCHazard = R6Class("LearnerSurvPCHazard",
 
   private = list(
     .train = function(task) {
-      python_packages = c("torch", "pycox", "numpy")
-      reticulate::py_require(python_packages)
-      assert_python_packages(python_packages)
+      assert_python_packages(c("torch", "pycox", "numpy"))
 
       pars = self$param_set$get_values(tags = "train")
       invoke(
@@ -149,9 +147,7 @@ LearnerSurvPCHazard = R6Class("LearnerSurvPCHazard",
     },
 
     .predict = function(task) {
-      python_packages = c("torch", "pycox", "numpy")
-      reticulate::py_require(python_packages)
-      assert_python_packages(python_packages)
+      assert_python_packages(c("torch", "pycox", "numpy"))
       pars = self$param_set$get_values(tags = "predict")
       newdata = ordered_features(task, self)
 
