+++ R/learner_tabpfn_classif_tabpfn.R
@@ -106,9 +106,7 @@ LearnerClassifTabPFN = R6Class("LearnerClassifTabPFN",
 
   private = list(
     .train = function(task) {
-      python_packages = c("torch", "tabpfn")
-      reticulate::py_require(python_packages)
-      assert_python_packages(python_packages)
+      assert_python_packages(c("torch", "tabpfn"))
       tabpfn = reticulate::import("tabpfn")
 
       pars = self$param_set$get_values(tags = "train")
@@ -145,7 +143,6 @@ LearnerClassifTabPFN = R6Class("LearnerClassifTabPFN",
     },
 
     .predict = function(task) {
-      reticulate::py_require("tabpfn")
       assert_python_packages("tabpfn")
       reticulate::import("tabpfn")
       model = self$model$fitted
