+++ R/learner_survivalmodels_surv_dnnsurv.R
@@ -123,9 +123,7 @@ LearnerSurvDNNSurv = R6Class("LearnerSurvDNNSurv",
 
   private = list(
     .train = function(task) {
-      python_packages = c("keras", "tensorflow")
-      reticulate::py_require(python_packages)
-      assert_python_packages(python_packages)
+      assert_python_packages(c("keras", "tensorflow"))
 
       pars = self$param_set$get_values(tags = "train")
       invoke(
@@ -139,9 +137,7 @@ LearnerSurvDNNSurv = R6Class("LearnerSurvDNNSurv",
     },
 
     .predict = function(task) {
-      python_packages = c("keras", "tensorflow")
-      reticulate::py_require(python_packages)
-      assert_python_packages(python_packages)
+      assert_python_packages(c("keras", "tensorflow"))
 
       pars = self$param_set$get_values(tags = "predict")
       newdata = ordered_features(task, self)
