+++ R/learner_survivalmodels_surv_parametric.R
@@ -110,9 +110,7 @@ LearnerSurvParametric = R6Class("LearnerSurvParametric",
 
   private = list(
     .train = function(task) {
-      python_packages = c("torch", "pycox", "numpy")
-      reticulate::py_require(python_packages)
-      assert_python_packages(python_packages)
+      assert_python_packages(c("torch", "pycox", "numpy"))
 
       pv = self$param_set$get_values(tags = "train")
       pv$weights = private$.get_weights(task)
@@ -127,9 +125,7 @@ LearnerSurvParametric = R6Class("LearnerSurvParametric",
     },
 
     .predict = function(task) {
-      python_packages = c("torch", "pycox", "numpy")
-      reticulate::py_require(python_packages)
-      assert_python_packages(python_packages)
+      assert_python_packages(c("torch", "pycox", "numpy"))
 
       pv = self$param_set$get_values(tags = "predict")
       newdata = as.data.frame(ordered_features(task, self))
