+++ R/learner_botorch_regr_singletaskgp.R
@@ -60,7 +60,7 @@ LearnerRegrBotorchSingleTaskGP = R6Class("LearnerRegrBotorchSingleTaskGP",
   ),
   private = list(
     .train = function(task) {
-      reticulate::py_require(c("torch", "botorch", "gpytorch"))
+      assert_python_packages(c("torch", "botorch", "gpytorch"))
       torch = reticulate::import("torch")
       botorch = reticulate::import("botorch")
       gpytorch = reticulate::import("gpytorch")
@@ -88,7 +88,7 @@ LearnerRegrBotorchSingleTaskGP = R6Class("LearnerRegrBotorchSingleTaskGP",
     },
 
     .predict = function(task) {
-      reticulate::py_require(c("torch", "botorch", "gpytorch"))
+      assert_python_packages(c("torch", "botorch", "gpytorch"))
       torch = reticulate::import("torch")
       pars = self$param_set$get_values(tags = "predict")
 
