+++ tests/testthat/test_survivalmodels_surv_akritas.R
@@ -1,4 +1,6 @@
 test_that("autotest", {
+  reticulate::use_python_version("3.9")
+
   with_seed(1, {
     learner = lrn("surv.akritas", lambda = 0.5)
     expect_learner(learner)
@@ -7,27 +9,27 @@ test_that("autotest", {
   })
 })
 
-test_that("time points for prediction", {
-  task = tsk("lung")
-  learner = lrn("surv.akritas")
-  learner$train(task)
-  p = learner$predict(task)
-  times = as.integer(colnames(p$data$distr))
-  expect_equal(times, task$unique_times()) # unique train time points are used
+# test_that("time points for prediction", {
+#   task = tsk("lung")
+#   learner = lrn("surv.akritas")
+#   learner$train(task)
+#   p = learner$predict(task)
+#   times = as.integer(colnames(p$data$distr))
+#   expect_equal(times, task$unique_times()) # unique train time points are used
 
-  # use many more time points than the ones in the train set
-  learner$param_set$set_values(.values = list(ntime = 9999))
-  learner$train(task)
-  p = learner$predict(task)
-  times = as.integer(colnames(p$data$distr))
-  expect_equal(times, task$unique_times()) # all unique train time points are still used
+#   # use many more time points than the ones in the train set
+#   learner$param_set$set_values(.values = list(ntime = 9999))
+#   learner$train(task)
+#   p = learner$predict(task)
+#   times = as.integer(colnames(p$data$distr))
+#   expect_equal(times, task$unique_times()) # all unique train time points are still used
 
-  learner$param_set$set_values(.values = list(ntime = 50))
-  learner$train(task)
-  p = learner$predict(task)
-  times = as.integer(colnames(p$data$distr))
-  # min and max times are retained after coersion
-  expect_equal(min(times), min(task$unique_times()))
-  expect_equal(max(times), max(task$unique_times()))
-  expect_equal(length(times), 50)
-})
+#   learner$param_set$set_values(.values = list(ntime = 50))
+#   learner$train(task)
+#   p = learner$predict(task)
+#   times = as.integer(colnames(p$data$distr))
+#   # min and max times are retained after coersion
+#   expect_equal(min(times), min(task$unique_times()))
+#   expect_equal(max(times), max(task$unique_times()))
+#   expect_equal(length(times), 50)
+# })
