+++ R/learner_survivalmodels_surv_akritas.R
@@ -56,7 +56,7 @@ LearnerSurvAkritas = R6Class("LearnerSurvAkritas",
 
   private = list(
     .train = function(task) {
-      assert_python_packages(c("torch", "pycox", "numpy"))
+      assert_python_packages(c("torch", "pycox", "numpy"), python_version = "3.10")
 
       pv = self$param_set$get_values(tags = "train")
       invoke(
@@ -69,7 +69,7 @@ LearnerSurvAkritas = R6Class("LearnerSurvAkritas",
     },
 
     .predict = function(task) {
-      assert_python_packages(c("torch", "pycox", "numpy"))
+      assert_python_packages(c("torch", "pycox", "numpy"), python_version = "3.10")
 
       pv = self$param_set$get_values(tags = "predict")
       newdata = ordered_features(task, self)
