+++ R/assertions.R
@@ -7,8 +7,8 @@
 #'   Python packages to check.
 #'
 #' @return (`character()`)\cr
-assert_python_packages = function(packages) {
-  reticulate::py_require(packages)
+assert_python_packages = function(packages, python_version = NULL) {
+  reticulate::py_require(packages, python_version = python_version)
   available = map_lgl(packages, reticulate::py_module_available)
   if (any(!available)) {
     stopf("Package %s not available.", packages[!available])
