### 2025-07-25 91a8f15
Apply suggestions from code review
