+++ R/learner_survivalmodels_surv_pchazard.R
@@ -133,7 +133,7 @@ LearnerSurvPCHazard = R6Class("LearnerSurvPCHazard",
 
   private = list(
     .train = function(task) {
-      assert_python_packages(c("torch", "pycox", "numpy"))
+      assert_python_packages(c("torch", "pycox", "numpy"), python_version = "3.10")
 
       pars = self$param_set$get_values(tags = "train")
       invoke(
@@ -147,7 +147,7 @@ LearnerSurvPCHazard = R6Class("LearnerSurvPCHazard",
     },
 
     .predict = function(task) {
-      assert_python_packages(c("torch", "pycox", "numpy"))
+      assert_python_packages(c("torch", "pycox", "numpy"), python_version = "3.10")
       pars = self$param_set$get_values(tags = "predict")
       newdata = ordered_features(task, self)
 
