+++ R/learner_survivalmodels_surv_coxtime.R
@@ -121,7 +121,7 @@ LearnerSurvCoxtime = R6Class("LearnerSurvCoxtime",
 
   private = list(
     .train = function(task) {
-      assert_python_packages(c("torch", "pycox", "numpy"))
+      assert_python_packages(c("torch", "pycox", "numpy"), python_version = "3.10")
 
       pars = self$param_set$get_values(tags = "train")
       invoke(
@@ -135,7 +135,7 @@ LearnerSurvCoxtime = R6Class("LearnerSurvCoxtime",
     },
 
     .predict = function(task) {
-      assert_python_packages(c("torch", "pycox", "numpy"))
+      assert_python_packages(c("torch", "pycox", "numpy"), python_version = "3.10")
 
       pars = self$param_set$get_values(tags = "predict")
       newdata = ordered_features(task, self)
