+++ R/learner_survivalmodels_surv_deephit.R
@@ -135,7 +135,7 @@ LearnerSurvDeephit = R6Class("LearnerSurvDeephit",
 
   private = list(
     .train = function(task) {
-      assert_python_packages(c("torch", "pycox", "numpy"))
+      assert_python_packages(c("torch", "pycox", "numpy"), python_version = "3.10")
 
       pars = self$param_set$get_values(tags = "train")
       invoke(
@@ -149,7 +149,7 @@ LearnerSurvDeephit = R6Class("LearnerSurvDeephit",
     },
 
     .predict = function(task) {
-      assert_python_packages(c("torch", "pycox", "numpy"))
+      assert_python_packages(c("torch", "pycox", "numpy"), python_version = "3.10")
 
       pars = self$param_set$get_values(tags = "predict")
       newdata = ordered_features(task, self)
