+++ attic/test_2/test_survivalmodels_surv_coxtime.R
@@ -20,12 +20,12 @@ test_that("autotest", {
     np$random$seed(1L)
     torch$manual_seed(1L)
 
-    learner = lrn("surv.deephit")
+    learner = lrn("surv.coxtime")
     expect_learner(learner)
 
     # single test fails randomly I think this depends on the python version
-    result = run_autotest(learner, check_replicable = FALSE, exclude = "sanity")
+    result = run_autotest(learner, check_replicable = FALSE, exclude = "sanity || feat_single")
   }))
 
-  expect_true(result)
+  expect_true(result, info = result$error)
 })
