+++ attic/test_2/test_survivalmodels_surv_dnnsurv.R
@@ -1,12 +1,10 @@
 skip_on_os("windows")
 
 test_that("autotest", {
-
   result = mirai::collect_mirai(mirai::mirai({
     library(mlr3)
     library(mlr3proba)
     library(mlr3extralearners)
-
 
     lapply(list.files(system.file("testthat", package = "mlr3"),
       pattern = "^helper.*\\.[rR]", full.names = TRUE), source)
@@ -17,16 +15,16 @@ test_that("autotest", {
     reticulate::py_require(c("numpy", "torch", "pycox"), python_version = "3.10")
     np = reticulate::import("numpy")
     torch = reticulate::import("torch")
-    set.seed(1)
+    set.seed(10)
     np$random$seed(1L)
     torch$manual_seed(1L)
 
-    learner = lrn("surv.deepsurv")
+    learner = lrn("surv.dnnsurv", cuts = 5)
     expect_learner(learner)
 
     # single test fails randomly I think this depends on the python version
-    result = run_autotest(learner, check_replicable = FALSE, exclude = "sanity")
+    result = run_autotest(learner, check_replicable = FALSE, exclude = "sanity || feat_single_integer")
   }))
 
-  expect_true(result)
+  expect_true(result, info = result$error)
 })
