+++ tests/testthat/test_survivalmodels_surv_akritas.R
@@ -19,27 +19,95 @@ test_that("autotest", {
   expect_true(result, info = result$error)
 })
 
-# test_that("time points for prediction", {
-#   task = tsk("lung")
-#   learner = lrn("surv.akritas")
-#   learner$train(task)
-#   p = learner$predict(task)
-#   times = as.integer(colnames(p$data$distr))
-#   expect_equal(times, task$unique_times()) # unique train time points are used
+test_that("time points for prediction", {
+  result = mirai::collect_mirai(mirai::mirai({
+    library(mlr3)
+    library(mlr3proba)
+    library(mlr3extralearners)
 
-#   # use many more time points than the ones in the train set
-#   learner$param_set$set_values(.values = list(ntime = 9999))
-#   learner$train(task)
-#   p = learner$predict(task)
-#   times = as.integer(colnames(p$data$distr))
-#   expect_equal(times, task$unique_times()) # all unique train time points are still used
+    lapply(list.files(system.file("testthat", package = "mlr3"),
+      pattern = "^helper.*\\.[rR]", full.names = TRUE), source)
 
-#   learner$param_set$set_values(.values = list(ntime = 50))
-#   learner$train(task)
-#   p = learner$predict(task)
-#   times = as.integer(colnames(p$data$distr))
-#   # min and max times are retained after coersion
-#   expect_equal(min(times), min(task$unique_times()))
-#   expect_equal(max(times), max(task$unique_times()))
-#   expect_equal(length(times), 50)
-# })
+    lapply(list.files(system.file("testthat", package = "mlr3proba"),
+      pattern = "^helper.*\\.[rR]", full.names = TRUE), source)
+
+    task = tsk("lung")
+    learner = lrn("surv.akritas")
+    learner$train(task)
+    p = learner$predict(task)
+    times = as.integer(colnames(p$data$distr))
+    testthat::expect_equal(times, task$unique_times()) # unique train time points are used
+  }))
+
+  expect_false(mirai::is_mirai_error(result), info = as.character(result))
+
+  result = mirai::collect_mirai(mirai::mirai({
+    library(mlr3)
+    library(mlr3proba)
+    library(mlr3extralearners)
+
+    lapply(list.files(system.file("testthat", package = "mlr3"),
+      pattern = "^helper.*\\.[rR]", full.names = TRUE), source)
+
+    lapply(list.files(system.file("testthat", package = "mlr3proba"),
+      pattern = "^helper.*\\.[rR]", full.names = TRUE), source)
+
+    task = tsk("lung")
+    learner = lrn("surv.akritas")
+    learner$train(task)
+    p = learner$predict(task)
+    times = as.integer(colnames(p$data$distr))
+    testthat::expect_equal(times, task$unique_times()) # unique train time points are used
+  }))
+
+  expect_false(mirai::is_mirai_error(result), info = as.character(result))
+
+  result = mirai::collect_mirai(mirai::mirai({
+    library(mlr3)
+    library(mlr3proba)
+    library(mlr3extralearners)
+
+    lapply(list.files(system.file("testthat", package = "mlr3"),
+      pattern = "^helper.*\\.[rR]", full.names = TRUE), source)
+
+    lapply(list.files(system.file("testthat", package = "mlr3proba"),
+      pattern = "^helper.*\\.[rR]", full.names = TRUE), source)
+
+    task = tsk("lung")
+    learner = lrn("surv.akritas")
+
+    # use many more time points than the ones in the train set
+    learner$param_set$set_values(.values = list(ntime = 9999))
+    learner$train(task)
+    p = learner$predict(task)
+    times = as.integer(colnames(p$data$distr))
[…]
