+++ attic/test_2/test_survivalmodels_surv_parametric.R
@@ -1,149 +1,0 @@
-test_that("autotest aft", {
-  result = mirai::collect_mirai(mirai::mirai({
-    library(mlr3)
-    library(mlr3proba)
-    library(mlr3extralearners)
-
-
-    lapply(list.files(system.file("testthat", package = "mlr3"),
-      pattern = "^helper.*\\.[rR]", full.names = TRUE), source)
-
-    lapply(list.files(system.file("testthat", package = "mlr3proba"),
-      pattern = "^helper.*\\.[rR]", full.names = TRUE), source)
-
-    set.seed(1)
-
-    learner = lrn("surv.parametric")
-    expect_learner(learner)
-
-    # single test fails randomly I think this depends on the python version
-    result = run_autotest(learner, check_replicable = FALSE, exclude = "utf8_feature_names")
-  }))
-
-  expect_true(result, info = result$error)
-})
-
-test_that("autotest ph", {
-  result = mirai::collect_mirai(mirai::mirai({
-    library(mlr3)
-    library(mlr3proba)
-    library(mlr3extralearners)
-
-    lapply(list.files(system.file("testthat", package = "mlr3"),
-      pattern = "^helper.*\\.[rR]", full.names = TRUE), source)
-
-    lapply(list.files(system.file("testthat", package = "mlr3proba"),
-      pattern = "^helper.*\\.[rR]", full.names = TRUE), source)
-
-    set.seed(1)
-
-    learner = lrn("surv.parametric", form = "ph")
-    expect_learner(learner)
-
-    # single test fails randomly I think this depends on the python version
-    result = run_autotest(learner, check_replicable = FALSE, exclude = "utf8_feature_names")
-  }))
-
-  expect_true(result, info = result$error)
-})
-
-test_that("autotest po", {
-  result = mirai::collect_mirai(mirai::mirai({
-    library(mlr3)
-    library(mlr3proba)
-    library(mlr3extralearners)
-
-    lapply(list.files(system.file("testthat", package = "mlr3"),
-      pattern = "^helper.*\\.[rR]", full.names = TRUE), source)
-
-    lapply(list.files(system.file("testthat", package = "mlr3proba"),
-      pattern = "^helper.*\\.[rR]", full.names = TRUE), source)
-
-    set.seed(1)
-
-    learner = lrn("surv.parametric", form = "po")
-    expect_learner(learner)
-
-    # single test fails randomly I think this depends on the python version
-    result = run_autotest(learner, check_replicable = FALSE, exclude = "utf8_feature_names")
-  }))
-
-  expect_true(result, info = result$error)
-})
-
-task = tsk("lung")
-
-# test_that("time points for prediction", {
-#   learner = lrn("surv.parametric")
-#   learner$train(task)
-#   p = learner$predict(task)
-#   times = as.integer(colnames(p$data$distr))
-#   expect_equal(times, task$unique_times()) # unique train time points are used
-
-#   # use many more time points than the ones in the train set
-#   learner$param_set$set_values(.values = list(ntime = 9999))
-#   learner$train(task)
-#   p = learner$predict(task)
-#   times = as.integer(colnames(p$data$distr))
-#   expect_equal(times, task$unique_times()) # all unique train time points are still used
-
-#   learner$param_set$set_values(.values = list(ntime = 50))
-#   p = learner$train(task)$predict(task)
-#   times = as.integer(colnames(p$data$distr))
-#   # min and max times are retained after coersion
-#   expect_equal(min(times), min(task$unique_times()))
-#   expect_equal(max(times), max(task$unique_times()))
[…]
