+++ attic/test_2/test_survivalmodels_surv_pchazard.R
@@ -27,3 +27,4 @@ test_that("autotest", {
 
   expect_true(result)
 })
+
