+++ R/learner_survivalmodels_surv_coxtime.R
@@ -1,10 +1,10 @@
-#' @title Survival DeepHit Learner
+#' @title Survival Cox-Time Learner
 #' @author RaphaelS1
-#' @name mlr_learners_surv.deephit
+#' @name mlr_learners_surv.coxtime
 #'
 #' @description
-#' Neural network 'Deephit' for survival analysis.
-#' Calls [survivalmodels::deephit()] from pacakge 'survivalmodels'.
+#' Cox-Time survival model.
+#' Calls [survivalmodels::coxtime()] from package 'survivalmodels'.
 #'
 #' @section Prediction types:
 #' This learner returns two prediction types:
@@ -14,23 +14,16 @@
 #' 2. `crank`: the expected mortality using [survivalmodels::surv_to_risk()].
 #'
 #' @template learner
-#' @templateVar id surv.deephit
+#' @templateVar id surv.coxtime
 #'
 #' @template install_survivalmodels
 #'
-#' @details
-#' Custom nets can be used in this learner either using the
-#'  [survivalmodels::build_pytorch_net] utility function or using `torch` via \CRANpkg{reticulate}.
-#'  The number of output channels depends on the number of discretised time-points, i.e.
-#'  the parameters `cuts` or `cutpoints`.
-#'
 #' @references
-#' `r format_bib("lee2018deephit")`
+#' `r format_bib("kvamme2019time")`
 #'
 #' @template seealso_learner
-#' @template simple_example
 #' @export
-LearnerSurvDeephit = R6Class("LearnerSurvDeephit",
+LearnerSurvCoxtime = R6Class("LearnerSurvCoxtime",
   inherit = mlr3proba::LearnerSurv,
 
   public = list(
@@ -40,11 +33,10 @@ LearnerSurvDeephit = R6Class("LearnerSurvDeephit",
 
       ps = ps(
         frac = p_dbl(default = 0, lower = 0, upper = 1, tags = "train"),
-        cuts = p_int(default = 10L, lower = 1L, tags = "train"),
-        cutpoints = p_uty(tags = "train"),
-        scheme = p_fct(default = "equidistant", levels = c("equidistant", "quantiles"),
-          tags = "train"),
-        cut_min = p_dbl(default = 0, lower = 0, tags = "train"),
+        standardize_time = p_lgl(default = FALSE, tags = "train"),
+        log_duration = p_lgl(default = FALSE, tags = "train"),
+        with_mean = p_lgl(default = TRUE, tags = "train"),
+        with_std = p_lgl(default = TRUE, tags = "train"),
         num_nodes = p_uty(default = c(32L, 32L), tags = "train"),
         batch_norm = p_lgl(default = TRUE, tags = "train"),
         dropout = p_dbl(lower = 0, upper = 1, tags = "train"),
@@ -54,10 +46,8 @@ LearnerSurvDeephit = R6Class("LearnerSurvDeephit",
             "rrelu", "relu", "selu", "sigmoid", "softmax", "softmax2d", "softmin",
             "softplus", "softshrink", "softsign", "tanh", "tanhshrink", "threshold"),
           tags = "train"),
-        custom_net = p_uty(tags = "train"),
         device = p_uty(tags = "train"),
-        mod_alpha = p_dbl(default = 0.2, lower = 0, upper = 1, tags = "train"),
-        sigma = p_dbl(default = 0.1, lower = 0, tags = "train"),
+        shrink = p_dbl(default = 0, lower = 0, tags = "train"),
         optimizer = p_fct(default = "adam",
           levels = c("adadelta", "adagrad", "adam", "adamax", "adamw", "asgd", "rmsprop", "rprop",
[…]
