+++ tests/testthat/test_survivalmodels_surv_parametric.R
@@ -1,5 +1,6 @@
 test_that("autotest aft", {
-  result = mirai::collect_mirai(mirai::mirai({
+  expect_true(callr::r(function() {
+    Sys.setenv(RETICULATE_PYTHON = "managed")
     library(mlr3)
     library(mlr3proba)
     library(mlr3extralearners)
@@ -18,13 +19,13 @@ test_that("autotest aft", {
 
     # single test fails randomly I think this depends on the python version
     result = run_autotest(learner, check_replicable = FALSE, exclude = "utf8_feature_names")
+    TRUE
   }))
-
-  expect_true(result, info = result$error)
 })
 
 test_that("autotest ph", {
-  result = mirai::collect_mirai(mirai::mirai({
+  expect_true(callr::r(function() {
+    Sys.setenv(RETICULATE_PYTHON = "managed")
     library(mlr3)
     library(mlr3proba)
     library(mlr3extralearners)
@@ -42,13 +43,13 @@ test_that("autotest ph", {
 
     # single test fails randomly I think this depends on the python version
     result = run_autotest(learner, check_replicable = FALSE, exclude = "utf8_feature_names")
+    TRUE
   }))
-
-  expect_true(result, info = result$error)
 })
 
 test_that("autotest po", {
-  result = mirai::collect_mirai(mirai::mirai({
+  expect_true(callr::r(function() {
+    Sys.setenv(RETICULATE_PYTHON = "managed")
     library(mlr3)
     library(mlr3proba)
     library(mlr3extralearners)
@@ -66,13 +67,13 @@ test_that("autotest po", {
 
     # single test fails randomly I think this depends on the python version
     result = run_autotest(learner, check_replicable = FALSE, exclude = "utf8_feature_names")
+    TRUE
   }))
-
-  expect_true(result, info = result$error)
 })
 
 test_that("time points for prediction", {
-  result = mirai::collect_mirai(mirai::mirai({
+  expect_true(callr::r(function() {
+    Sys.setenv(RETICULATE_PYTHON = "managed")
     library(mlr3)
     library(mlr3proba)
     library(mlr3extralearners)
@@ -89,11 +90,11 @@ test_that("time points for prediction", {
     p = learner$predict(task)
     times = as.integer(colnames(p$data$distr))
     testthat::expect_equal(times, task$unique_times()) # unique train time points are used
+    TRUE
   }))
 
-  expect_false(mirai::is_mirai_error(result), info = as.character(result))
-
-  result = mirai::collect_mirai(mirai::mirai({
+  expect_true(callr::r(function() {
+    Sys.setenv(RETICULATE_PYTHON = "managed")
     library(mlr3)
     library(mlr3proba)
     library(mlr3extralearners)
@@ -112,12 +113,11 @@ test_that("time points for prediction", {
     p = learner$predict(task)
     times = as.integer(colnames(p$data$distr))
     testthat::expect_equal(times, task$unique_times()) # all unique train time points are still used
+    TRUE
   }))
 
-  expect_false(mirai::is_mirai_error(result), info = as.character(result))
-
-
-  result = mirai::collect_mirai(mirai::mirai({
+  expect_true(callr::r(function() {
+    Sys.setenv(RETICULATE_PYTHON = "managed")
     library(mlr3)
     library(mlr3proba)
     library(mlr3extralearners)
@@ -137,13 +137,14 @@ test_that("time points for prediction", {
     testthat::expect_equal(min(times), min(task$unique_times()))
     testthat::expect_equal(max(times), max(task$unique_times()))
     testthat::expect_equal(length(times), 50)
+    TRUE
   }))
 
[…]
