+++ tests/testthat/helper.R
@@ -38,9 +38,3 @@ s4_helper = function(x) {
   formals(f) = pairlist()
   return(f())
 }
-
-# add survival lung task with missing features
-lung = survival::lung
-lung$status = lung$status - 1
-lung_missings = mlr3proba::TaskSurv$new("lung", backend = lung, time = "time", event = "status")
-
