+++ tests/testthat/test_survivalmodels_surv_pchazard.R
@@ -1,11 +1,10 @@
 skip_on_os("windows")
 
 test_that("autotest", {
-  result = mirai::collect_mirai(mirai::mirai({
+  expect_true(callr::r(function() {
     library(mlr3)
     library(mlr3proba)
     library(mlr3extralearners)
-
 
     lapply(list.files(system.file("testthat", package = "mlr3"),
       pattern = "^helper.*\\.[rR]", full.names = TRUE), source)
@@ -20,12 +19,11 @@ test_that("autotest", {
     np$random$seed(1L)
     torch$manual_seed(1L)
 
-    learner = lrn("surv.loghaz")
+    learner = lrn("surv.pchazard")
     expect_learner(learner)
 
-    # single test fails randomly I think this depends on the python version
-    result = run_autotest(learner, check_replicable = FALSE, exclude = "sanity")
+    run_autotest(learner, check_replicable = FALSE, exclude = "sanity")
+    TRUE
   }))
-
-  expect_true(result, info = result$error)
 })
+

[…]
