+++ tests/testthat/test_fastai_classif_fastai.R
@@ -6,215 +6,215 @@ test_that("autotest", {
   expect_true(result, info = result$error)
 })
 
-# test_that("eval protocol", {
-#   learner = lrn("classif.fastai")
-#   task = tsk("sonar")
-#   learner$train(task)
-#   expect_true(is.list(learner$state$eval_protocol))
-# })
+test_that("eval protocol", {
+  learner = lrn("classif.fastai")
+  task = tsk("sonar")
+  learner$train(task)
+  expect_true(is.list(learner$state$eval_protocol))
+})
 
 
-# test_that("validation and inner tuning works", {
+test_that("validation and inner tuning works", {
 
 
-#   task = tsk("spam")
+  task = tsk("spam")
 
-#   # with n_epoch and patience parameter
-#   learner = lrn("classif.fastai",
-#     n_epoch = 10,
-#     patience = 1,
-#     validate = 0.2
-#   )
+  # with n_epoch and patience parameter
+  learner = lrn("classif.fastai",
+    n_epoch = 10,
+    patience = 1,
+    validate = 0.2
+  )
 
-#   learner$train(task)
-#   expect_named(learner$state$eval_protocol, c("epoch", "train_loss", "valid_loss", "accuracy"))
-#   expect_list(learner$internal_valid_scores, types = "numeric")
-#   expect_equal(names(learner$internal_valid_scores), "accuracy")
-#   expect_list(learner$internal_tuned_values, types = "integerish")
-#   expect_equal(names(learner$internal_tuned_values), "n_epoch")
+  learner$train(task)
+  expect_named(learner$state$eval_protocol, c("epoch", "train_loss", "valid_loss", "accuracy"))
+  expect_list(learner$internal_valid_scores, types = "numeric")
+  expect_equal(names(learner$internal_valid_scores), "accuracy")
+  expect_list(learner$internal_tuned_values, types = "integerish")
+  expect_equal(names(learner$internal_tuned_values), "n_epoch")
 
-#   # without validate parameter
-#   learner$validate = NULL
-#   expect_error(learner$train(task), "field 'validate'")
+  # without validate parameter
+  learner$validate = NULL
+  expect_error(learner$train(task), "field 'validate'")
 
-#   # with patience parameter
-#   learner = lrn("classif.fastai",
-#     n_epoch = 10,
-#     validate = 0.2
-#   )
+  # with patience parameter
+  learner = lrn("classif.fastai",
+    n_epoch = 10,
+    validate = 0.2
+  )
 
-#   learner$train(task)
-#   expect_equal(learner$internal_tuned_values, NULL)
-#   expect_named(learner$state$eval_protocol, c("epoch", "train_loss", "valid_loss", "accuracy"))
-#   expect_list(learner$internal_valid_scores, types = "numeric")
-#   expect_equal(names(learner$internal_valid_scores), "accuracy")
+  learner$train(task)
+  expect_equal(learner$internal_tuned_values, NULL)
+  expect_named(learner$state$eval_protocol, c("epoch", "train_loss", "valid_loss", "accuracy"))
+  expect_list(learner$internal_valid_scores, types = "numeric")
+  expect_equal(names(learner$internal_valid_scores), "accuracy")
 
-#   # internal tuning
-#   learner = lrn("classif.fastai",
-#     n_epoch = to_tune(upper = 20, internal = TRUE),
-#     validate = 0.2
-#   )
-#   s = learner$param_set$search_space()
-#   expect_error(learner$param_set$convert_internal_search_space(s), "patience")
-#   learner$param_set$set_values(n_epoch = 10)
-#   learner$param_set$disable_internal_tuning("n_epoch")
[…]
