+++ tests/testthat/test_survivalmodels_surv_pchazard.R
@@ -2,6 +2,7 @@ skip_on_os("windows")
 
 test_that("autotest", {
   expect_true(callr::r(function() {
+    Sys.setenv(RETICULATE_PYTHON = "managed")
     library(mlr3)
     library(mlr3proba)
     library(mlr3extralearners)
