+++ man/mlr_learners_regr.mars.Rd
@@ -1,0 +1,141 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/learner_mda_regr_mars.R
+\name{mlr_learners_regr.mars}
+\alias{mlr_learners_regr.mars}
+\alias{LearnerRegrMars}
+\title{Regression Mars Learner}
+\description{
+Multivariate Adaptive Regression Splines.
+Calls \code{\link[mda:mars]{mda::mars()}} from \CRANpkg{mda}.
+}
+\section{Dictionary}{
+
+This \link[mlr3:Learner]{Learner} can be instantiated via \link[mlr3:mlr_sugar]{lrn()}:
+
+\if{html}{\out{<div class="sourceCode">}}\preformatted{lrn("regr.mars")
+}\if{html}{\out{</div>}}
+}
+
+\section{Meta Information}{
+
+\itemize{
+\item Task type: \dQuote{regr}
+\item Predict Types: \dQuote{response}
+\item Feature Types: \dQuote{integer}, \dQuote{numeric}
+\item Required Packages: \CRANpkg{mlr3}, \CRANpkg{mlr3extralearners}, \CRANpkg{mda}
+}
+}
+
+\section{Parameters}{
+\tabular{lllll}{
+   Id \tab Type \tab Default \tab Levels \tab Range \cr
+   degree \tab integer \tab 1 \tab  \tab \eqn{[1, \infty)}{[1, Inf)} \cr
+   nk \tab integer \tab - \tab  \tab \eqn{[1, \infty)}{[1, Inf)} \cr
+   penalty \tab numeric \tab 2 \tab  \tab \eqn{[0, \infty)}{[0, Inf)} \cr
+   thresh \tab numeric \tab 0.001 \tab  \tab \eqn{[0, \infty)}{[0, Inf)} \cr
+   prune \tab logical \tab TRUE \tab TRUE, FALSE \tab - \cr
+   trace.mars \tab logical \tab FALSE \tab TRUE, FALSE \tab - \cr
+   forward.step \tab logical \tab FALSE \tab TRUE, FALSE \tab - \cr
+}
+}
+
+\examples{
+\dontshow{if (mlr3misc::require_namespaces(lrn("regr.mars")$packages, quietly = TRUE)) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
+# Define the Learner
+learner = mlr3::lrn("regr.mars")
+print(learner)
+
+# Define a Task
+task = mlr3::tsk("mtcars")
+
+# Create train and test set
+ids = mlr3::partition(task)
+
+# Train the learner on the training ids
+learner$train(task, row_ids = ids$train)
+
+print(learner$model)
+
+
+# Make predictions for the test rows
+predictions = learner$predict(task, row_ids = ids$test)
+
+# Score the predictions
+predictions$score()
+\dontshow{\}) # examplesIf}
+}
+\references{
+Friedman, H J (1991).
+\dQuote{Multivariate adaptive regression splines.}
+\emph{The annals of statistics}, \bold{19}(1), 1--67.
+}
+\seealso{
+\itemize{
+\item \link[mlr3misc:Dictionary]{Dictionary} of \link[mlr3:Learner]{Learners}: \link[mlr3:mlr_learners]{mlr3::mlr_learners}.
+\item \code{as.data.table(mlr_learners)} for a table of available \link[mlr3:Learner]{Learners} in the running session (depending on the loaded packages).
+\item Chapter in the \href{https://mlr3book.mlr-org.com/}{mlr3book}: \url{https://mlr3book.mlr-org.com/basics.html#learners}
+\item \CRANpkg{mlr3learners} for a selection of recommended learners.
+\item \CRANpkg{mlr3cluster} for unsupervised clustering learners.
[…]
