+++ man/mlr_learners_regr.glmboost.Rd
@@ -1,0 +1,154 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/learner_mboost_regr_glmboost.R
+\name{mlr_learners_regr.glmboost}
+\alias{mlr_learners_regr.glmboost}
+\alias{LearnerRegrGLMBoost}
+\title{Boosted Generalized Linear Regression Learner}
+\description{
+Fit a generalized linear regression model using a boosting algorithm.
+Calls \code{\link[mboost:glmboost]{mboost::glmboost()}} from \CRANpkg{mboost}.
+}
+\section{Dictionary}{
+
+This \link[mlr3:Learner]{Learner} can be instantiated via \link[mlr3:mlr_sugar]{lrn()}:
+
+\if{html}{\out{<div class="sourceCode">}}\preformatted{lrn("regr.glmboost")
+}\if{html}{\out{</div>}}
+}
+
+\section{Meta Information}{
+
+\itemize{
+\item Task type: \dQuote{regr}
+\item Predict Types: \dQuote{response}
+\item Feature Types: \dQuote{integer}, \dQuote{numeric}, \dQuote{factor}, \dQuote{ordered}
+\item Required Packages: \CRANpkg{mlr3}, \CRANpkg{mlr3extralearners}, \CRANpkg{mboost}
+}
+}
+
+\section{Parameters}{
+\tabular{lllll}{
+   Id \tab Type \tab Default \tab Levels \tab Range \cr
+   family \tab character \tab Gaussian \tab Gaussian, Laplace, Huber, Poisson, GammaReg, NBinomial, Hurdle, custom \tab - \cr
+   custom.family \tab untyped \tab - \tab  \tab - \cr
+   nuirange \tab untyped \tab c(0, 100) \tab  \tab - \cr
+   d \tab numeric \tab NULL \tab  \tab \eqn{(-\infty, \infty)}{(-Inf, Inf)} \cr
+   center \tab logical \tab TRUE \tab TRUE, FALSE \tab - \cr
+   mstop \tab integer \tab 100 \tab  \tab \eqn{(-\infty, \infty)}{(-Inf, Inf)} \cr
+   nu \tab numeric \tab 0.1 \tab  \tab \eqn{(-\infty, \infty)}{(-Inf, Inf)} \cr
+   risk \tab character \tab inbag \tab inbag, oobag, none \tab - \cr
+   oobweights \tab untyped \tab NULL \tab  \tab - \cr
+   trace \tab logical \tab FALSE \tab TRUE, FALSE \tab - \cr
+   stopintern \tab untyped \tab FALSE \tab  \tab - \cr
+   na.action \tab untyped \tab stats::na.omit \tab  \tab - \cr
+   contrasts.arg \tab untyped \tab - \tab  \tab - \cr
+}
+}
+
+\section{Offset}{
+
+If a \code{Task} contains a column with the \code{offset} role, it is automatically
+incorporated via the \code{offset} argument in \code{mboost}'s training function.
+No offset is applied during prediction for this learner.
+}
+
+\examples{
+\dontshow{if (mlr3misc::require_namespaces(lrn("regr.glmboost")$packages, quietly = TRUE)) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
+# Define the Learner
+learner = mlr3::lrn("regr.glmboost")
+print(learner)
+
+# Define a Task
+task = mlr3::tsk("mtcars")
+
+# Create train and test set
+ids = mlr3::partition(task)
+
+# Train the learner on the training ids
+learner$train(task, row_ids = ids$train)
+
+print(learner$model)
+
+
+# Make predictions for the test rows
+predictions = learner$predict(task, row_ids = ids$test)
+
+# Score the predictions
[…]

[…]
