+++ DESCRIPTION
@@ -74,6 +74,7 @@ Suggests:
     klaR,
     knitr,
     ks,
+    lgr,
     LiblineaR,
     lightgbm (>= 4.5.0),
     lme4,
