+++ tests/testthat/test_survivalmodels_surv_parametric.R
@@ -184,7 +184,7 @@ test_that("'form' affects survival prediction", {
     learner$train(task)
     p = learner$predict(task)
     testthat::expect_true(inherits(p$distr, "Matdist"))
-    p$distr$survival(task$unique_times())
+    surv_aft = p$distr$survival(task$unique_times())
     checkmate::expect_matrix(surv_aft, nrows = length(task$unique_times()), ncols = task$nrow,  any.missing = FALSE) # [times x obs]
   }))
 
@@ -206,7 +206,7 @@ test_that("'form' affects survival prediction", {
     learner$train(task)
     p = learner$predict(task)
     testthat::expect_true(inherits(p$distr, "Matdist"))
-    p$distr$survival(task$unique_times())
+    surv_tobit = p$distr$survival(task$unique_times())
     checkmate::expect_matrix(surv_tobit, nrows = length(task$unique_times()), ncols = task$nrow, any.missing = FALSE) # [times x obs]
   }))
 
@@ -228,7 +228,7 @@ test_that("'form' affects survival prediction", {
     learner$train(task)
     p = learner$predict(task)
     testthat::expect_true(inherits(p$distr, "Matdist"))
-    p$distr$survival(task$unique_times())
+    surv_ph = p$distr$survival(task$unique_times())
     checkmate::expect_matrix(surv_ph, nrows = length(task$unique_times()), ncols = task$nrow, any.missing = FALSE) # [times x obs]
   }))
 
@@ -250,7 +250,7 @@ test_that("'form' affects survival prediction", {
     learner$train(task)
     p = learner$predict(task)
     testthat::expect_true(inherits(p$distr, "Matdist"))
-    p$distr$survival(task$unique_times())
+    surv_po = p$distr$survival(task$unique_times())
     checkmate::expect_matrix(surv_po, nrows = length(task$unique_times()), ncols = task$nrow, any.missing = FALSE) # [times x obs]
   }))
 })
