+++ attic/devcmdcheck.yml
@@ -1,74 +1,0 @@
-# r cmd check workflow of the mlr3 ecosystem v0.3.1
-# https://github.com/mlr-org/actions
-on:
-  workflow_dispatch:
-    inputs:
-      debug_enabled:
-        type: boolean
-        description: 'Run the build with tmate debugging enabled'
-        required: false
-        default: false
-  push:
-    branches:
-      - main
-  pull_request:
-    branches:
-      - main
-  schedule:
-    - cron:  '0 4 * * 1'
-
-name: Dev Check
-
-jobs:
-  check-package:
-    runs-on: ${{ matrix.config.os }}
-
-    name: ${{ matrix.config.dev-package }}
-
-    strategy:
-      fail-fast: false
-      matrix:
-        config:
-          - {os: ubuntu-latest, r: 'release', dev-package: "mlr-org/paradox', 'mlr-org/mlr3learners', 'mlr-org/mlr3pipelines', 'mlr-org/mlr3cluster"}
-
-    env:
-      GITHUB_PAT: ${{ secrets.GITHUB_TOKEN }}
-
-    steps:
-      - uses: actions/checkout@v4
-
-      - uses: r-lib/actions/setup-pandoc@v2
-
-      - uses: r-lib/actions/setup-r@v2
-        with:
-          r-version: ${{ matrix.config.r }}
-
-      - uses: r-lib/actions/setup-r-dependencies@v2
-        with:
-          extra-packages: any::rcmdcheck
-          needs: check
-
-      - uses: mxschmitt/action-tmate@v3
-        if: ${{ github.event_name == 'workflow_dispatch' && inputs.debug_enabled }}
-        with:
-          limit-access-to-actor: true
-
-      - name: Install Python packages
-        run: |
-          pak::pkg_install('rstudio/reticulate')
-          reticulate::install_miniconda(force = TRUE)
-          reticulate::use_condaenv('r-reticulate')
-          pak::pkg_install(c('eagerai/fastai', 'keras'))
-          fastai::install_fastai(gpu = FALSE, cuda_version = '11.6', overwrite = FALSE)
-          reticulate::py_install("pytorch-cpu")
-          reticulate::py_install("tensorflow-cpu", pip = TRUE)
-          survivalmodels::install_pycox(pip = TRUE, install_torch = FALSE)
-          survivalmodels::install_keras(pip = TRUE, install_tensorflow = FALSE)
-          reticulate::py_install("tabpfn")
-        shell: Rscript {0}
-
-      - name: Install dev versions
-        run: pak::pkg_install(c('${{ matrix.config.dev-package }}'))
-        shell: Rscript {0}
-
-      - uses: r-lib/actions/check-r-package@v2
