+++ R/learner_randomForestSRC_regr_rfsrc.R
@@ -73,7 +73,6 @@ LearnerRegrRandomForestSRC = R6Class("LearnerRegrRandomForestSRC",
           default = "na.omit", levels = c("na.omit", "na.impute"),
           tags = c("train", "predict")),
         nimpute = p_int(default = 1L, lower = 1L, tags = "train"),
-        ntime = p_int(lower = 1L, tags = "train"),
         proximity = p_fct(
           default = "FALSE",
           levels = c("FALSE", "TRUE", "inbag", "oob", "all"),
