+++ tests/testthat/test_paramtest_randomForestSRC_regr_rfsrc.R
@@ -10,6 +10,7 @@ test_that("paramtest regr.rfsrc train", {
     "cores", # set as option(rf.cores)
     "sampsize.ratio", # alternative to sampsize
     "mtry.ratio", # alternative to mtry
+    "ntime", # for survival forests only
     "cause" # for competing risk forests only
   )
 
