+++ R/learner_RWeka_classif_AdaBoostM1.R
@@ -108,4 +108,3 @@ LearnerClassifAdaBoostM1 = R6Class("LearnerClassifAdaBoostM1",
 )
 
 .extralrns_dict$add("classif.AdaBoostM1", LearnerClassifAdaBoostM1)
-
