+++ R/helpers_rweka.R
@@ -61,7 +61,8 @@ rweka_predict = function(newdata, pars, predict_type, model) {
 marshal_model.Weka_classifier = function(model, inplace = FALSE, ...) {
   rJava::.jcache(model$classifier)
   structure(list(
-    marshaled = model, packages = c("mlr3", "mlr3extralearners")),
+    marshaled = model,
+    packages = c("mlr3", "mlr3extralearners")),
     class = c("Weka_classifier_marshaled", "marshaled")
   )
 }
