+++ R/helpers_rweka.R
@@ -60,9 +60,7 @@ rweka_predict = function(newdata, pars, predict_type, model) {
 #' @export
 marshal_model.Weka_classifier = function(model, inplace = FALSE, ...) {
   rJava::.jcache(model$classifier)
-  structure(list(
-    marshaled = model,
-    packages = c("mlr3", "mlr3extralearners")),
+  structure(list( marshaled = model, packages = c("mlr3", "mlr3extralearners")),
     class = c("Weka_classifier_marshaled", "marshaled")
   )
 }
