+++ R/learner_survivalmodels_surv_deephit.R
@@ -134,11 +134,7 @@ LearnerSurvDeephit = R6Class("LearnerSurvDeephit",
 
   private = list(
     .train = function(task) {
-<<<<<<< HEAD
-      assert_python_packages(c("torch", "pycox", "numpy"))
-=======
       assert_python_packages(c("torch", "pycox", "numpy"), python_version = "3.10")
->>>>>>> 695e1e920008e5b76b20118737e5c02f1f5669f8
 
       pars = self$param_set$get_values(tags = "train")
       invoke(
@@ -152,11 +148,7 @@ LearnerSurvDeephit = R6Class("LearnerSurvDeephit",
     },
 
     .predict = function(task) {
-<<<<<<< HEAD
-      assert_python_packages(c("torch", "pycox", "numpy"))
-=======
       assert_python_packages(c("torch", "pycox", "numpy"), python_version = "3.10")
->>>>>>> 695e1e920008e5b76b20118737e5c02f1f5669f8
 
       pars = self$param_set$get_values(tags = "predict")
       newdata = ordered_features(task, self)
