+++ R/learner_survivalmodels_surv_akritas.R
@@ -55,11 +55,7 @@ LearnerSurvAkritas = R6Class("LearnerSurvAkritas",
 
   private = list(
     .train = function(task) {
-<<<<<<< HEAD
-      assert_python_packages(c("torch", "pycox", "numpy"))
-=======
       assert_python_packages(c("torch", "pycox", "numpy"), python_version = "3.10")
->>>>>>> 695e1e920008e5b76b20118737e5c02f1f5669f8
 
       pv = self$param_set$get_values(tags = "train")
       invoke(
@@ -72,11 +68,7 @@ LearnerSurvAkritas = R6Class("LearnerSurvAkritas",
     },
 
     .predict = function(task) {
-<<<<<<< HEAD
-      assert_python_packages(c("torch", "pycox", "numpy"))
-=======
       assert_python_packages(c("torch", "pycox", "numpy"), python_version = "3.10")
->>>>>>> 695e1e920008e5b76b20118737e5c02f1f5669f8
 
       pv = self$param_set$get_values(tags = "predict")
       newdata = ordered_features(task, self)
