+++ R/learner_survivalmodels_surv_loghaz.R
@@ -132,11 +132,7 @@ LearnerSurvLogisticHazard = R6Class("LearnerSurvLogisticHazard",
 
   private = list(
     .train = function(task) {
-<<<<<<< HEAD
-      assert_python_packages(c("torch", "pycox", "numpy"))
-=======
       assert_python_packages(c("torch", "pycox", "numpy"), python_version = "3.10")
->>>>>>> 695e1e920008e5b76b20118737e5c02f1f5669f8
       pars = self$param_set$get_values(tags = "train")
       invoke(
         survivalmodels::loghaz,
@@ -149,11 +145,7 @@ LearnerSurvLogisticHazard = R6Class("LearnerSurvLogisticHazard",
     },
 
     .predict = function(task) {
-<<<<<<< HEAD
-      assert_python_packages(c("torch", "pycox", "numpy"))
-=======
       assert_python_packages(c("torch", "pycox", "numpy"), python_version = "3.10")
->>>>>>> 695e1e920008e5b76b20118737e5c02f1f5669f8
       pars = self$param_set$get_values(tags = "predict")
       newdata = ordered_features(task, self)
 
