+++ R/learner_survivalmodels_surv_dnnsurv.R
@@ -122,11 +122,7 @@ LearnerSurvDNNSurv = R6Class("LearnerSurvDNNSurv",
 
   private = list(
     .train = function(task) {
-<<<<<<< HEAD
-      assert_python_packages(c("keras", "tensorflow"))
-=======
       assert_python_packages(c("keras", "tensorflow"), python_version = "3.10")
->>>>>>> 695e1e920008e5b76b20118737e5c02f1f5669f8
 
       pars = self$param_set$get_values(tags = "train")
       invoke(
@@ -140,11 +136,7 @@ LearnerSurvDNNSurv = R6Class("LearnerSurvDNNSurv",
     },
 
     .predict = function(task) {
-<<<<<<< HEAD
-      assert_python_packages(c("keras", "tensorflow"))
-=======
       assert_python_packages(c("keras", "tensorflow"), python_version = "3.10")
->>>>>>> 695e1e920008e5b76b20118737e5c02f1f5669f8
 
       pars = self$param_set$get_values(tags = "predict")
       newdata = ordered_features(task, self)
