### 2025-07-28 c8d2146
remove ::: from fastai learner
