+++ tests/testthat/test_botorch_regr_singletaskgp.R
@@ -1,6 +1,6 @@
-test_that("autotest: regr.botorch_singletaskgp", {
-  learner = mlr3::lrn("regr.botorch_singletaskgp")
-  expect_learner(learner)
-  result = run_autotest(learner)
-  expect_true(result, info = result$error)
-})
+# test_that("autotest: regr.botorch_singletaskgp", {
+#   learner = mlr3::lrn("regr.botorch_singletaskgp")
+#   expect_learner(learner)
+#   result = run_autotest(learner)
+#   expect_true(result, info = result$error)
+# })
