### 2025-07-28 a1b9b0f
add riskRegression in Suggests
