### 2025-07-28 7275b02
fix CI note on doc marshaling for fastai
