+++ man/mlr_learners_classif.fastai.Rd
@@ -26,10 +26,10 @@ After loading a marshaled \code{lrn("classif.fastai")} into R again, you then ne
 
 \itemize{
 \item \code{n_epoch}:
-Needs to be set for \code{\link[fastai:reexports]{fastai::fit()}} to work.
+Needs to be set for \code{\link[fastai:fit]{fastai::fit()}} to work.
 If no value is given, it is set to 5.
 \item \code{eval_metric}:
-Needs to be set for \code{\link[fastai:reexports]{fastai::predict()}} to work.
+Needs to be set for \code{\link[fastai:predict]{fastai::predict()}} to work.
 If no value is given, it is set to \code{fastai::accuracy()}.
 }
 }
@@ -178,7 +178,7 @@ Marshal the learner's model.
 \if{html}{\out{<div class="arguments">}}
 \describe{
 \item{\code{...}}{(any)\cr
-Additional arguments passed to \code{\link[=marshal_model]{marshal_model()}}.}
+Additional arguments passed to \code{\link[mlr3:marshaling]{marshal_model()}}.}
 }
 \if{html}{\out{</div>}}
 }
@@ -196,7 +196,7 @@ Unmarshal the learner's model.
 \if{html}{\out{<div class="arguments">}}
 \describe{
 \item{\code{...}}{(any)\cr
-Additional arguments passed to \code{\link[=unmarshal_model]{unmarshal_model()}}.}
+Additional arguments passed to \code{\link[mlr3:marshaling]{marshal_model()}}.}
 }
 \if{html}{\out{</div>}}
 }
