### 2025-07-28 632487f
enable selected_features property with guard against prediction bug
