+++ tests/testthat/test_randomForestSRC_classif_rfsrc.R
@@ -12,6 +12,7 @@ test_that("autotest", {
   with_seed(1, {
     learner = lrn("classif.rfsrc", ntree = 20, importance = "random", na.action = "na.impute")
     expect_learner(learner)
+    learner$properties = setdiff(learner$properties, "selected_features")
 
     result = run_autotest(learner, exclude = "uf8_feature_names")
     expect_true(result, info = result$error)
