+++ R/learner_randomForestSRC_classif_rfsrc.R
@@ -123,11 +123,8 @@ LearnerClassifRandomForestSRC = R6Class("LearnerClassifRandomForestSRC",
         feature_types = c("logical", "integer", "numeric", "factor"),
         predict_types = c("response", "prob"),
         param_set = ps,
-        # selected features is possible but there's a bug somewhere in rfsrc so that the model
-        # can be trained but not predicted. so public method retained but property not included
-        properties = c(
-          "weights", "missings", "importance", "oob_error",
-          "twoclass", "multiclass"),
+        properties = c("weights", "missings", "importance", "oob_error",
+                       "selected_features", "twoclass", "multiclass"),
         man = "mlr3extralearners::mlr_learners_classif.rfsrc",
         label = "Random Forest"
       )
@@ -148,9 +145,13 @@ LearnerClassifRandomForestSRC = R6Class("LearnerClassifRandomForestSRC",
 
     #' @description
     #' Selected features are extracted from the model slot `var.used`.
+    #'
+    #' **Note**: Due to a known issue in `randomForestSRC`, enabling `var.used = "all.trees"`
+    #' causes prediction to fail. Therefore, this setting should be used exclusively
+    #' for feature selection purposes and not when prediction is required.
     #' @return `character()`.
     selected_features = function() {
-      if (is.null(self$model$var.used) & !is.null(self$model)) {
+      if (is.null(self$model$var.used) && !is.null(self$model)) {
         stopf("Set 'var.used' to 'all.trees'.")
       }
 
@@ -182,6 +183,11 @@ LearnerClassifRandomForestSRC = R6Class("LearnerClassifRandomForestSRC",
     .predict = function(task) {
       newdata = data.table::setDF(ordered_features(task, self))
       pars = self$param_set$get_values(tags = "predict")
+
+      if (!is.null(pars$var.used) && pars$var.used == "all.trees") {
+        stop("Prediction is not supported when var.used = 'all.trees'. Use this setting only when extracting selected features.") #nolint
+      }
+
       cores = pars$cores %??% 1L
       pars$cores = NULL
 
