### 2025-07-29 41c9aca
feat: marshall rweka models (#395)
