+++ R/learner_RWeka_classif_decision_stump.R
@@ -52,10 +52,33 @@ LearnerClassifDecisionStump = R6Class("LearnerClassifDecisionStump",
         feature_types = c("logical", "integer", "numeric", "factor", "ordered"),
         predict_types = c("response", "prob"),
         param_set = param_set,
-        properties = c("missings", "multiclass", "twoclass"),
+        properties = c("missings", "multiclass", "twoclass", "marshal"),
         man = "mlr3extralearners::mlr_learners_classif.decision_stump",
         label = "Decision Stump"
       )
+    },
+
+    #' @description
+    #' Marshal the learner's model.
+    #' @param ... (any)\cr
+    #'   Additional arguments passed to [`marshal_model()`].
+    marshal = function(...) {
+      learner_marshal(.learner = self, ...)
+    },
+    #' @description
+    #' Unmarshal the learner's model.
+    #' @param ... (any)\cr
+    #'   Additional arguments passed to [`unmarshal_model()`].
+    unmarshal = function(...) {
+      learner_unmarshal(.learner = self, ...)
+    }
+  ),
+
+  active = list(
+    #' @field marshaled (`logical(1)`)\cr
+    #' Whether the learner has been marshaled.
+    marshaled = function() {
+      learner_marshaled(self)
     }
   ),
 
