+++ tests/testthat/test_survivalmodels_surv_parametric.R
@@ -236,5 +236,3 @@ test_that("'form' affects survival prediction", {
     TRUE
   }))
 })
-
-
