+++ tests/testthat/test_survivalmodels_surv_akritas.R
@@ -106,5 +106,3 @@ test_that("time points for prediction", {
     TRUE
   }))
 })
-
-
