+++ tests/testthat/test_survivalmodels_surv_pchazard.R
@@ -25,4 +25,3 @@ test_that("autotest", {
     TRUE
   }))
 })
-
