### 2025-07-29 fa15f8f
fix and update test for selected features
