+++ R/learner_randomForestSRC_regr_rfsrc.R
@@ -115,9 +115,7 @@ LearnerRegrRandomForestSRC = R6Class("LearnerRegrRandomForestSRC",
         feature_types = c("logical", "integer", "numeric", "factor"),
         predict_types = "response",
         param_set = ps,
-        # selected features is possible but there's a bug somewhere in rfsrc so that the model
-        # can be trained but not predicted. so public method retained but property not included
-        properties = c("weights", "missings", "importance", "oob_error"),
+        properties = c("weights", "missings", "importance", "oob_error", "selected_features"),
         man = "mlr3extralearners::mlr_learners_regr.rfsrc",
         label = "Random Forest"
       )
@@ -137,10 +135,18 @@ LearnerRegrRandomForestSRC = R6Class("LearnerRegrRandomForestSRC",
 
     #' @description
     #' Selected features are extracted from the model slot `var.used`.
+    #'
+    #' **Note**: Due to a known issue in `randomForestSRC`, enabling `var.used = "all.trees"`
+    #' causes prediction to fail. Therefore, this setting should be used exclusively
+    #' for feature selection purposes and not when prediction is required.
     #' @return `character()`.
     selected_features = function() {
-      if (is.null(self$model$var.used) & !is.null(self$model)) {
-        stopf("Set 'var.used' to 'all.trees'.")
+      if (is.null(self$model)) {
+        stopf("No model stored")
+      }
+
+      if (is.null(self$model$var.used)) {
+        stopf("Set parameter 'var.used' to 'all.trees'.")
       }
 
       names(self$model$var.used)
@@ -170,14 +176,19 @@ LearnerRegrRandomForestSRC = R6Class("LearnerRegrRandomForestSRC",
 
     .predict = function(task) {
       newdata = ordered_features(task, self)
-      pars = self$param_set$get_values(tags = "predict")
-      cores = pars$cores %??% 1L
-      pars$cores = NULL
+      pv = self$param_set$get_values(tags = "predict")
+
+      if (!is.null(pv$var.used) && pv$var.used == "all.trees") {
+        stopf("Prediction is not supported when var.used = 'all.trees'. Use this setting only when extracting selected features.") #nolint
+      }
+
+      cores = pv$cores %??% 1L
+      pv$cores = NULL
 
       list(
         response = invoke(predict,
           object = self$model, newdata = newdata,
-          .args = pars, .opts = list(rf.cores = cores))$predicted
+          .args = pv, .opts = list(rf.cores = cores))$predicted
       )
     }
   )
