+++ tests/testthat/test_randomForestSRC_classif_rfsrc.R
@@ -1,3 +1,15 @@
+test_that("autotest", {
+  with_seed(1, {
+    learner = lrn("classif.rfsrc", ntree = 20, importance = "random", na.action = "na.impute")
+    expect_learner(learner)
+    # remove property as prediction doesn't work due to rsfrc bug
+    learner$properties = setdiff(learner$properties, "selected_features")
+
+    result = run_autotest(learner, exclude = "uf8_feature_names")
+    expect_true(result, info = result$error)
+  })
+})
+
 test_that("importance/selected", {
   task = tsk("iris")
   learner = lrn("classif.rfsrc", ntree = 20)
@@ -11,18 +23,6 @@ test_that("importance/selected", {
   learner$train(task)
   expect_character(learner$selected_features())
   expect_error(learner$predict(task), "Prediction is not supported")
-})
-
-test_that("autotest", {
-  with_seed(1, {
-    learner = lrn("classif.rfsrc", ntree = 20, importance = "random", na.action = "na.impute")
-    expect_learner(learner)
-    # remove property as prediction doesn't work due to rsfrc bug
-    learner$properties = setdiff(learner$properties, "selected_features")
-
-    result = run_autotest(learner, exclude = "uf8_feature_names")
-    expect_true(result, info = result$error)
-  })
 })
 
 test_that("convert_ratio", {
