+++ tests/testthat/test_randomForestSRC_surv_rfsrc.R
@@ -2,6 +2,9 @@ test_that("autotest", {
   with_seed(1, {
     learner = lrn("surv.rfsrc", ntree = 20, importance = "random", na.action = "na.impute")
     expect_learner(learner)
+    # remove property as prediction doesn't work due to rsfrc bug
+    learner$properties = setdiff(learner$properties, "selected_features")
+
     result = run_autotest(learner, check_replicable = FALSE, N = 100)
     expect_true(result, info = result$error)
   })
@@ -13,7 +16,7 @@ test_that("importance/selected", {
     learner = lrn("surv.rfsrc", ntree = 20)
     learner$train(task)
     expect_error(learner$importance(), "Set 'importance'")
-    expect_error(learner$selected_features(), "Set 'var.used'")
+    expect_error(learner$selected_features(), "Set parameter 'var.used'")
 
     learner$param_set$values = list(
       var.used = "all.trees", importance = "random"
@@ -21,5 +24,6 @@ test_that("importance/selected", {
     learner$train(task)
     expect_character(learner$selected_features())
     expect_numeric(learner$importance(), names = "named")
+    expect_error(learner$predict(task), "Prediction is not supported")
   })
 })
