### 2025-07-29 04d961f
remove unneeded call to py_require() in akritas and parametric survival learners
