+++ man/mlr_learners_classif.rfsrc.Rd
@@ -46,7 +46,6 @@ This \link[mlr3:Learner]{Learner} can be instantiated via \link[mlr3:mlr_sugar]{
    sampsize.ratio \tab numeric \tab - \tab  \tab \eqn{[0, 1]}{[0, 1]} \cr
    na.action \tab character \tab na.omit \tab na.omit, na.impute \tab - \cr
    nimpute \tab integer \tab 1 \tab  \tab \eqn{[1, \infty)}{[1, Inf)} \cr
-   ntime \tab integer \tab - \tab  \tab \eqn{[1, \infty)}{[1, Inf)} \cr
    proximity \tab character \tab FALSE \tab FALSE, TRUE, inbag, oob, all \tab - \cr
    distance \tab character \tab FALSE \tab FALSE, TRUE, inbag, oob, all \tab - \cr
    forest.wt \tab character \tab FALSE \tab FALSE, TRUE, inbag, oob, all \tab - \cr
@@ -186,6 +185,10 @@ Named \code{numeric()}.
 \if{latex}{\out{\hypertarget{method-LearnerClassifRandomForestSRC-selected_features}{}}}
 \subsection{Method \code{selected_features()}}{
 Selected features are extracted from the model slot \code{var.used}.
+
+\strong{Note}: Due to a known issue in \code{randomForestSRC}, enabling \code{var.used = "all.trees"}
+causes prediction to fail. Therefore, this setting should be used exclusively
+for feature selection purposes and not when prediction is required.
 \subsection{Usage}{
 \if{html}{\out{<div class="r">}}\preformatted{LearnerClassifRandomForestSRC$selected_features()}\if{html}{\out{</div>}}
 }
