+++ R/learner_tabpfn_classif_tabpfn.R
@@ -85,14 +85,14 @@ LearnerClassifTabPFN = R6Class("LearnerClassifTabPFN",
     #' @param ... (any)\cr
     #'   Additional arguments passed to [`marshal_model()`][mlr3::marshaling()].
     marshal = function(...) {
-      mlr3::learner_marshal(.learner = self, ...)
+      learner_marshal(.learner = self, ...)
     },
     #' @description
     #' Unmarshal the learner's model.
     #' @param ... (any)\cr
     #'   Additional arguments passed to [`unmarshal_model()`][mlr3::marshaling()].
     unmarshal = function(...) {
-      mlr3::learner_unmarshal(.learner = self, ...)
+      learner_unmarshal(.learner = self, ...)
     }
   ),
 
@@ -100,7 +100,7 @@ LearnerClassifTabPFN = R6Class("LearnerClassifTabPFN",
     #' @field marshaled (`logical(1)`)\cr
     #' Whether the learner has been marshaled.
     marshaled = function() {
-      mlr3::learner_marshaled(self)
+      learner_marshaled(self)
     }
   ),
 
