+++ R/learner_fastai_classif_fastai.R
@@ -101,14 +101,14 @@ LearnerClassifFastai = R6Class("LearnerClassifFastai",
     #' @param ... (any)\cr
     #'   Additional arguments passed to [`marshal_model()`][mlr3::marshaling()].
     marshal = function(...) {
-      mlr3::learner_marshal(.learner = self, ...)
+      learner_marshal(.learner = self, ...)
     },
     #' @description
     #' Unmarshal the learner's model.
     #' @param ... (any)\cr
     #'   Additional arguments passed to [`marshal_model()`][mlr3::marshaling()].
     unmarshal = function(...) {
-      mlr3::learner_unmarshal(.learner = self, ...)
+      learner_unmarshal(.learner = self, ...)
     }
   ),
 
@@ -141,7 +141,7 @@ LearnerClassifFastai = R6Class("LearnerClassifFastai",
     #' @field marshaled (`logical(1)`)
     #' Whether the learner has been marshaled.
     marshaled = function() {
-      mlr3::learner_marshaled(self)
+      learner_marshaled(self)
     }
   ),
 
