+++ R/learner_xgboost_surv_xgboost_cox.R
@@ -154,14 +154,14 @@ LearnerSurvXgboostCox = R6Class("LearnerSurvXgboostCox",
     #' @param ... (any)\cr
     #'   Additional arguments passed to [`marshal_model()`][mlr3::marshaling()].
     marshal = function(...) {
-      mlr3::learner_marshal(.learner = self, ...)
+      learner_marshal(.learner = self, ...)
     },
     #' @description
     #' Unmarshal the learner's model.
     #' @param ... (any)\cr
     #'   Additional arguments passed to [`unmarshal_model()`][mlr3::marshaling()].
     unmarshal = function(...) {
-      mlr3::learner_unmarshal(.learner = self, ...)
+      learner_unmarshal(.learner = self, ...)
     }
   ),
 
@@ -189,7 +189,7 @@ LearnerSurvXgboostCox = R6Class("LearnerSurvXgboostCox",
     #' @field marshaled (`logical(1)`)\cr
     #' Whether the learner has been marshaled.
     marshaled = function() {
-      mlr3::learner_marshaled(self)
+      learner_marshaled(self)
     }
   ),
 
