+++ R/learner_RWeka_classif_PART.R
@@ -77,14 +77,14 @@ LearnerClassifPART = R6Class("LearnerClassifPART",
     #' @description
     #' Marshal the learner's model.
     #' @param ... (any)\cr
-    #'   Additional arguments passed to [`marshal_model()`].
+    #'   Additional arguments passed to [`mlr3::marshal_model()`].
     marshal = function(...) {
       learner_marshal(.learner = self, ...)
     },
     #' @description
     #' Unmarshal the learner's model.
     #' @param ... (any)\cr
-    #'   Additional arguments passed to [`unmarshal_model()`].
+    #'   Additional arguments passed to [`mlr3::unmarshal_model()`].
     unmarshal = function(...) {
       learner_unmarshal(.learner = self, ...)
     }
