### 2025-07-30 e117228
fixed model call
