### 2025-07-30 73ce37b
test: remove old checks
