+++ tests/testthat/test_tabpfn_classif_tabpfn.R
@@ -1,15 +1,4 @@
-tryCatch({
-  reticulate::use_condaenv("r-reticulate", required = TRUE)
-}, error = function(e) {
-  skip("r-reticulate conda environment not found.")
-})
-if (!reticulate::py_module_available("torch") || !reticulate::py_module_available("tabpfn")) {
-  skip("torch or tabpfn not available for testing.")
-}
-
 test_that("autotest", {
-
-
   learner = lrn("classif.tabpfn")
   expect_learner(learner)
   # reproducibility is not guaranteed, hence check_replicable = FALSE
@@ -18,8 +7,6 @@ test_that("autotest", {
 })
 
 test_that("marshaling works for classif.tabpfn", {
-
-
   learner = lrn("classif.tabpfn")
   task = tsk("iris")
   # expect_marshalable_learner(learner, task)
@@ -46,8 +33,6 @@ test_that("marshaling works for classif.tabpfn", {
 })
 
 test_that("categorical feature columns are encoded correctly", {
-
-
   n = 6
   task = as_task_classif(
     data.frame(
@@ -66,8 +51,6 @@ test_that("categorical feature columns are encoded correctly", {
 })
 
 test_that("device selection works", {
-
-
   torch = reticulate::import("torch")
 
   learner = lrn("classif.tabpfn", device = "cpu")
@@ -103,8 +86,6 @@ test_that("inference_precision works", {
 })
 
 test_that("checks for memory saving mode work", {
-
-
   learner = lrn("classif.tabpfn")
   task = tsk("iris")
 
