### 2025-07-30 74f199c
fix predict_newdata_fast autotest fail
