### 2025-07-30 d1f764c
fix autotest
