### 2025-08-01 b22c787
feat: added learner ExhaustiveSearch (classif, regr) (#443)
