+++ man/mlr_learners_regr.exhaustive_search.Rd
@@ -1,0 +1,168 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/learner_ExhaustiveSearch_regr_lm.R
+\name{mlr_learners_regr.exhaustive_search}
+\alias{mlr_learners_regr.exhaustive_search}
+\alias{LearnerRegrExhaustiveSearch}
+\title{Linear Regression with Exhaustive Search}
+\description{
+Linear Regression, where the features used in the model are selected by running Exhaustive Search.
+
+Calls \code{\link[ExhaustiveSearch:ExhaustiveSearch]{ExhaustiveSearch::ExhaustiveSearch()}} from \CRANpkg{ExhaustiveSearch}.
+}
+\details{
+Internal validation not yet implemented.
+}
+\section{Initial parameter values}{
+
+\itemize{
+\item \code{family}:
+\itemize{
+\item Actual default: NULL
+\item Adjusted default: "gaussian"
+\item Reason for change: To comply with mlr3 architecture, we differentiate between classification and regression learners.
+}
+\item \code{nThreads}:
+\itemize{
+\item Actual default: NULL
+\item Adjusted default: 1
+\item Reason for change: Suppressing the automatic internal parallelization if
+\code{cv.folds} > 0.
+}
+\item \code{quietly}:
+\itemize{
+\item Actual default: FALSE
+\item Adjusted default: TRUE
+\item Reason for change: Suppression of constant printing to console
+}
+}
+}
+
+\section{Dictionary}{
+
+This \link[mlr3:Learner]{Learner} can be instantiated via \link[mlr3:mlr_sugar]{lrn()}:
+
+\if{html}{\out{<div class="sourceCode">}}\preformatted{lrn("regr.exhaustive_search")
+}\if{html}{\out{</div>}}
+}
+
+\section{Meta Information}{
+
+\itemize{
+\item Task type: \dQuote{regr}
+\item Predict Types: \dQuote{response}, \dQuote{se}
+\item Feature Types: \dQuote{logical}, \dQuote{integer}, \dQuote{numeric}, \dQuote{character}, \dQuote{factor}
+\item Required Packages: \CRANpkg{mlr3}, \CRANpkg{mlr3extralearners}, \CRANpkg{ExhaustiveSearch}
+}
+}
+
+\section{Parameters}{
+\tabular{lllll}{
+   Id \tab Type \tab Default \tab Levels \tab Range \cr
+   family \tab character \tab - \tab gaussian, binomial \tab - \cr
+   performanceMeasure \tab character \tab - \tab MSE, AIC \tab - \cr
+   combsUpTo \tab integer \tab - \tab  \tab \eqn{[1, \infty)}{[1, Inf)} \cr
+   nResults \tab integer \tab 5000 \tab  \tab \eqn{[1, \infty)}{[1, Inf)} \cr
+   nThreads \tab integer \tab - \tab  \tab \eqn{[1, \infty)}{[1, Inf)} \cr
+   testSetIDs \tab integer \tab - \tab  \tab \eqn{[1, \infty)}{[1, Inf)} \cr
+   errorVal \tab untyped \tab -1 \tab  \tab - \cr
+   quietly \tab logical \tab - \tab TRUE, FALSE \tab - \cr
+   checkLarge \tab logical \tab TRUE \tab TRUE, FALSE \tab - \cr
+}
+}
+
+\examples{
+# define learner
+learner = LearnerRegrExhaustiveSearch$new()
+learner$predict_type = "se"
+
+# define task
+tsk_cars = tsk("mtcars")
+
+# train learner
+learner$train(tsk_cars)
+
+# extract selected features
+learner$selected_features()
+
+# predict on training task
[…]
