+++ NAMESPACE
@@ -22,6 +22,7 @@ export(LearnerClassifDecisionStump)
 export(LearnerClassifDecisionTable)
 export(LearnerClassifDiagLda)
 export(LearnerClassifEarth)
+export(LearnerClassifExhaustiveSearch)
 export(LearnerClassifEvtree)
 export(LearnerClassifFNN)
 export(LearnerClassifFastai)
@@ -86,6 +87,7 @@ export(LearnerRegrCubist)
 export(LearnerRegrDecisionStump)
 export(LearnerRegrDecisionTable)
 export(LearnerRegrEarth)
+export(LearnerRegrExhaustiveSearch)
 export(LearnerRegrEvtree)
 export(LearnerRegrFNN)
 export(LearnerRegrGAMBoost)

[…]
