### 2025-08-01 9c2fdd8
fix: TabPFN parameters (#449)
