+++ tests/testthat/test_tabpfn_regr_tabpfn.R
@@ -1,6 +1,4 @@
 test_that("autotest", {
-
-
   learner = lrn("regr.tabpfn")
   expect_learner(learner)
   # reproducibility is not guaranteed, hence check_replicable = FALSE
@@ -9,8 +7,6 @@ test_that("autotest", {
 })
 
 test_that("marshaling works for regr.tabpfn", {
-
-
   learner = lrn("regr.tabpfn")
   task = tsk("mtcars")
   # expect_marshalable_learner(learner, task)
@@ -37,8 +33,6 @@ test_that("marshaling works for regr.tabpfn", {
 })
 
 test_that("categorical feature columns are encoded correctly", {
-
-
   n = 6
   task = as_task_regr(
     data.frame(
@@ -57,8 +51,6 @@ test_that("categorical feature columns are encoded correctly", {
 })
 
 test_that("device selection works", {
-
-
   torch = reticulate::import("torch")
 
   learner = lrn("regr.tabpfn", device = "cpu")
@@ -71,9 +63,30 @@ test_that("device selection works", {
   expect_identical(learner$model$fitted$device, "auto")
 })
 
+
+test_that("inference_precision works", {
+  torch = reticulate::import("torch")
+  task = tsk("mtcars")
+
+  # No test for the "autocast" option, because it is not supported on cpu.
+  # But we test all possible torch dtypes.
+  dtypes = c(
+    "float32", "float",
+    "float64", "double",
+    "float16", "half",
+    "bfloat16"
+  )
+  lapply(dtypes, function(dtype) {
+    learner = lrn("regr.tabpfn", inference_precision = paste0("torch.", dtype))
+    expect_invisible(learner$train(task))
+    actual_dtype = learner$model$fitted$inference_precision
+    expected_dtype = reticulate::py_get_attr(torch, dtype)
+    # actual and expected should be the same torch.dtype object
+    expect_identical(reticulate::py_id(actual_dtype), reticulate::py_id(expected_dtype))
+  })
+})
+
 test_that("checks for memory saving mode work", {
-
-
   learner = lrn("regr.tabpfn")
   task = tsk("mtcars")
 
@@ -91,4 +104,33 @@ test_that("checks for memory saving mode work", {
   learner$param_set$set_values(memory_saving_mode = 50)
   learner$train(task)
   expect_identical(learner$model$fitted$memory_saving_mode, 50)
+})
+
+test_that("random_state works", {
+  task = tsk("mtcars")
+
+  # different seeds => slightly different predictions
+  learner1 = lrn("regr.tabpfn", predict_type = "quantiles")
+  learner1$param_set$set_values(random_state = "None")
+  learner1$train(task)
+  learner2 = lrn("regr.tabpfn", predict_type = "quantiles")
+  learner2$param_set$set_values(random_state = "None")
+  learner2$train(task)
+  expect_character(all.equal(learner1$predict(task), learner2$predict(task)))
+
+  # same seed (default: 0) => (almost) same predictions
+  learner1 = lrn("regr.tabpfn", predict_type = "quantiles")
+  learner1$train(task)
+  learner2 = lrn("regr.tabpfn", predict_type = "quantiles")
+  learner2$train(task)
+  expect_equal(learner1$predict(task), learner2$predict(task))
+
+  # same seed => (almost) same predictions
+  learner1 = lrn("regr.tabpfn", predict_type = "quantiles")
+  learner1$param_set$set_values(random_state = 42L)
+  learner1$train(task)
+  learner2 = lrn("regr.tabpfn", predict_type = "quantiles")
+  learner2$param_set$set_values(random_state = 42L)
+  learner2$train(task)
+  expect_equal(learner1$predict(task), learner2$predict(task))
 })
