+++ NAMESPACE
@@ -22,8 +22,8 @@ export(LearnerClassifDecisionStump)
 export(LearnerClassifDecisionTable)
 export(LearnerClassifDiagLda)
 export(LearnerClassifEarth)
-export(LearnerClassifExhaustiveSearch)
 export(LearnerClassifEvtree)
+export(LearnerClassifExhaustiveSearch)
 export(LearnerClassifFNN)
 export(LearnerClassifFastai)
 export(LearnerClassifGAMBoost)
@@ -87,8 +87,8 @@ export(LearnerRegrCubist)
 export(LearnerRegrDecisionStump)
 export(LearnerRegrDecisionTable)
 export(LearnerRegrEarth)
-export(LearnerRegrExhaustiveSearch)
 export(LearnerRegrEvtree)
+export(LearnerRegrExhaustiveSearch)
 export(LearnerRegrFNN)
 export(LearnerRegrGAMBoost)
 export(LearnerRegrGBM)
