### 2025-08-01 f599b15
add numpy assertion to fastai
