### 2025-08-01 272c5ad
fastai not found?
