### 2025-08-01 2b310c2
check fastai learner
