### 2025-08-01 99affb9
check fastai learner
