### 2025-08-01 4a0374e
fix fastai module availability
