### 2025-08-01 1a17e61
replate assert with py_require in fastai
