### 2025-08-05 29a3dc1
tutorial for internal validation and tuning
