### 2025-08-05 04eaa9f
add reticulate to tutorial
