### 2025-08-05 21cd004
add marshaling of reticulate to tutorial
