### 2025-08-05 e446baf
refactor fastai tests with callr
