+++ tests/testthat/test_fastai_classif_fastai.R
@@ -1,214 +1,273 @@
 test_that("autotest", {
-  learner = lrn("classif.fastai", layers = c(200, 100))
-  expect_learner(learner, check_man = FALSE)
-  # results not replicable, because torch seed must be set in the python backend
-  result = run_autotest(learner, check_replicable = FALSE)
-  expect_true(result, info = result$error)
+  expect_true(callr::r(function() {
+    Sys.setenv(RETICULATE_PYTHON = "managed")
+    library(mlr3)
+    library(mlr3extralearners)
+
+    lapply(list.files(system.file("testthat", package = "mlr3"),
+                      pattern = "^helper.*\\.[rR]", full.names = TRUE), source)
+
+    reticulate::py_require(c("numpy", "fastai"), python_version = "3.10")
+
+    learner = lrn("classif.fastai", layers = c(200, 100))
+    expect_learner(learner, check_man = FALSE)
+    # results not replicable, because torch seed must be set in the python backend
+    result = run_autotest(learner, check_replicable = FALSE)
+    testthat::expect_true(result, info = result$error)
+    TRUE
+  }))
 })
 
 test_that("eval protocol", {
-  learner = lrn("classif.fastai")
-  task = tsk("sonar")
-  learner$train(task)
-  expect_true(is.list(learner$state$eval_protocol))
+  expect_true(callr::r(function() {
+    Sys.setenv(RETICULATE_PYTHON = "managed")
+    library(mlr3)
+    library(mlr3extralearners)
+
+    lapply(list.files(system.file("testthat", package = "mlr3"),
+                      pattern = "^helper.*\\.[rR]", full.names = TRUE), source)
+
+    reticulate::py_require(c("numpy", "fastai"), python_version = "3.10")
+
+    learner = lrn("classif.fastai")
+    task = tsk("sonar")
+    learner$train(task)
+    expect_true(is.list(learner$state$eval_protocol))
+    TRUE
+  }))
 })
 
 
 test_that("validation and inner tuning works", {
-  task = tsk("spam")
+  expect_true(callr::r(function() {
+    Sys.setenv(RETICULATE_PYTHON = "managed")
+    library(mlr3)
+    library(mlr3extralearners)
 
-  # with n_epoch and patience parameter
-  learner = lrn("classif.fastai",
-    n_epoch = 10,
-    patience = 1,
-    validate = 0.2
-  )
+    lapply(list.files(system.file("testthat", package = "mlr3"),
+                      pattern = "^helper.*\\.[rR]", full.names = TRUE), source)
 
-  learner$train(task)
-  expect_named(learner$state$eval_protocol, c("epoch", "train_loss", "valid_loss", "accuracy"))
-  expect_list(learner$internal_valid_scores, types = "numeric")
-  expect_equal(names(learner$internal_valid_scores), "accuracy")
-  expect_list(learner$internal_tuned_values, types = "integerish")
-  expect_equal(names(learner$internal_tuned_values), "n_epoch")
+    reticulate::py_require(c("numpy", "fastai"), python_version = "3.10")
 
-  # without validate parameter
-  learner$validate = NULL
-  expect_error(learner$train(task), "field 'validate'")
+    task = tsk("spam")
 
-  # with patience parameter
-  learner = lrn("classif.fastai",
-    n_epoch = 10,
-    validate = 0.2
-  )
+    # with n_epoch and patience parameter
+    learner = lrn("classif.fastai",
+                  n_epoch = 10,
+                  patience = 1,
+                  validate = 0.2
+    )
 
-  learner$train(task)
-  expect_equal(learner$internal_tuned_values, NULL)
-  expect_named(learner$state$eval_protocol, c("epoch", "train_loss", "valid_loss", "accuracy"))
[…]
