### 2025-08-09 #8331 @RichardLitt
@kelly-sovacool Are you able to finish your review?
