### 2025-10-24 #8331 @sebffischer
@RichardLitt I think I have attended everything. @kelly-sovacool let me know whether you consider everything addressed and thanks for the review again!
