### 2025-11-14 #8331 @danielskatz
I can't check this in detail now but it looks like a label (fig? Table? Bib entry?) is being reused / is defined twice. Check the md and bib files
