### 2025-11-14 #8331 @bblodfon
@RichardLitt I removed the duplicated reference
